### 2023-09-25 #5874 @editorialbot
[PRE REVIEW]: HiddenMarkovModels.jl: generic, fast and reliable state space modeling
**Submitting author:** @gdalle (<a href="http://orcid.org/0000-0003-4866-1687">Guillaume Dalle</a>)
**Repository:** https://github.com/gdalle/HiddenMarkovModels.jl
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.3.0
