+++ Project.toml
@@ -1,7 +1,7 @@
 name = "HiddenMarkovModels"
 uuid = "84ca31d5-effc-45e0-bfda-5a68cd981f47"
 authors = ["Guillaume Dalle"]
-version = "0.3.0"
+version = "0.3.1"
 
 [deps]
 ChainRulesCore = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"
@@ -27,7 +27,9 @@ HiddenMarkovModelsHMMBaseExt = "HMMBase"
 ChainRulesCore = "1.16"
 DensityInterface = "0.4"
 Distributions = "0.25"
+LinearAlgebra = "1.6"
 PrecompileTools = "1.1"
+Random = "1.6"
 RequiredInterfaces = "0.1.3"
 Requires = "1.3"
 SimpleUnPack = "1.1"
