### 2023-10-30 90664bb
Stdlib compat + Documenter v1 + bump to 0.3.1 (#54)
* Add stdlib compat, bump to 0.3.1

* Fix docs

* Fix broken links

* Remove empty benchmarks warning
