+++ docs/make.jl
@@ -22,6 +22,12 @@ open(joinpath(joinpath(@__DIR__, "src"), "index.md"), "w") do io
     end
 end
 
+alt_pages = if get(ENV, "HMM_BENCHMARKS_DONE", false)
+    ["Features" => "alt_features.md", "Performance" => "alt_performance.md"]
+else
+    ["Features" => "alt_features.md"]
+end
+
 pages = [
     "Home" => "index.md",
     "Essentials" => ["Background" => "background.md", "API reference" => "api.md"],
@@ -30,28 +36,25 @@ pages = [
         "Custom HMM" => "tuto_custom.md",
         "Debugging" => "debugging.md",
     ],
-    "Alternatives" =>
-        ["Features" => "alt_features.md", "Performance" => "alt_performance.md"],
+    "Alternatives" => alt_pages,
     "Advanced" => ["Formulas" => "formulas.md", "Roadmap" => "roadmap.md"],
 ]
 
 fmt = Documenter.HTML(;
     prettyurls=get(ENV, "CI", "false") == "true",
+    repolink="https://github.com/gdalle/HiddenMarkovModels.jl",
     canonical="https://gdalle.github.io/HiddenMarkovModels.jl",
-    edit_link="main",
     assets=String[],
 )
 
-makedocs(
-    bib;
+makedocs(;
     modules=[HiddenMarkovModels],
     authors="Guillaume Dalle",
-    repo="https://github.com/gdalle/HiddenMarkovModels.jl/blob/{commit}{path}#{line}",
     sitename="HiddenMarkovModels.jl",
     format=fmt,
     pages=pages,
-    linkcheck=false,
-    strict=false,
+    plugins=[bib],
+    pagesonly=true,
 )
 
 deploydocs(; repo="github.com/gdalle/HiddenMarkovModels.jl", devbranch="main")
