+++ src/types/abstract_hmm.jl
@@ -8,7 +8,7 @@ Abstract supertype for an HMM amenable to simulation, inference and learning.
 - `initial_distribution(hmm)`
 - `transition_matrix(hmm)`
 - `obs_distribution(hmm, i)`
-- `fit!(hmm, init_count, trans_count, obs_seq, state_marginals)` (optional)
+- `fit!(hmm, obs_seqs, fbs)` (optional)
 
 # Applicable methods
 
@@ -48,6 +48,11 @@ The returned object `dist` must implement
 """
 function obs_distribution end
 
+"""
+    StatsAPI.fit!(hmm::AbstractHMM, obs_seqs, fbs)
+"""
+StatsAPI.fit!  # TODO: docstring
+
 function Base.rand(rng::AbstractRNG, hmm::AbstractHMM, T::Integer)
     mc = MarkovChain(hmm)
     state_seq = rand(rng, mc, T)
@@ -62,33 +67,4 @@ end
 
 function MarkovChain(hmm::AbstractHMM)
     return MarkovChain(initial_distribution(hmm), transition_matrix(hmm))
-end
-
-"""
-    PermutedHMM{H<:AbstractHMM}
-
-Wrapper around an `AbstractHMM` that permutes its states.
-
-This is computationally inefficient and mostly useful for evaluation.
-
-# Fields
-
-- `hmm:H`: the old HMM
-- `perm::Vector{Int}`: a permutation such that state `i` in the new HMM corresponds to state `perm[i]` in the old.
-"""
-struct PermutedHMM{H<:AbstractHMM} <: AbstractHMM
-    hmm::H
-    perm::Vector{Int}
-end
-
-Base.length(p::PermutedHMM) = length(p.hmm)
-
-HMMs.initial_distribution(p::PermutedHMM) = initial_distribution(p.hmm)[p.perm]
-
-function HMMs.transition_matrix(p::PermutedHMM)
-    return transition_matrix(p.hmm)[p.perm, :][:, p.perm]
-end
-
-function HMMs.obs_distribution(p::PermutedHMM, i::Integer)
-    return obs_distribution(p.hmm, p.perm[i])
 end
