+++ src/inference/forward_backward.jl
@@ -15,6 +15,7 @@ The following fields are internals and subject to change:
 - `α::Matrix{R}`: scaled forward variables `α[i,t]` proportional to `ℙ(Y[1:t], X[t]=i)` (up to a function of `t`)
 - `β::Matrix{R}`: scaled backward variables `β[i,t]` proportional to `ℙ(Y[t+1:T] | X[t]=i)` (up to a function of `t`)
 - `c::Vector{R}`: forward variable inverse normalizations `c[t] = 1 / sum(α[:,t])`
+- `logB::Matrix{R}`: observation loglikelihoods `logB[i,t]`
 - `logm::Vector{R}`: maximum of the observation loglikelihoods `logm[t] = maximum(logB[:, t])`
 - `Bscaled::Matrix{R}`: numerically stabilized observation likelihoods `Bscaled[i,t] = exp.(logB[i,t] - logm[t])`
 - `Bβscaled::Matrix{R}`: numerically stabilized product `Bβscaled[i,t] = Bscaled[i,t] * β[i,t]`
@@ -25,6 +26,7 @@ struct ForwardBackwardStorage{R}
     γ::Matrix{R}
     ξ::Array{R,3}
     c::Vector{R}
+    logB::Matrix{R}
     logm::Vector{R}
     Bscaled::Matrix{R}
     Bβscaled::Matrix{R}
@@ -46,38 +48,66 @@ function loglikelihood(fbs::Vector{ForwardBackwardStorage{R}}) where {R}
     return logL
 end
 
-function initialize_forward_backward(p, A, logB)
+function initialize_forward_backward(hmm, obs_seq)
+    init = initialization(hmm)
+    trans = transition_matrix(hmm)
+    dists = obs_distributions(hmm)
+    useless_logdensity_value = logdensityof(dists[1], obs_seq[1])
+
     N, T = size(logB)
-    R = promote_type(eltype(p), eltype(A), eltype(logB))
+    R = promote_type(eltype(init), eltype(trans), typeof(useless_logdensity_value))
     α = Matrix{R}(undef, N, T)
     β = Matrix{R}(undef, N, T)
     γ = Matrix{R}(undef, N, T)
     ξ = Array{R,3}(undef, N, N, T - 1)
     c = Vector{R}(undef, T)
+    logB = Matrix{R}(undef, N, T)
     logm = Vector{R}(undef, T)
     Bscaled = Matrix{R}(undef, N, T)
     Bβscaled = Matrix{R}(undef, N, T)
-    return ForwardBackwardStorage(α, β, γ, ξ, c, logm, Bscaled, Bβscaled)
+    return ForwardBackwardStorage(α, β, γ, ξ, c, logB, logm, Bscaled, Bβscaled)
+end
+
+function loglikelihoods!(
+    dists::AbstractVector, logB::AbstractMatrix, hmm::AbstractHMM, obs_seq
+)
+    for t in eachindex(obs_seqs, axes(logB, 2))
+        obs_distributions!(dists, hmm, t)
+        logB[:, t] .= logdensityof.(dists, Ref(obs_seq[t]))
+    end
+    check_no_nan(logB)
+    check_no_inf(logB)
+    return nothing
+end
+
+function loglikelihoods(hmm::AbstractHMM, obs_seq)
+    T, N = length(obs_seq), length(hmm)
+    dists = obs_distributions(hmm)
+    useless_logdensity_value = logdensityof(dists[1], obs_seq[1])
+    logB = Matrix{typeof(useless_logdensity_value)}(undef, N, T)
+    loglikelihoods!(dists, logB, hmm, obs_seq)
+    return logB
 end
 
 function initialize_forward_backward(hmm::AbstractHMM, logB)
-    p = initial_distribution(hmm)
[…]
