+++ test/allocations.jl
@@ -11,28 +11,29 @@ function test_allocations(hmm; T)
     @unpack state_seq, obs_seq = rand(hmm, T)
 
     ## Forward
-    logb = HMMs.loglikelihoods_vec(hmm, obs_seq[1])
+    logb = HiddenMarkovModels.loglikelihoods_vec(hmm, obs_seq[1])
     αₜ = zeros(N)
     αₜ₊₁ = zeros(N)
-    allocs = @ballocated HMMs.forward!($αₜ, $αₜ₊₁, $logb, $p, $A, $hmm, $obs_seq)
+    allocs = @ballocated HiddenMarkovModels.forward!(
+        $αₜ, $αₜ₊₁, $logb, $p, $A, $hmm, $obs_seq
+    )
     @test allocs == 0
 
     ## Viterbi
-    logb = HMMs.loglikelihoods_vec(hmm, obs_seq[1])
+    logb = HiddenMarkovModels.loglikelihoods_vec(hmm, obs_seq[1])
     δₜ = zeros(N)
     δₜ₋₁ = zeros(N)
     δA_tmp = zeros(N)
     ψ = zeros(Int, N, T)
     q = zeros(Int, T)
-    allocs = @ballocated HMMs.viterbi!(
+    allocs = @ballocated HiddenMarkovModels.viterbi!(
         $q, $δₜ, $δₜ₋₁, $δA_tmp, $ψ, $logb, $p, $A, $hmm, $obs_seq
     )
     @test allocs == 0
 
     ## Forward-backward
-    logB = HMMs.loglikelihoods(hmm, obs_seq)
-    fb = HMMs.initialize_forward_backward(p, A, logB)
-    allocs = @ballocated HMMs.forward_backward!($fb, $p, $A, $logB)
+    fb = HiddenMarkovModels.initialize_forward_backward(hmm, obs_seq)
+    allocs = @ballocated HiddenMarkovModels.forward_backward!($fb, $hmm, $obs_seq)
     @test allocs == 0
 end
 

[…]
