+++ v0.3.1/assets/benchmark/results/setup.txt
@@ -1,0 +1,54 @@
+Julia Version 1.9.3
+Commit bed2cd540a1 (2023-08-24 14:43 UTC)
+Build Info:
+  Official https://julialang.org/ release
+Platform Info:
+  OS: Linux (x86_64-linux-gnu)
+  CPU: 4 × AMD EPYC 7763 64-Core Processor
+  WORD_SIZE: 64
+  LIBM: libopenlibm
+  LLVM: libLLVM-14.0.6 (ORCJIT, znver3)
+  Threads: 1 on 4 virtual cores
+Environment:
+  JULIA_PKG_SERVER_REGISTRY_PREFERENCE = eager
+  JULIA_IMAGE_THREADS = 1
+  JULIA_PYTHONCALL_EXE = /home/runner/work/HiddenMarkovModels.jl/HiddenMarkovModels.jl/benchmark/.CondaPkg/env/bin/python
+
+# Multithreading
+
+Julia threads = 1
+OpenBLAS threads = 2
+Pytorch threads = 2
+
+# Julia packages
+
+Status `~/work/HiddenMarkovModels.jl/HiddenMarkovModels.jl/benchmark/Project.toml`
+  [6e4b80f9] BenchmarkTools v1.3.2
+  [336ed68f] CSV v0.10.11
+⌃ [992eb4ea] CondaPkg v0.2.18
+  [a93c6f00] DataFrames v1.6.1
+⌃ [31c24e10] Distributions v0.25.100
+  [b2b3ca75] HMMBase v1.0.7
+  [84ca31d5] HiddenMarkovModels v0.3.0 `..`
+  [32113eaa] PkgBenchmark v0.2.12
+  [91a5bcdd] Plots v1.39.0
+⌃ [6099a3de] PythonCall v0.9.14
+⌃ [295af30f] Revise v3.5.3
+  [ce78b400] SimpleUnPack v1.1.0
+  [b77e0a4c] InteractiveUtils
+  [37e2e46d] LinearAlgebra
+  [9a3f8284] Random
+Info Packages marked with ⌃ have new versions available and may be upgradable.
+
+# Python packages
+
+CondaPkg Status /home/runner/work/HiddenMarkovModels.jl/HiddenMarkovModels.jl/benchmark/CondaPkg.toml
+Environment
+  /home/runner/work/HiddenMarkovModels.jl/HiddenMarkovModels.jl/benchmark/.CondaPkg/env
+Packages
+  ipykernel v6.26.0
+  numpy v1.26.0
+  pytorch v2.0.0
+Pip packages
+  hmmlearn v0.3.0
+  pomegranate v1.0.3
