+++ ext/HiddenMarkovModelsChainRulesCoreExt.jl
@@ -9,7 +9,8 @@ using SimpleUnPack
 function _params_and_loglikelihoods(hmm::AbstractHMM, obs_seq)
     p = initialization(hmm)
     A = transition_matrix(hmm)
-    logB = HiddenMarkovModels.loglikelihoods(hmm, obs_seq)
+    d = obs_distributions(hmm)
+    logB = reduce(hcat, logdensityof.(d, Ref(obs)) for obs in obs_seq)
     return p, A, logB
 end
 
