+++ docs/make.jl
@@ -22,7 +22,14 @@ open(joinpath(joinpath(@__DIR__, "src"), "index.md"), "w") do io
     end
 end
 
-alt_pages = if get(ENV, "HMM_BENCHMARKS_DONE", false)
+benchmarks_done = (
+    length(readdir(joinpath(@__DIR__, "src", "assets", "benchmark", "plots"))) >
+    1 &  # plots present
+    length(readdir(joinpath(@__DIR__, "src", "assets", "benchmark", "results"))) >
+    1  # results present
+)
+
+alt_pages = if benchmarks_done
     ["Features" => "alt_features.md", "Performance" => "alt_performance.md"]
 else
     ["Features" => "alt_features.md"]
