+++ Project.toml
@@ -27,6 +27,7 @@ HiddenMarkovModelsHMMBaseExt = "HMMBase"
 ChainRulesCore = "1.16"
 DensityInterface = "0.4"
 Distributions = "0.25"
+DocStringExtensions = "0.9"
 LinearAlgebra = "1.6"
 PrecompileTools = "1.1"
 Random = "1.6"
