### 2023-11-03 145df1c
Fix docs benchmark (#58)
