### 2023-11-06 2d343e2
Finish BW storage
