+++ src/utils/probvec.jl
@@ -4,6 +4,11 @@ end
 
 sum_to_one!(x) = x ./= sum(x)
 
+"""
+    rand_prob_vec([rng,] N)
+
+Generate a random probability distribution of size `N`.
+"""
 function rand_prob_vec(rng::AbstractRNG, N)
     p = rand(rng, N)
     sum_to_one!(p)
