+++ docs/src/api.md
@@ -10,7 +10,6 @@ HiddenMarkovModels
 AbstractHiddenMarkovModel
 HiddenMarkovModel
 AbstractHMM
-PermutedHMM
 HMM
 ```
 
@@ -38,8 +37,15 @@ forward_backward
 
 ```@docs
 fit!
-fit
 baum_welch
+```
+
+## Misc
+
+```@docs
+check_hmm
+rand_prob_vec
+rand_trans_mat
 ```
 
 ## Internals
@@ -51,6 +57,7 @@ HiddenMarkovModels.ForwardBackwardStorage
 HiddenMarkovModels.BaumWelchStorage
 HiddenMarkovModels.fit_element_from_sequence!
 HiddenMarkovModels.LightDiagNormal
+HiddenMarkovModels.PermutedHMM
 ```
 
 ## Notations
