### 2023-11-06 51fca9c
Separate single sequence case
