+++ src/inference/baum_welch.jl
@@ -1,147 +1,123 @@
-"""
-$(TYPEDEF)
-
-Store Baum-Welch quantities with element type `R` and observation type `O`.
-
-Unlike the other storage types, this one is relative to multiple sequences.
-
-# Fields
-
-$(TYPEDFIELDS)
-"""
-struct BaumWelchStorage{R,O}
-    "one `ForwardBackwardStorage` for each observation sequence"
-    fbs::Vector{ForwardBackwardStorage{R}}
-    "number of iterations performed"
-    iteration::RefValue{Int}
-    "history of total loglikelihood values throughout the algorithm"
-    logL_evolution::Vector{R}
-    "concatenation of `γ` matrices for all observation sequences (useful to avoid allocations in fitting)"
-    state_marginals_concat::Matrix{R}
-    "concatenation of observation sequences (useful to avoid allocations in fitting)"
-    obs_seqs_concat::Vector{O}
-    "temporal limits of each observation sequence in the concatenations"
-    limits::Vector{Int}
-end
-
-function initialize_baum_welch(
-    hmm::AbstractHMM, obs_seqs::Vector{<:Vector}, nb_seqs::Integer; max_iterations::Integer
+function baum_welch_has_converged(
+    logL_evolution::Vector; atol::Real, loglikelihood_increasing::Bool
 )
-    if nb_seqs != length(obs_seqs)
-        throw(ArgumentError("nb_seqs != length(obs_seqs)"))
+    if length(logL_evolution) >= 2
+        logL, logL_prev = logL_evolution[end], logL_evolution[end - 1]
+        progress = logL - logL_prev
+        if loglikelihood_increasing && progress < 0
+            error("Loglikelihood decreased in Baum-Welch")
+        elseif progress < atol
+            return true
+        end
     end
-    N, T = length(hmm), sum(length, obs_seqs)
-    R = eltype(hmm, obs_seqs[1][1])
-    fbs = [initialize_forward_backward(hmm, obs_seqs[k]) for k in eachindex(obs_seqs)]
-    iteration = Ref(0)
-    logL_evolution = Vector{R}(undef, max_iterations)
-    state_marginals_concat = Matrix{R}(undef, N, T)
-    obs_seqs_concat = reduce(vcat, obs_seqs)
-    limits = vcat(0, cumsum(length.(obs_seqs)))
-    return BaumWelchStorage(
-        fbs, iteration, logL_evolution, state_marginals_concat, obs_seqs_concat, limits
-    )
+    return false
 end
 
 function baum_welch!(
+    fb::ForwardBackwardStorage,
+    logL_evolution::Vector,
     hmm::AbstractHMM,
-    bw::BaumWelchStorage,
-    obs_seqs::Vector{<:Vector};
+    obs_seq::Vector;
     atol::Real,
     max_iterations::Integer,
-    check_loglikelihood_increasing::Bool,
+    loglikelihood_increasing::Bool,
 )
-    @unpack (
-        fbs, iteration, logL_evolution, state_marginals_concat, obs_seqs_concat, limits
-    ) = bw
-    iteration[] = 0
+    for _ in 1:max_iterations
+        forward_backward!(fb, hmm, obs_seq)
+        push!(logL_evolution, fb.logL[])
+        fit!(hmm, (fb,), (obs_seq,), fb, obs_seq)
+        if baum_welch_has_converged(logL_evolution; atol, loglikelihood_increasing)
+            break
+        end
+    end
+    return nothing
+end
 
-    while iteration[] < max_iterations
-        # E step
+function baum_welch!(
+    fbs::Vector{<:ForwardBackwardStorage},
+    fb_concat::ForwardBackwardStorage,
[…]
