+++ src/types/hmm.jl
@@ -41,8 +41,9 @@ initialization(hmm::HMM) = hmm.init
 transition_matrix(hmm::HMM) = hmm.trans
 obs_distributions(hmm::HMM) = hmm.dists
 
-function StatsAPI.fit!(hmm::HMM, bw::BaumWelchStorage)
-    @unpack fbs, state_marginals_concat, obs_seqs_concat = bw
+function StatsAPI.fit!(
+    hmm::HMM, fbs, obs_seqs, fb_concat::ForwardBackwardStorage, obs_seqs_concat::Vector
+)
     # Initialization
     hmm.init .= zero(eltype(hmm.init))
     for k in eachindex(fbs)
@@ -51,15 +52,11 @@ function StatsAPI.fit!(hmm::HMM, bw::BaumWelchStorage)
     sum_to_one!(hmm.init)
     # Transition matrix
     hmm.trans .= zero(eltype(hmm.trans))
-    for k in eachindex(fbs)
-        sum!(hmm.trans, fbs[k].ξ; init=false)
-    end
+    sum!(hmm.trans, fb_concat.ξ; init=false)
     foreach(sum_to_one!, eachrow(hmm.trans))
     #  Observation distributions
     for i in eachindex(hmm.dists)
-        fit_element_from_sequence!(
-            hmm.dists, i, obs_seqs_concat, view(state_marginals_concat, i, :)
-        )
+        fit_element_from_sequence!(hmm.dists, i, obs_seqs_concat, view(fb_concat.γ, i, :))
     end
     check_hmm(hmm)
     return nothing
