+++ src/utils/check.jl
@@ -65,3 +65,9 @@ function check_hmm(hmm::AbstractHMM)
     check_dists(d)
     return nothing
 end
+
+function check_lengths(obs_seqs::Vector{<:Vector}, nb_seqs::Integer)
+    if nb_seqs != length(obs_seqs)
+        throw(ArgumentError("nb_seqs != length(obs_seqs)"))
+    end
+end
