### 2023-11-09 d12ea1f
Reactivate precompile workload and fix viterbi
