+++ src/HiddenMarkovModels.jl
@@ -67,31 +67,19 @@ if !isdefined(Base, :get_extension)
     end
 end
 
-# @compile_workload begin
-#     N, D, T = 5, 3, 100
-#     p = rand_prob_vec(N)
-#     A = rand_trans_mat(N)
-#     dists = [LightDiagNormal(randn(D), ones(D)) for i in 1:N]
-#     hmm = HMM(p, A, dists)
+@compile_workload begin
+    N, D, T = 3, 2, 100
+    p = rand_prob_vec(N)
+    A = rand_trans_mat(N)
+    dists = [LightDiagNormal(randn(D), ones(D)) for i in 1:N]
+    hmm = HMM(p, A, dists)
+    obs_seq = rand(hmm, T).obs_seq
 
-#     obs_seq = rand(hmm, T).obs_seq
-#     obs_seqs = [rand(hmm, T).obs_seq for _ in 1:3]
-#     nb_seqs = 3
-
-#     logdensityof(hmm, obs_seq)
-#     logdensityof(hmm, obs_seqs, nb_seqs)
-
-#     forward(hmm, obs_seq)
-#     forward(hmm, obs_seqs, nb_seqs)
-
-#     viterbi(hmm, obs_seq)
-#     viterbi(hmm, obs_seqs, nb_seqs)
-
-#     forward_backward(hmm, obs_seq)
-#     forward_backward(hmm, obs_seqs, nb_seqs)
-
-#     baum_welch(hmm, obs_seq; max_iterations=2, atol=-Inf)
-#     baum_welch(hmm, obs_seqs, nb_seqs; max_iterations=2, atol=-Inf)
-# end
+    logdensityof(hmm, obs_seq)
+    forward(hmm, obs_seq)
+    viterbi(hmm, obs_seq)
+    forward_backward(hmm, obs_seq)
+    baum_welch(hmm, obs_seq; max_iterations=2, atol=-Inf)
+end
 
 end
