### 2023-11-09 0142f6d
Reactivate sparse allocation test
