+++ test/arrays.jl
@@ -22,11 +22,10 @@ hmm_init = HMM(p, A, d_init);
 
 obs_seq = rand(hmm, T).obs_seq;
 
-γ, ξ, logL = forward_backward(hmm, obs_seq);
+γ, logL = forward_backward(hmm, obs_seq);
 hmm_est, logL_evolution = @inferred baum_welch(hmm_init, obs_seq);
 
 @testset "Sparse" begin
-    @test eltype(ξ) <: AbstractSparseArray
     @test typeof(hmm_est) == typeof(hmm_init)
     @test nnz(transition_matrix(hmm_est)) <= nnz(transition_matrix(hmm))
 end
@@ -42,10 +41,9 @@ hmm = HMM(p, A, d);
 hmm_init = HMM(p, A, d_init);
 obs_seq = rand(hmm, T).obs_seq;
 
-γ, ξ, logL = forward_backward(hmm, obs_seq);
+γ, logL = forward_backward(hmm, obs_seq);
 hmm_est, logL_evolution = @inferred baum_welch(hmm_init, obs_seq);
 
 @testset "Static" begin
-    @test eltype(ξ) <: StaticArray
     @test typeof(hmm_est) == typeof(hmm_init)
 end
