### 2023-11-10 c68f8e9
Fix tests for 1.6
