### 2023-11-10 0b1eef1
Complete revamp (#59)
* Revamp

* More changes

* Correctness and type stability

* Allocs

* BaumWelch storage

* Finish BW storage

* Fix docs

* Separate single sequence case

* Finish revamp except custom tuto

* Reactivate precompile workload and fix viterbi

* Reactivate sparse allocation test

* Fix FB output

* Skip allocations test sparse

* Fix tests for 1.6
