+++ test/quality.jl
@@ -1,4 +1,0 @@
-using Aqua: test_all
-using HiddenMarkovModels
-
-test_all(HiddenMarkovModels; ambiguities=false)
