+++ src/utils/fit.jl
@@ -14,23 +14,23 @@ end
 function fit_element_from_sequence!(
     dists::AbstractVector{D}, i::Integer, x::AbstractVector, w::AbstractVector
 ) where {D<:Distribution}
-    dists[i] = _fit_from_sequence(D, x, w)
+    dists[i] = fit_from_sequence(D, x, w)
     return nothing
 end
 
-function _fit_from_sequence(
+function fit_from_sequence(
     ::Type{D}, x_nums::AbstractVector, w::AbstractVector
 ) where {D<:UnivariateDistribution}
     return fit(D, x_nums, w)
 end
 
-function _fit_from_sequence(
+function fit_from_sequence(
     ::Type{D}, x_vecs::AbstractVector{<:AbstractVector}, w::AbstractVector
 ) where {D<:MultivariateDistribution}
     return fit(D, reduce(hcat, x_vecs), w)
 end
 
-function _fit_from_sequence(
+function fit_from_sequence(
     ::Type{D}, x_mats::AbstractVector{<:AbstractMatrix}, w::AbstractVector
 ) where {D<:MatrixDistribution}
     return fit(D, reduce(dcat, x_mats), w)
