+++ benchmark/benchmarks.jl
@@ -1,16 +1,13 @@
-using Pkg
-Pkg.activate(@__DIR__)
-
-using Revise
-includet("utils/suite.jl")
+using HMMBenchmark
+using BenchmarkTools
 
 SUITE = define_suite(;
     implems=("HMMs.jl",),
     algos=("logdensity", "viterbi", "forward_backward", "baum_welch"),
     N_vals=2:2:20,
     D_vals=10,
-    T_vals=1000,
-    K_vals=100,
+    T_vals=100,
+    K_vals=10,
     I=10,
 )
 

[…]
