+++ src/inference/forward_backward.jl
@@ -7,7 +7,7 @@ This storage is relative to a single sequence.
 
 # Fields
 
-The only fields useful outside of the algorithm are `γ`, `ξ` and `logL`.
+The only fields useful outside of the algorithm are `γ` and `logL`, as well as `init_count` and `trans_count`.
 
 $(TYPEDFIELDS)
 """
@@ -20,18 +20,20 @@ struct ForwardBackwardStorage{R,M<:AbstractMatrix{R}}
     β::Matrix{R}
     "posterior state marginals `γ[i,t] = ℙ(X[t]=i | Y[1:T])`"
     γ::Matrix{R}
-    "posterior transition marginals `ξ[t][i,j] = ℙ(X[t:t+1]=(i,j) | Y[1:T])`"
-    ξ::Vector{M}
     "forward message inverse normalizations `c[t] = 1 / sum(α[:,t])`"
     c::Vector{R}
     "observation loglikelihoods `logB[i,t] = ℙ(Y[t] | X[t]=i)`"
     logB::Matrix{R}
     "maximum of the observation loglikelihoods `logm[t] = maximum(logB[:, t])`"
     logm::Vector{R}
-    "numerically stabilized observation likelihoods `B̃[i,t] = exp.(logB[i,t] - logm[t])`"
-    B̃::Matrix{R}
-    "product `B̃β[i,t] = B̃[i,t] * β[i,t]`"
-    B̃β::Matrix{R}
+    "numerically stabilized observation likelihoods `B[i,t] = exp.(logB[i,t] - logm[t])`"
+    B::Matrix{R}
+    "product `Bβ[i,t] = B[i,t] * β[i,t]`"
+    Bβ::Matrix{R}
+    "posterior initialization count"
+    init_count::Vector{R}
+    "posterior transition count"
+    trans_count::M
 end
 
 Base.eltype(::ForwardBackwardStorage{R}) where {R} = R
@@ -41,53 +43,47 @@ function initialize_forward_backward(hmm::AbstractHMM, obs_seq::Vector)
     N, T = length(hmm), length(obs_seq)
     A = transition_matrix(hmm)
     R = eltype(hmm, obs_seq[1])
-    M = typeof(similar(A, R))
 
     logL = RefValue{R}(zero(R))
     α = Matrix{R}(undef, N, T)
     β = Matrix{R}(undef, N, T)
     γ = Matrix{R}(undef, N, T)
-    ξ = Vector{M}(undef, T - 1)
-    for t in 1:(T - 1)
-        ξ[t] = similar(A, R)
-    end
     c = Vector{R}(undef, T)
     logB = Matrix{R}(undef, N, T)
     logm = Vector{R}(undef, T)
-    B̃ = Matrix{R}(undef, N, T)
-    B̃β = Matrix{R}(undef, N, T)
-
-    return ForwardBackwardStorage{R,M}(logL, α, β, γ, ξ, c, logB, logm, B̃, B̃β)
+    B = Matrix{R}(undef, N, T)
+    Bβ = Matrix{R}(undef, N, T)
+    init_count = Vector{R}(undef, N)
+    trans_count = similar(A, R)
+    M = typeof(trans_count)
+    return ForwardBackwardStorage{R,M}(
+        logL, α, β, γ, c, logB, logm, B, Bβ, init_count, trans_count
+    )
 end
 
 function update_likelihoods!(fb::ForwardBackwardStorage, hmm::AbstractHMM, obs_seq::Vector)
-    d = obs_distributions(hmm)
-    @unpack logB, logm, B̃ = fb
-    N, T = length(hmm), duration(fb)
-
[…]
