+++ src/utils/check.jl
@@ -56,13 +56,11 @@ Verify that `hmm` satisfies basic assumptions.
 function check_hmm(hmm::AbstractHMM)
     p = initialization(hmm)
     A = transition_matrix(hmm)
-    d = obs_distributions(hmm)
-    if !all(==(length(hmm)), (length(p), size(A, 1), size(A, 2), length(d)))
+    if !all(==(length(hmm)), (length(p), size(A, 1), size(A, 2)))
         throw(DimensionMismatch("Incoherent sizes"))
     end
     check_prob_vec(p)
     check_trans_mat(A)
-    check_dists(d)
     return nothing
 end
 
