+++ test/allocations.jl
@@ -1,6 +1,6 @@
 using BenchmarkTools
 using HiddenMarkovModels
-using HiddenMarkovModels.Test
+using HiddenMarkovModels.HMMTest
 import HiddenMarkovModels as HMMs
 using Test
 
