+++ src/inference/forward_backward.jl
@@ -7,7 +7,7 @@ This storage is relative to a single sequence.
 
 # Fields
 
-The only fields useful outside of the algorithm are `γ`, `logL`, `init_count` and `trans_count`.
+The only fields useful outside of the algorithm are `γ`, `logL`, `init_count` and `trans_count`, the rest does not belong to the public API.
 
 $(TYPEDFIELDS)
 """
@@ -20,6 +20,8 @@ struct ForwardBackwardStorage{R,M<:AbstractMatrix{R}}
     β::Matrix{R}
     "posterior state marginals `γ[i,t] = ℙ(X[t]=i | Y[1:T])`"
     γ::Matrix{R}
+    "posterior transition marginals `ξ[t][i,j] = ℙ(X[t]=i, X[t+1]=j | Y[1:T])`"
+    ξ::Vector{M}
     "forward message inverse normalizations `c[t] = 1 / sum(α[:,t])`"
     c::Vector{R}
     "observation loglikelihoods `logB[i,t] = ℙ(Y[t] | X[t]=i)`"
@@ -30,39 +32,43 @@ struct ForwardBackwardStorage{R,M<:AbstractMatrix{R}}
     B::Matrix{R}
     "product `Bβ[i,t] = B[i,t] * β[i,t]`"
     Bβ::Matrix{R}
-    "posterior initialization count"
-    init_count::Vector{R}
-    "posterior transition count"
-    trans_count::M
 end
 
+"""
+    initialize_forward_backward(hmm, obs_seq)
+"""
 function initialize_forward_backward(hmm::AbstractHMM, obs_seq::Vector)
     N, T = length(hmm), length(obs_seq)
     A = transition_matrix(hmm)
     R = eltype(hmm, obs_seq[1])
+    M = typeof(similar(A, R))
 
     logL = RefValue{R}(zero(R))
     α = Matrix{R}(undef, N, T)
     β = Matrix{R}(undef, N, T)
     γ = Matrix{R}(undef, N, T)
+    ξ = Vector{M}(undef, T - 1)
+    for t in 1:(T - 1)
+        ξ[t] = similar(A, R)
+    end
     c = Vector{R}(undef, T)
     logB = Matrix{R}(undef, N, T)
     logm = Vector{R}(undef, T)
     B = Matrix{R}(undef, N, T)
     Bβ = Matrix{R}(undef, N, T)
-    init_count = Vector{R}(undef, N)
-    trans_count = similar(A, R)
-    M = typeof(trans_count)
-    return ForwardBackwardStorage{R,M}(
-        logL, α, β, γ, c, logB, logm, B, Bβ, init_count, trans_count
-    )
+    return ForwardBackwardStorage{R,M}(logL, α, β, γ, ξ, c, logB, logm, B, Bβ)
 end
 
-function forward_backward!(fb::ForwardBackwardStorage, hmm::AbstractHMM, obs_seq::Vector)
+"""
+    forward_backward!(storage, hmm, obs_seq)
+"""
+function forward_backward!(
+    storage::ForwardBackwardStorage, hmm::AbstractHMM, obs_seq::Vector
+)
     p = initialization(hmm)
     A = transition_matrix(hmm)
     T = length(obs_seq)
-    @unpack α, β, c, γ, logB, logm, B, Bβ, init_count, trans_count = fb
+    @unpack logL, α, β, c, γ, ξ, logB, logm, B, Bβ = storage
 
     # Observation loglikelihoods
     for (logb, obs) in zip(eachcol(logB), obs_seq)
@@ -97,53 +103,33 @@ function forward_backward!(fb::ForwardBackwardStorage, hmm::AbstractHMM, obs_seq
     # Marginals
     γ .= α .* β ./ c'
     check_finite(γ)
-
-    # Sufficient stats
[…]
