+++ src/inference/baum_welch.jl
@@ -22,6 +22,15 @@ struct BaumWelchStorage{R,M<:AbstractMatrix{R}}
     limits::Vector{Int}
 end
 
+function check_nb_seqs(obs_seqs::Vector{<:Vector}, nb_seqs::Integer)
+    if nb_seqs != length(obs_seqs)
+        throw(ArgumentError("Incoherent sizes provided: `nb_seqs != length(obs_seqs)`"))
+    end
+end
+
+"""
+    initialize_baum_welch(hmm, obs_seqs, nb_seqs)
+"""
 function initialize_baum_welch(
     hmm::AbstractHMM, obs_seqs::Vector{<:Vector}, nb_seqs::Integer
 )
@@ -36,6 +45,9 @@ function initialize_baum_welch(
     return BaumWelchStorage(init_count, trans_count, state_marginals_concat, limits)
 end
 
+"""
+    initialize_logL_evolution(hmm, obs_seqs, nb_seqs; max_iterations)
+"""
 function initialize_logL_evolution(
     hmm::AbstractHMM, obs_seqs::Vector{<:Vector}, nb_seqs::Integer; max_iterations::Integer
 )
@@ -47,16 +59,19 @@ function initialize_logL_evolution(
 end
 
 function update_sufficient_statistics!(
-    bw::BaumWelchStorage{R}, fbs::Vector{<:ForwardBackwardStorage}
+    bw_storage::BaumWelchStorage{R}, fb_storages::Vector{<:ForwardBackwardStorage}
 ) where {R}
-    @unpack init_count, trans_count, state_marginals_concat, limits = bw
+    @unpack init_count, trans_count, state_marginals_concat, limits = bw_storage
     init_count .= zero(R)
     trans_count .= zero(R)
     state_marginals_concat .= zero(R)
-    for k in eachindex(fbs)  # TODO: ThreadsX?
-        init_count .+= fbs[k].init_count
-        mynonzeros(trans_count) .+= mynonzeros(fbs[k].trans_count)
-        state_marginals_concat[:, (limits[k] + 1):limits[k + 1]] .= fbs[k].γ
+    for k in eachindex(fb_storages)  # TODO: ThreadsX?
+        @unpack γ, ξ = fb_storages[k]
+        init_count .+= @view γ[:, 1]
+        for t in eachindex(ξ)
+            mynonzeros(trans_count) .+= mynonzeros(ξ[t])
+        end
+        state_marginals_concat[:, (limits[k] + 1):limits[k + 1]] .= γ
     end
     return nothing
 end
@@ -76,15 +91,23 @@ function baum_welch_has_converged(
     return false
 end
 
-function StatsAPI.fit!(hmm::AbstractHMM, bw::BaumWelchStorage, obs_seqs_concat::Vector)
-    return fit!(
-        hmm, bw.init_count, bw.trans_count, obs_seqs_concat, bw.state_marginals_concat
-    )
+function StatsAPI.fit!(
+    hmm::AbstractHMM, bw_storage::BaumWelchStorage, obs_seqs_concat::Vector
+)
+    @unpack init_count, trans_count, state_marginals_concat = bw_storage
+    return fit!(hmm, init_count, trans_count, obs_seqs_concat, state_marginals_concat)
 end
 
+"""
+    baum_welch!(
+        fb_storages, bw_storage, logL_evolution,
+        hmm, obs_seqs, obs_seqs_concat;
+        atol, max_iterations, loglikelihood_increasing
+    )
+"""
 function baum_welch!(
-    fbs::Vector{<:ForwardBackwardStorage},
-    bw::BaumWelchStorage,
+    fb_storages::Vector{<:ForwardBackwardStorage},
+    bw_storage::BaumWelchStorage,
[…]
