+++ test/arrays.jl
@@ -19,14 +19,12 @@ d_init = [Normal(i + randn(), 1.0) for i in 1:N];
 
 hmm = HMM(p, A, d);
 hmm_init = HMM(p, A, d_init);
-
 obs_seq = rand(hmm, T).obs_seq;
 
-γ, logL = forward_backward(hmm, obs_seq);
-hmm_est, logL_evolution = @inferred baum_welch(hmm_init, obs_seq);
-
 @testset "Sparse" begin
-    @test typeof(hmm_est) == typeof(hmm_init)
+    γ, ξ, logL = @inferred forward_backward(hmm, obs_seq)
+    hmm_est, logL_evolution = @inferred baum_welch(hmm_init, obs_seq)
+    @test eltype(ξ) == typeof(transition_matrix(hmm))
     @test nnz(transition_matrix(hmm_est)) <= nnz(transition_matrix(hmm))
 end
 
@@ -41,9 +39,8 @@ hmm = HMM(p, A, d);
 hmm_init = HMM(p, A, d_init);
 obs_seq = rand(hmm, T).obs_seq;
 
-γ, logL = forward_backward(hmm, obs_seq);
-hmm_est, logL_evolution = @inferred baum_welch(hmm_init, obs_seq);
-
 @testset "Static" begin
-    @test typeof(hmm_est) == typeof(hmm_init)
+    γ, ξ, logL = @inferred forward_backward(hmm, obs_seq)
+    hmm_est, logL_evolution = @inferred baum_welch(hmm_init, obs_seq)
+    @test eltype(ξ) == typeof(transition_matrix(hmm))
 end

[…]
