### 2023-11-11 39be362
Remove obs_sample
