### 2023-11-11 fffad07
Minor fixes (#60)
* Remove obs_distributions from interface

* Fix stuff

* Minor fixes

* Stuff

* Change dependecy status ChainRulesCore

* CRC compat

* Remove obs_sample
