+++ test/correctness.jl
@@ -6,58 +6,45 @@ using SimpleUnPack
 using Test
 
 function test_correctness(hmm, hmm_init; T)
-    obs_seq1 = rand(hmm, T).obs_seq
-    obs_seq2 = rand(hmm, T).obs_seq
-    obs_mat1 = collect(reduce(hcat, obs_seq1)')
-    obs_mat2 = collect(reduce(hcat, obs_seq2)')
-
-    nb_seqs = 2
-    obs_seqs = [obs_seq1, obs_seq2]
+    obs_seq = rand(hmm, T).obs_seq
+    obs_mat = collect(reduce(hcat, obs_seq)')
 
     hmm_base = HMMBase.HMM(deepcopy(hmm))
     hmm_init_base = HMMBase.HMM(deepcopy(hmm_init))
 
     @testset "Logdensity" begin
-        logL1_base = HMMBase.forward(hmm_base, obs_mat1)[2]
-        logL2_base = HMMBase.forward(hmm_base, obs_mat2)[2]
-        logL = logdensityof(hmm, obs_seqs, nb_seqs)
-        @test logL ≈ logL1_base + logL2_base
+        logL_base = HMMBase.forward(hmm_base, obs_mat)[2]
+        logL = logdensityof(hmm, obs_seq)
+        @test logL ≈ logL_base
     end
 
     @testset "Forward" begin
-        (α1_base, logL1_base), (α2_base, logL2_base) = [
-            HMMBase.forward(hmm_base, obs_mat1), HMMBase.forward(hmm_base, obs_mat2)
-        ]
-        (α1, logL1), (α2, logL2) = forward(hmm, obs_seqs, nb_seqs)
-        @test isapprox(α1, α1_base[end, :])
-        @test isapprox(α2, α2_base[end, :])
-        @test logL1 ≈ logL1_base
-        @test logL2 ≈ logL2_base
+        α_base, logL_base = HMMBase.forward(hmm_base, obs_mat)
+        α, logL = forward(hmm, obs_seq)
+        @test isapprox(α, α_base[end, :])
+        @test logL ≈ logL_base
     end
 
     @testset "Viterbi" begin
-        q1_base = HMMBase.viterbi(hmm_base, obs_mat1)
-        q2_base = HMMBase.viterbi(hmm_base, obs_mat2)
-        q1, q2 = viterbi(hmm, obs_seqs, nb_seqs)
-        @test isequal(q1, q1_base)
-        @test isequal(q2, q2_base)
+        q_base = HMMBase.viterbi(hmm_base, obs_mat)
+        q = viterbi(hmm, obs_seq)
+        # Viterbi decoding can vary in case of (infrequent) ties
+        @test mean(q .== q_base) > 0.9
     end
 
     @testset "Forward-backward" begin
-        γ1_base = HMMBase.posteriors(hmm_base, obs_mat1)
-        γ2_base = HMMBase.posteriors(hmm_base, obs_mat2)
-        (γ1, _), (γ2, _) = forward_backward(hmm, obs_seqs, nb_seqs)
-        @test isapprox(γ1, γ1_base')
-        @test isapprox(γ2, γ2_base')
+        γ_base = HMMBase.posteriors(hmm_base, obs_mat)
+        γ, _, _ = forward_backward(hmm, obs_seq)
+        @test isapprox(γ, γ_base')
     end
 
     @testset "Baum-Welch" begin
         hmm_est_base, hist_base = HMMBase.fit_mle(
-            hmm_init_base, obs_mat1; maxiter=10, tol=-Inf
+            hmm_init_base, obs_mat; maxiter=10, tol=-Inf
         )
         logL_evolution_base = hist_base.logtots
[…]
