+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"debugging/#Debugging","page":"Debugging","title":"Debugging","text":"","category":"section"},{"location":"debugging/#Numerical-overflow","page":"Debugging","title":"Numerical overflow","text":"","category":"section"},{"location":"debugging/","page":"Debugging","title":"Debugging","text":"The most frequent error you will encounter is an OverflowError during forward-backward, telling you that \"some values are infinite / NaN\". This can happen for a variety of reasons, so here are a few leads worth investigating:","category":"page"},{"location":"debugging/","page":"Debugging","title":"Debugging","text":"Increase the duration of the sequence / the number of sequences (to get more data)\nReduce the number of states (to make every one of them useful)\nAdd a prior to your transition matrix / observation distributions (to avoid degenerate behavior like zero variance in a Gaussian)\nPick a better initialization (to start closer to the supposed ground truth)\nUse LogarithmicNumbers.jl in strategic places (to guarantee numerical stability). Note that these numbers don't play nicely with Distributions.jl, so you may have to roll out your own observation distribution.","category":"page"},{"location":"api/#API-reference","page":"API reference","title":"API reference","text":"","category":"section"},{"location":"api/","page":"API reference","title":"API reference","text":"HiddenMarkovModels","category":"page"},{"location":"api/#HiddenMarkovModels","page":"API reference","title":"HiddenMarkovModels","text":"HiddenMarkovModels\n\nA Julia package for HMM modeling, simulation, inference and learning.\n\nExports\n\nAbstractHMM\nAbstractHiddenMarkovModel\nHMM\nHiddenMarkovModel\nbaum_welch\ncheck_hmm\nfit!\nforward\nforward_backward\ninitialization\nlogdensityof\nobs_distributions\nrand_prob_vec\nrand_trans_mat\ntransition_matrix\nviterbi\n\n\n\n\n\n","category":"module"},{"location":"api/#Types","page":"API reference","title":"Types","text":"","category":"section"},{"location":"api/","page":"API reference","title":"API reference","text":"AbstractHiddenMarkovModel\nHiddenMarkovModel\nAbstractHMM\nHMM","category":"page"},{"location":"api/#HiddenMarkovModels.AbstractHiddenMarkovModel","page":"API reference","title":"HiddenMarkovModels.AbstractHiddenMarkovModel","text":"AbstractHiddenMarkovModel\n\nAbstract supertype for an HMM amenable to simulation, inference and learning.\n\nInterface\n\nTo create your own subtype of AbstractHiddenMarkovModel, you need to implement the following methods:\n\nlength(hmm)\neltype(hmm, obs)\ninitialization(hmm)\ntransition_matrix(hmm)\nobs_distributions(hmm)\nfit!(hmm, init_count, trans_count, obs_seq, state_marginals) (optional)\n\nApplicable functions\n\nAny HMM object which satisfies the interface can be given as input to the following functions:\n\nrand(rng, hmm, T)\nlogdensityof(hmm, obs_seq)\nforward(hmm, obs_seq)\nviterbi(hmm, obs_seq)\nforward_backward(hmm, obs_seq)\nbaum_welch(hmm, obs_seq) (if the optional fit! is implemented)\n\n\n\n\n\n","category":"type"},{"location":"api/#HiddenMarkovModels.HiddenMarkovModel","page":"API reference","title":"HiddenMarkovModels.HiddenMarkovModel","text":"struct HiddenMarkovModel{I<:(AbstractVector), T<:(AbstractMatrix), D<:(AbstractVector)} <: AbstractHiddenMarkovModel\n\n
[…]
