### 2023-11-11 7994cbb
Reactivate multithreaded BW (#62)
