### 2023-11-14 46c7dc5
Put benchmarks inside HMMBenchmark subpackage (#63)
* Move benchmarks

* Docs build on tags
