### 2023-11-24 #5874 @arfon
:wave: @jbytecode, would you be willing to edit this submission for JOSS?
