### 2023-11-24 #5874 @gdalle
Hi! Since the submission was in the backlog, I took the opportunity to revamp the package significantly, and as a result the paper and benchmarks are outdated. Can we maybe put this submission on hold? Or is it easier for your logistics if I withdraw and resubmit later?
