### 2023-11-24 #5874 @jbytecode
@gdalle - How much time do you need to reach the final version of the manuscript?
