+++ src/inference/forward_backward.jl
@@ -7,7 +7,7 @@ This storage is relative to a single sequence.
 
 # Fields
 
-The only fields useful outside of the algorithm are `γ`, `logL`, `init_count` and `trans_count`, the rest does not belong to the public API.
+The only fields useful outside of the algorithm are `γ`, `ξ`, `logL`, the rest does not belong to the public API.
 
 $(TYPEDFIELDS)
 """
@@ -18,61 +18,66 @@ struct ForwardBackwardStorage{R,M<:AbstractMatrix{R}}
     α::Matrix{R}
     "scaled backward messsages `β[i,t]` proportional to `ℙ(Y[t+1:T] | X[t]=i)` (up to a function of `t`)"
     β::Matrix{R}
+    "forward message inverse normalizations `c[t] = 1 / sum(α[:,t])`"
+    c::Vector{R}
     "posterior state marginals `γ[i,t] = ℙ(X[t]=i | Y[1:T])`"
     γ::Matrix{R}
     "posterior transition marginals `ξ[t][i,j] = ℙ(X[t]=i, X[t+1]=j | Y[1:T])`"
     ξ::Vector{M}
-    "forward message inverse normalizations `c[t] = 1 / sum(α[:,t])`"
-    c::Vector{R}
     "observation loglikelihoods `logB[i,t] = ℙ(Y[t] | X[t]=i)`"
     logB::Matrix{R}
     "maximum of the observation loglikelihoods `logm[t] = maximum(logB[:, t])`"
     logm::Vector{R}
-    "numerically stabilized observation likelihoods `B[i,t] = exp.(logB[i,t] - logm[t])`"
+    "numerically stabilized observation likelihoods `B[i,t] = exp(logB[i,t] - logm[t])`"
     B::Matrix{R}
-    "product `Bβ[i,t] = B[i,t] * β[i,t]`"
-    Bβ::Matrix{R}
+    "scratch storage space"
+    scratch::Vector{R}
 end
 
 """
     initialize_forward_backward(hmm, obs_seq)
 """
-function initialize_forward_backward(hmm::AbstractHMM, obs_seq::Vector)
+function initialize_forward_backward(
+    hmm::AbstractHMM, obs_seq::Vector; transition_marginals=true
+)
     N, T = length(hmm), length(obs_seq)
-    A = transition_matrix(hmm)
+    A = transition_matrix(hmm, 1)
     R = eltype(hmm, obs_seq[1])
     M = typeof(similar(A, R))
 
     logL = RefValue{R}(zero(R))
     α = Matrix{R}(undef, N, T)
     β = Matrix{R}(undef, N, T)
+    c = Vector{R}(undef, T)
     γ = Matrix{R}(undef, N, T)
     ξ = Vector{M}(undef, T - 1)
-    for t in 1:(T - 1)
-        ξ[t] = similar(A, R)
+    if transition_marginals
+        for t in 1:(T - 1)
+            ξ[t] = similar(transition_matrix(hmm, t), R)
+        end
     end
-    c = Vector{R}(undef, T)
     logB = Matrix{R}(undef, N, T)
     logm = Vector{R}(undef, T)
     B = Matrix{R}(undef, N, T)
-    Bβ = Matrix{R}(undef, N, T)
-    return ForwardBackwardStorage{R,M}(logL, α, β, γ, ξ, c, logB, logm, B, Bβ)
+    scratch = Vector{R}(undef, N)
+    return ForwardBackwardStorage{R,M}(logL, α, β, c, γ, ξ, logB, logm, B, scratch)
 end
 
 """
     forward_backward!(storage, hmm, obs_seq)
 """
 function forward_backward!(
-    storage::ForwardBackwardStorage, hmm::AbstractHMM, obs_seq::Vector
[…]
