+++ src/utils/lightdiagnormal.jl
@@ -1,7 +1,7 @@
 """
 $(TYPEDEF)
 
-An HMMs-compatible implementation of a multivariate normal distribution with diagonal covariance, enabling allocation-free estimation.
+An HMMs-compatible implementation of a multivariate normal distribution with diagonal covariance, enabling allocation-free in-place estimation.
 
 This is not part of the public API and is expected to change.
 
@@ -38,7 +38,7 @@ function Base.rand(rng::AbstractRNG, dist::LightDiagNormal{T1,T2}) where {T1,T2}
     return dist.σ .* randn(rng, T, length(dist)) .+ dist.μ
 end
 
-function DensityInterface.logdensityof(dist::LightDiagNormal, x)
+function DensityInterface.logdensityof(dist::LightDiagNormal, x::Number)
     a = -sum(abs2, (x[i] - dist.μ[i]) / dist.σ[i] for i in eachindex(x, dist.μ, dist.σ))
     b = -sum(dist.logσ)
     logd = (a / 2) + b
@@ -49,16 +49,15 @@ function StatsAPI.fit!(dist::LightDiagNormal{T1,T2}, x, w) where {T1,T2}
     w_tot = sum(w)
     dist.μ .= zero(T1)
     dist.σ .= zero(T2)
-    for i in eachindex(x, w)
-        dist.μ .+= x[i] .* w[i]
+    for (xᵢ, wᵢ) in zip(x, w)
+        dist.μ .+= xᵢ .* wᵢ
+        dist.σ .+= abs2.(xᵢ) .* wᵢ
     end
     dist.μ ./= w_tot
-    for i in eachindex(x, w)
-        dist.σ .+= abs2.(x[i] .- dist.μ) .* w[i]
-    end
     dist.σ ./= w_tot
-    check_positive(dist.σ)
+    dist.σ .-= abs2(dist.μ)
     dist.σ .= sqrt.(dist.σ)
     dist.logσ .= log.(dist.σ)
+    check_positive(dist.σ)
     return nothing
 end
