+++ src/types/periodic_hmm.jl
@@ -1,29 +1,22 @@
 """
 $(TYPEDEF)
 
-Basic implementation of a time-heterogeneous HMM with periodic transition matrices and observation distributions.
+Basic implementation of a time-heterogeneous HMM with periodic transition matrices and observation distributions. 
+
+The period is the first type parameter `L`.
 
 # Fields
 
 $(TYPEDFIELDS)
 """
-struct PeriodicHiddenMarkovModel{
-    period<:Integer,V<:AbstractVector,M<:AbstractMatrix,VD<:AbstractVector
-} <: AbstractHMM
+struct PeriodicHMM{L,V<:AbstractVector,M<:AbstractMatrix,VD<:AbstractVector} <: AbstractHMM
     "initial state probabilities"
     init::V
     "one state transition matrix per time"
-    trans_periodic::NTuple{period,M}
+    trans_periodic::NTuple{L,M}
     "one vector of observation distributions per time (must be amenable to `logdensityof` and `rand`)"
-    dists_periodic::NTuple{period,VD}
+    dists_periodic::NTuple{L,VD}
 end
-
-"""
-    PeriodicHMM
-
-Alias for the type `HiddenMarkovModel`.
-"""
-const PeriodicHMM = PeriodicHiddenMarkovModel
 
 function Base.copy(phmm::PeriodicHMM)
     return PeriodicHMM(
@@ -34,28 +27,107 @@ end
 Base.length(phmm::PeriodicHMM) = length(phmm.init)
 initialization(phmm::PeriodicHMM) = phmm.init
 
-function transition_matrix(phmm::PeriodicHMM{period}, t::Integer) where {period}
-    return phmm.trans_periodic[(t - 1) % period + 1]
+function transition_matrix(phmm::PeriodicHMM{L}, t::Integer) where {L}
+    return phmm.trans_periodic[(t - 1) % L + 1]
 end
 
-function obs_distributions(phmm::PeriodicHMM{period}, t::Integer) where {period}
-    return phmm.dists_periodic[(t - 1) % period + 1]
+function obs_distributions(phmm::PeriodicHMM{L}, t::Integer) where {L}
+    return phmm.dists_periodic[(t - 1) % L + 1]
 end
 
-function StatsAPI.fit!(
-    phmm::PeriodicHMM,
-    init_count::Vector,
-    trans_count::AbstractMatrix,
-    obs_seq::Vector,
-    state_marginals::Matrix,
-)
+## Fitting
+
+struct BaumWelchStoragePeriodicHMM{L,O,R} <: AbstractBaumWelchStorage
+    obs_seqs_concat_periodic::NTuple{L,Vector{O}}
+    state_marginals_concat_periodic::NTuple{L,Matrix{R}}
+end
+
+function initialize_baum_welch(
+    ::PeriodicHMM{L},
+    fb_storages::Vector{<:ForwardBackwardStorage},
+    obs_seqs::Vector{<:Vector},
+) where {L}
+    obs_seqs_concat_periodic = ntuple(
+        l -> reduce(vcat, obs_seqs[k][l:L:end] for k in eachindex(obs_seqs)), L
+    )
+    state_marginals_concat_periodic = ntuple(
+        l -> reduce(hcat, fb_storages[k].γ[:, l:L:end] for k in eachindex(fb_storages)), L
+    )
+    return BaumWelchStoragePeriodicHMM(
+        obs_seqs_concat_periodic, state_marginals_concat_periodic
+    )
+end
+
+function update_baum_welch!(
+    bw_storage::BaumWelchStoragePeriodicHMM{L},
+    fb_storages::Vector{<:ForwardBackwardStorage},
+    obs_seqs::Vector{<:Vector},
+) where {L}
[…]
