+++ src/inference/baum_welch.jl
@@ -1,49 +1,4 @@
-"""
-$(TYPEDEF)
-
-Store Baum-Welch quantities with element type `R`.
-
-This storage is relative to a several sequences.
-
-# Fields
-
-These fields (except `limits`) are passed to the `fit!(hmm, ...)` method along with `obs_seqs_concat`. 
-
-$(TYPEDFIELDS)
-"""
-struct BaumWelchStorage{R,M<:AbstractMatrix{R}}
-    "posterior initialization counts for each state"
-    init_count::Vector{R}
-    "posterior transition counts for each state"
-    trans_count::M
-    "concatenation along time of the state marginal matrices `γ[i,t] = ℙ(X[t]=i | Y[1:T])` for all observation sequences"
-    state_marginals_concat::Matrix{R}
-    "temporal separations between observation sequences: `state_marginals_concat[limits[k]+1:limits[k+1]]` refers to sequence `k`"
-    limits::Vector{Int}
-end
-
-function check_nb_seqs(obs_seqs::Vector{<:Vector}, nb_seqs::Integer)
-    if nb_seqs != length(obs_seqs)
-        throw(ArgumentError("Incoherent sizes provided: `nb_seqs != length(obs_seqs)`"))
-    end
-end
-
-"""
-    initialize_baum_welch(hmm, obs_seqs, nb_seqs)
-"""
-function initialize_baum_welch(
-    hmm::AbstractHMM, obs_seqs::Vector{<:Vector}, nb_seqs::Integer
-)
-    check_nb_seqs(obs_seqs, nb_seqs)
-    N, T_concat = length(hmm), sum(length, obs_seqs)
-    A = transition_matrix(hmm, 1)
-    R = eltype(hmm, obs_seqs[1][1])
-    init_count = Vector{R}(undef, N)
-    trans_count = similar(A, R)
-    state_marginals_concat = Matrix{R}(undef, N, T_concat)
-    limits = vcat(0, cumsum(length.(obs_seqs)))
-    return BaumWelchStorage(init_count, trans_count, state_marginals_concat, limits)
-end
+abstract type AbstractBaumWelchStorage end
 
 """
     initialize_logL_evolution(hmm, obs_seqs, nb_seqs; max_iterations)
@@ -56,24 +11,6 @@ function initialize_logL_evolution(
     logL_evolution = R[]
     sizehint!(logL_evolution, max_iterations)
     return logL_evolution
-end
-
-function update_sufficient_statistics!(
-    bw_storage::BaumWelchStorage{R}, fb_storages::Vector{<:ForwardBackwardStorage}
-) where {R}
-    @unpack init_count, trans_count, state_marginals_concat, limits = bw_storage
-    init_count .= zero(R)
-    trans_count .= zero(R)
-    state_marginals_concat .= zero(R)
-    for k in eachindex(fb_storages)  # TODO: ThreadsX?
-        @unpack γ, ξ = fb_storages[k]
-        init_count .+= @view γ[:, 1]
-        for t in eachindex(ξ)
-            mynonzeros(trans_count) .+= mynonzeros(ξ[t])
-        end
-        state_marginals_concat[:, (limits[k] + 1):limits[k + 1]] .= γ
-    end
-    return nothing
 end
 
 function baum_welch_has_converged(
@@ -91,27 +28,19 @@ function baum_welch_has_converged(
     return false
 end
 
-function StatsAPI.fit!(
-    hmm::AbstractHMM, bw_storage::BaumWelchStorage, obs_seqs_concat::Vector
-)
-    @unpack init_count, trans_count, state_marginals_concat = bw_storage
[…]
