+++ src/inference/forward_backward.jl
@@ -7,13 +7,11 @@ This storage is relative to a single sequence.
 
 # Fields
 
-The only fields useful outside of the algorithm are `γ`, `ξ`, `logL`, the rest does not belong to the public API.
+The only fields useful outside of the algorithm are `γ`, `ξ`, the rest does not belong to the public API.
 
 $(TYPEDFIELDS)
 """
 struct ForwardBackwardStorage{R,M<:AbstractMatrix{R}}
-    "total loglikelihood"
-    logL::RefValue{R}
     "scaled forward messsages `α[i,t]` proportional to `ℙ(Y[1:t], X[t]=i)` (up to a function of `t`)"
     α::Matrix{R}
     "scaled backward messsages `β[i,t]` proportional to `ℙ(Y[t+1:T] | X[t]=i)` (up to a function of `t`)"
@@ -41,11 +39,10 @@ function initialize_forward_backward(
     hmm::AbstractHMM, obs_seq::Vector; transition_marginals=true
 )
     N, T = length(hmm), length(obs_seq)
-    A = transition_matrix(hmm, 1)
+    trans = transition_matrix(hmm, 1)
     R = eltype(hmm, obs_seq[1])
-    M = typeof(similar(A, R))
+    M = typeof(similar(trans, R))
 
-    logL = RefValue{R}(zero(R))
     α = Matrix{R}(undef, N, T)
     β = Matrix{R}(undef, N, T)
     c = Vector{R}(undef, T)
@@ -60,7 +57,7 @@ function initialize_forward_backward(
     logm = Vector{R}(undef, T)
     B = Matrix{R}(undef, N, T)
     scratch = Vector{R}(undef, N)
-    return ForwardBackwardStorage{R,M}(logL, α, β, c, γ, ξ, logB, logm, B, scratch)
+    return ForwardBackwardStorage{R,M}(α, β, c, γ, ξ, logB, logm, B, scratch)
 end
 
 """
@@ -73,26 +70,25 @@ function forward_backward!(
     transition_marginals::Bool=true,
 )
     T = length(obs_seq)
-    @unpack logL, α, β, c, γ, ξ, logB, logm, B, scratch = storage
+    @unpack α, β, c, γ, ξ, logB, logm, B, scratch = storage
 
     # Observation loglikelihoods then likelihoods
     @views for t in 1:T
         obs_logdensities!(logB[:, t], hmm, t, obs_seq[t])
     end
-    check_right_finite(logB)
     maximum!(logm', logB)
     B .= exp.(logB .- logm')
 
     # Forward
     @views begin
-        p = initialization(hmm)
-        α[:, 1] .= p .* B[:, 1]
+        init = initialization(hmm)
+        α[:, 1] .= init .* B[:, 1]
         c[1] = inv(sum(α[:, 1]))
         lmul!(c[1], α[:, 1])
     end
     @views for t in 1:(T - 1)
-        A = transition_matrix(hmm, t)
-        mul!(α[:, t + 1], A', α[:, t])
+        trans = transition_matrix(hmm, t)
+        mul!(α[:, t + 1], trans', α[:, t])
         α[:, t + 1] .*= B[:, t + 1]
         c[t + 1] = inv(sum(α[:, t + 1]))
         lmul!(c[t + 1], α[:, t + 1])
@@ -101,13 +97,13 @@ function forward_backward!(
     # Backward and transition marginals
     β[:, T] .= c[T]
     @views for t in (T - 1):-1:1
-        A = transition_matrix(hmm, t)
+        trans = transition_matrix(hmm, t)
         scratch .= B[:, t + 1] .* β[:, t + 1]  # Bβ
-        mul!(β[:, t], A, scratch)
+        mul!(β[:, t], trans, scratch)
[…]
