### 2023-11-30 9d50e54
Better periodic example
