+++ examples/periodic.jl
@@ -1,89 +1,51 @@
+# # Periodic HMM
+
+using Distributions
+using HiddenMarkovModels
+import HiddenMarkovModels as HMMs
+using Plots
+using SimpleUnPack
+using StatsAPI
+
+# ## Structure
+
 """
-$(TYPEDEF)
+    PeriodicHMM{L}
 
-Basic implementation of a time-heterogeneous HMM with periodic transition matrices and observation distributions. 
-
-The period is the first type parameter `L`.
-
-# Fields
-
-$(TYPEDFIELDS)
+Basic implementation of a periodic HMM with time-dependent transition matrices and observation distributions, repeating every `L` time steps.
 """
 struct PeriodicHMM{L,V<:AbstractVector,M<:AbstractMatrix,VD<:AbstractVector} <: AbstractHMM
-    "initial state probabilities"
     init::V
-    "one state transition matrix per time"
     trans_periodic::NTuple{L,M}
-    "one vector of observation distributions per time (must be amenable to `logdensityof` and `rand`)"
     dists_periodic::NTuple{L,VD}
 end
 
-function Base.copy(phmm::PeriodicHMM)
-    return PeriodicHMM(
-        copy(phmm.init), copy(phmm.trans_periodic), copy(phmm.dists_periodic)
-    )
-end
+period(::PeriodicHMM{L}) where {L} = L
 
 Base.length(phmm::PeriodicHMM) = length(phmm.init)
-initialization(phmm::PeriodicHMM) = phmm.init
+HMMs.initialization(phmm::PeriodicHMM) = phmm.init
 
-function transition_matrix(phmm::PeriodicHMM{L}, t::Integer) where {L}
-    return phmm.trans_periodic[(t - 1) % L + 1]
+function HMMs.transition_matrix(phmm::PeriodicHMM, t::Integer)
+    return phmm.trans_periodic[(t - 1) % period(hmm) + 1]
 end
 
-function obs_distributions(phmm::PeriodicHMM{L}, t::Integer) where {L}
-    return phmm.dists_periodic[(t - 1) % L + 1]
+function HMMs.obs_distributions(phmm::PeriodicHMM, t::Integer)
+    return phmm.dists_periodic[(t - 1) % period(hmm) + 1]
 end
 
 ## Fitting
 
-struct BaumWelchStoragePeriodicHMM{L,O,R} <: AbstractBaumWelchStorage
-    obs_seqs_concat_periodic::NTuple{L,Vector{O}}
-    state_marginals_concat_periodic::NTuple{L,Matrix{R}}
+struct BaumWelchStoragePeriodicHMM <: HMMs.AbstractBaumWelchStorage end
+function HMMs.initialize_baum_welch(::PeriodicHMM, fb_storages, obs_seqs)
+    return BaumWelchStoragePeriodicHMM()
 end
 
-function initialize_baum_welch(
-    ::PeriodicHMM{L},
-    fb_storages::Vector{<:ForwardBackwardStorage},
-    obs_seqs::Vector{<:Vector},
-) where {L}
-    obs_seqs_concat_periodic = ntuple(
-        l -> reduce(vcat, obs_seqs[k][l:L:end] for k in eachindex(obs_seqs)), L
-    )
-    state_marginals_concat_periodic = ntuple(
-        l -> reduce(hcat, fb_storages[k].γ[:, l:L:end] for k in eachindex(fb_storages)), L
-    )
-    return BaumWelchStoragePeriodicHMM(
-        obs_seqs_concat_periodic, state_marginals_concat_periodic
-    )
-end
-
-function update_baum_welch!(
[…]
