+++ examples/periodic.jl
@@ -3,22 +3,19 @@
 using Distributions
 using HiddenMarkovModels
 import HiddenMarkovModels as HMMs
-using Plots
+#md using Plots
 using SimpleUnPack
 using StatsAPI
 
 # ## Structure
 
-"""
-    PeriodicHMM{L}
-
-Basic implementation of a periodic HMM with time-dependent transition matrices and observation distributions, repeating every `L` time steps.
-"""
 struct PeriodicHMM{L,V<:AbstractVector,M<:AbstractMatrix,VD<:AbstractVector} <: AbstractHMM
     init::V
     trans_periodic::NTuple{L,M}
     dists_periodic::NTuple{L,VD}
 end
+
+#-
 
 period(::PeriodicHMM{L}) where {L} = L
 
@@ -36,18 +33,19 @@ end
 ## Fitting
 
 struct BaumWelchStoragePeriodicHMM <: HMMs.AbstractBaumWelchStorage end
+
 function HMMs.initialize_baum_welch(::PeriodicHMM, fb_storages, obs_seqs)
     return BaumWelchStoragePeriodicHMM()
 end
 
+#-
+
 function fit_states!(hmm::PeriodicHMM, fb_storages::Vector{<:HMMs.ForwardBackwardStorage})
     L = period(hmm)
-    # Reset
     hmm.init .= 0
     for l in 1:L
         hmm.trans_periodic[l] .= 0
     end
-    # Accumulate sufficient stats
     for k in eachindex(fb_storages)
         @unpack γ, ξ = fb_storages[k]
         hmm.init .+= view(γ, :, 1)
@@ -56,7 +54,6 @@ function fit_states!(hmm::PeriodicHMM, fb_storages::Vector{<:HMMs.ForwardBackwar
             hmm.trans_periodic[l] .+= ξ[t]
         end
     end
-    # Normalize
     hmm.init ./= sum(hmm.init)
     for l in 1:L
         hmm.trans_periodic[l] ./= sum(hmm.trans_periodic[l]; dims=2)
@@ -64,25 +61,30 @@ function fit_states!(hmm::PeriodicHMM, fb_storages::Vector{<:HMMs.ForwardBackwar
     return nothing
 end
 
+#-
+
 function fit_observations!(
     hmm::PeriodicHMM,
     fb_storages::Vector{<:HMMs.ForwardBackwardStorage},
     obs_seqs::Vector{<:Vector},
 )
+    L = period(hmm)
     for l in 1:L
-        obs_seq_periodic = reduce(vcat, obs_seqs[k][l:L:end] for k in eachindex(obs_seqs))
-        state_marginals_periodic = reduce(
-            hcat, fb_storages[k].γ[:, l:L:end] for k in eachindex(fb_storages)
-        )
         for i in 1:length(hmm)
+            obs_seq_periodic = reduce(
+                vcat, obs_seqs[k][l:L:end] for k in eachindex(obs_seqs)
+            )
+            state_marginals_periodic = reduce(
+                vcat, fb_storages[k].γ[i, l:L:end] for k in eachindex(fb_storages)
+            )
             D = typeof(hmm.dists_periodic[l][i])
-            x = obs_seq_periodic
-            w = view(state_marginals_periodic, i, :)
-            hmm.dists_periodic[l][i] = fit(D, x, w)
+            hmm.dists_periodic[l][i] = fit(D, obs_seq_periodic, state_marginals_periodic)
         end
     end
[…]
