+++ ext/HiddenMarkovModelsHMMBaseExt.jl
@@ -1,6 +1,6 @@
 module HiddenMarkovModelsHMMBaseExt
 
-using HiddenMarkovModels
+using HiddenMarkovModels: HiddenMarkovModels
 using HMMBase: HMMBase
 
 function HiddenMarkovModels.HMM(hmm_base::HMMBase.HMM)
@@ -15,6 +15,18 @@ function HMMBase.HMM(hmm::HiddenMarkovModels.HMM)
     A = deepcopy(hmm.trans)
     B = deepcopy(hmm.dists)
     return HMMBase.HMM(a, A, B)
+end
+
+function HiddenMarkovModels.initialization(hmm_base::HMMBase.HMM)
+    return hmm_base.a
+end
+
+function HiddenMarkovModels.transition_matrix(hmm_base::HMMBase.HMM, ::Integer)
+    return hmm_base.A
+end
+
+function HiddenMarkovModels.obs_distributions(hmm_base::HMMBase.HMM, ::Integer)
+    return hmm_base.B
 end
 
 end
