+++ src/inference/baum_welch.jl
@@ -1,20 +1,56 @@
 abstract type AbstractBaumWelchStorage end
 
 """
-    initialize_logL_evolution(hmm, obs_seqs; max_iterations)
+$(TYPEDEF)
+
+# Fields
+
+$(TYPEDFIELDS)
 """
-function initialize_logL_evolution(
-    hmm::AbstractHMM, obs_seqs::Vector{<:Vector}; max_iterations::Integer
-)
+struct BaumWelchStorage{O,R,M} <: AbstractBaumWelchStorage
+    fb_storages::Vector{ForwardBackwardStorage{R,M}}
+    obs_seqs_concat::Vector{O}
+    state_marginals_concat::Matrix{R}
+    seq_limits::Vector{Int}
+    logL_evolution::Vector{R}
+end
+
+function initialize_baum_welch(hmm::AbstractHMM, obs_seqs::MultiSeq; max_iterations=0)
+    O = typeof(obs_seqs[1][1])
     R = eltype(hmm, obs_seqs[1][1])
+    fb_storages = initialize_forward_backward(hmm, obs_seqs)
+    obs_seqs_concat = Vector{O}(undef, sum(length, obs_seqs))
+    state_marginals_concat = Matrix{R}(undef, length(hmm), sum(length, obs_seqs))
+    seq_limits = vcat(0, cumsum(length.(obs_seqs)))
+    for k in eachindex(obs_seqs, fb_storages)
+        obs_seqs_concat[(seq_limits[k] + 1):seq_limits[k + 1]] .= obs_seqs[k]
+    end
     logL_evolution = R[]
     sizehint!(logL_evolution, max_iterations)
-    return logL_evolution
+    bw_storage = BaumWelchStorage(
+        fb_storages, obs_seqs_concat, state_marginals_concat, seq_limits, logL_evolution
+    )
+    return bw_storage
+end
+
+function update_baum_welch!(
+    bw_storage::BaumWelchStorage, hmm::AbstractHMM, obs_seqs::MultiSeq
+)
+    @unpack fb_storages, state_marginals_concat, logL_evolution, seq_limits = bw_storage
+    forward_backward!(fb_storages, hmm, obs_seqs)
+    logL = zero(eltype(logL_evolution))
+    for k in eachindex(obs_seqs, fb_storages)
+        state_marginals_concat[:, (seq_limits[k] + 1):seq_limits[k + 1]] .= fb_storages[k].γ
+        logL += fb_storages[k].logL[]
+    end
+    push!(logL_evolution, logL)
+    return nothing
 end
 
 function baum_welch_has_converged(
-    logL_evolution::Vector; atol::Real, loglikelihood_increasing::Bool
+    bw_storage::BaumWelchStorage; atol::Real, loglikelihood_increasing::Bool
 )
+    @unpack logL_evolution = bw_storage
     if length(logL_evolution) >= 2
         logL, logL_prev = logL_evolution[end], logL_evolution[end - 1]
         progress = logL - logL_prev
@@ -29,29 +65,22 @@ end
 
 """
     baum_welch!(
-        fb_storages, bw_storage, logL_evolution,
-        hmm, obs_seqs;
+        bw_storage, hmm, MultiSeq(obs_seqs);
         atol, max_iterations, loglikelihood_increasing
     )
 """
 function baum_welch!(
-    hmm::AbstractHMM,
-    fb_storages::Vector{<:ForwardBackwardStorage{R}},
     bw_storage::AbstractBaumWelchStorage,
-    logL_evolution::Vector,
-    obs_seqs::Vector{<:Vector};
+    hmm::AbstractHMM,
+    obs_seqs::MultiSeq;
     atol::Real,
     max_iterations::Integer,
[…]
