### 2023-12-01 325417b
Periodic HMM
