+++ test/runtests.jl
@@ -18,6 +18,7 @@ examples_path = joinpath(dirname(@__DIR__), "examples")
         end
 
         @testset "Code linting" begin
+            using Zygote
             JET.test_package(HiddenMarkovModels; target_defined_modules=true)
         end
 
