+++ .github/workflows/benchmark.yml
@@ -8,7 +8,7 @@ on:
 jobs:
   Benchmark:
     runs-on: ubuntu-latest
-    if: contains(github.event.pull_request.labels.*.name, 'run-benchmark')
+    if: contains(github.event.pull_request.labels.*.name, 'run benchmark')
     steps:
       - uses: actions/checkout@v2
       - uses: julia-actions/setup-julia@latest
