+++ Project.toml
@@ -7,6 +7,7 @@ version = "0.4.0"
 ChainRulesCore = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"
 DensityInterface = "b429d917-457f-4dbc-8f4c-0cc954292b1d"
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
+FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 PrecompileTools = "aea7be01-6a6a-4083-8856-8a6e6704d82a"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
@@ -28,6 +29,7 @@ ChainRulesCore = "1.16"
 DensityInterface = "0.4"
 Distributions = "0.25"
 DocStringExtensions = "0.9"
+FillArrays = "1.9"
 HMMBase = "1"
 LinearAlgebra = "1.6"
 PrecompileTools = "1.1"

[…]
