+++ src/HiddenMarkovModels.jl
@@ -36,6 +36,7 @@ include("utils/lightdiagnormal.jl")
 include("utils/lightcategorical.jl")
 include("utils/multiseq.jl")
 include("utils/control.jl")
+include("utils/test.jl")
 
 include("inference/forward.jl")
 include("inference/viterbi.jl")
@@ -45,7 +46,6 @@ include("inference/logdensity.jl")
 include("inference/chainrules.jl")
 
 include("types/hmm.jl")
-include("types/periodic_hmm.jl")
 
 if !isdefined(Base, :get_extension)
     function __init__()
@@ -58,6 +58,6 @@ if !isdefined(Base, :get_extension)
     end
 end
 
-include("precompile.jl")
+# include("precompile.jl")
 
 end
