+++ examples/basics.jl
@@ -1,32 +1,28 @@
 # # Basics
 
-# ## Setup
-
-# Imports
-
 using Distributions
 using HiddenMarkovModels
 using Random
 using Test  #src
 
-# Random seed
+#-
 
 rng = Random.default_rng()
 Random.seed!(rng, 63)
 
-# ## Model creation
+#-
 
 init = [0.4, 0.6]
 trans = [0.7 0.3; 0.2 0.8]
 dists = [Normal(-1, 0.3), Normal(1, 0.3)]
 hmm = HMM(init, trans, dists)
 
-# ## Simulation
+#-
 
 T = 100
 state_seq, obs_seq = rand(rng, hmm, T)
 
-# ## Inference on a single sequence
+#-
 
 viterbi(hmm, obs_seq)
 
@@ -42,11 +38,7 @@ forward(hmm, obs_seq)
 
 forward_backward(hmm, obs_seq)
 
-# ## Learning from several sequences
-
-obs_seqs = [rand(rng, hmm, k * 100).obs_seq for k in 1:3]
-
-# Baum-Welch needs an initial guess
+#-
 
 init_guess = [0.5, 0.5]
 trans_guess = [0.5 0.5; 0.5 0.5]
@@ -55,6 +47,7 @@ hmm_guess = HMM(init_guess, trans_guess, dists_guess)
 
 #-
 
+obs_seqs = [rand(rng, hmm, rand(T:2T)).obs_seq for k in 1:10]
 hmm_est, logL_evolution = baum_welch(hmm_guess, MultiSeq(obs_seqs))
 
 #-
