+++ examples/periodic.jl
@@ -7,8 +7,7 @@ using SimpleUnPack
 using StatsAPI
 using Test  #src
 
-# ## Structure
-
+#-
 struct PeriodicHMM{L,T<:Number,D} <: AbstractHMM
     init::Vector{T}
     trans_periodic::NTuple{L,Matrix{T}}
@@ -29,42 +28,40 @@ function HMMs.obs_distributions(hmm::PeriodicHMM{L}, t::Integer) where {L}
     return hmm.dists_periodic[(t - 1) % L + 1]
 end
 
-# Construction
+#-
 
 init = [0.4, 0.6]
 trans_periodic = ([0.8 0.2; 0.2 0.8], [0.6 0.4; 0.4 0.6])
 dists_periodic = ([Normal(-1), Normal(+1)], [Normal(-2), Normal(+2)])
 hmm = PeriodicHMM(init, trans_periodic, dists_periodic)
 
-# Simulation
+#-
 
-control_seq = 1:100
+T = 100
+control_seq = 1:T
 state_seq, obs_seq = rand(hmm, control_seq)
 
-# Inference
+#-
 
 viterbi(hmm, obs_seq, control_seq)
 forward(hmm, obs_seq, control_seq)
 logdensityof(hmm, obs_seq, control_seq)
 forward_backward(hmm, obs_seq, control_seq)
 
-## Fitting
+#-
 
-function fit_states!(hmm::PeriodicHMM{L}, bw_storage::HMMs.BaumWelchStorage)
+function fit_states!(hmm::PeriodicHMM{L,T}, bw_storage::HMMs.BaumWelchStorage) where {L,T}
     @unpack fb_storages, obs_seqs_concat, state_marginals_concat, seq_limits = bw_storage
-    K = length(fb_storages)
-    N = length(hmm)
-
-    hmm.init .= zero(eltype(hmm.init))
+    K, N = length(fb_storages), length(hmm)
+    hmm.init .= zero(T)
     for l in 1:L
-        hmm.trans_periodic[l] .= zero(eltype(hmm.trans_periodic[l]))
+        hmm.trans_periodic[l] .= zero(T)
     end
     for k in 1:K
         @unpack γ, ξ = fb_storages[k]
-        T = size(γ, 2)
         hmm.init .+= view(γ, :, 1)
-        for t in 1:(T - 1)
-            hmm.trans_periodic[(t - 1) % L + 1] .+= ξ[t]
+        for l in 1:L
+            @views hmm.trans_periodic[l] .+= sum(ξ[l:L:end])
         end
     end
     hmm.init ./= sum(hmm.init)
@@ -73,11 +70,9 @@ function fit_states!(hmm::PeriodicHMM{L}, bw_storage::HMMs.BaumWelchStorage)
     end
 end
 
-function fit_observations!(hmm::PeriodicHMM{L}, bw_storage::HMMs.BaumWelchStorage)
+function fit_observations!(hmm::PeriodicHMM{L}, bw_storage::HMMs.BaumWelchStorage) where {L}
     @unpack fb_storages, obs_seqs_concat, state_marginals_concat, seq_limits = bw_storage
-    K = length(fb_storages)
-    N = length(hmm)
-
+    K, N = length(fb_storages), length(hmm)
     for l in 1:L
         indices_l = reduce(vcat, (seq_limits[k] + l):L:seq_limits[k + 1] for k in 1:K)
         obs_seq_periodic = view(obs_seqs_concat, indices_l)
@@ -101,16 +96,19 @@ function StatsAPI.fit!(
     return nothing
 end
 
-# Test
+#-
 
[…]
