+++ src/inference/baum_welch.jl
@@ -1,16 +1,26 @@
 """
 $(TYPEDEF)
 
+This storage is relative to multiple sequences.
+
 # Fields
+
+Only the fields with a description are part of the public API.
 
 $(TYPEDFIELDS)
 """
 struct BaumWelchStorage{O,C,R,M}
-    fb_storages::Vector{ForwardBackwardStorage{R,M}}
-    obs_seqs_concat::Vector{O}
-    control_seqs_concat::Vector{C}
-    state_marginals_concat::Matrix{R}
+    "observation sequence `k` corresponds to indices `seq_limits[k]+1:seq_limits[k+1]`"
     seq_limits::Vector{Int}
+    "concatenation of observation sequences (for fitting) `obs_seqs_concat[seq_limits[k]+t] = Yₖ[t]`"
+    obs_seqs_concat::Vector{O}
+    "concatenation of control sequences (for fitting) `control_seqs_concat[seq_limits[k]+t] = Uₖ[t]`"
+    control_seqs_concat::Vector{C}
+    "horizontal concatenation of state marginals `state_marginals_concat[i,seq_limits[k]+t] = ℙ(Xₖ[t]=i | Yₖ[1:T], Uₖ[1:T])"
+    state_marginals_concat::Matrix{R}
+    "one storage for each observation sequence"
+    fb_storages::Vector{ForwardBackwardStorage{R,M}}
+    "loglikelihood of the observation sequence at each iteration of the Baum-Welch algorithm"
     logL_evolution::Vector{R}
 end
 
@@ -23,23 +33,23 @@ function initialize_baum_welch(
     O = typeof(obs_seqs[1][1])
     C = typeof(control_seqs[1][1])
     R = eltype(hmm, obs_seqs[1][1], control_seqs[1][1])
-    fb_storages = initialize_forward_backward(hmm, obs_seqs, control_seqs)
     obs_seqs_concat = Vector{O}(undef, sum(length, obs_seqs))
     control_seqs_concat = Vector{C}(undef, sum(length, control_seqs))
     state_marginals_concat = Matrix{R}(undef, length(hmm), sum(length, obs_seqs))
     seq_limits = vcat(0, cumsum(length.(obs_seqs)))
-    for k in eachindex(obs_seqs, fb_storages)
+    for k in eachindex(sequences(obs_seqs), sequences(control_seqs))
         obs_seqs_concat[(seq_limits[k] + 1):seq_limits[k + 1]] .= obs_seqs[k]
         control_seqs_concat[(seq_limits[k] + 1):seq_limits[k + 1]] .= control_seqs[k]
     end
+    fb_storages = initialize_forward_backward(hmm, obs_seqs, control_seqs)
     logL_evolution = R[]
     sizehint!(logL_evolution, max_iterations)
     bw_storage = BaumWelchStorage(
-        fb_storages,
+        seq_limits,
         obs_seqs_concat,
         control_seqs_concat,
         state_marginals_concat,
-        seq_limits,
+        fb_storages,
         logL_evolution,
     )
     return bw_storage
