+++ src/inference/viterbi.jl
@@ -1,8 +1,6 @@
 """
 $(TYPEDEF)
 
-This storage is relative to a single sequence.
-
 # Fields
 
 Only the fields with a description are part of the public API.
@@ -10,114 +8,97 @@ Only the fields with a description are part of the public API.
 $(TYPEDFIELDS)
 """
 struct ViterbiStorage{R}
-    "joint loglikelihood of the sequence of observations and the most likely state sequence"
+    "most likely state sequence `q[t] = argmaxᵢ ℙ(X[t]=i | Y[1:T])`"
+    q::Vector{Int}
+    "joint loglikelihood of the observation sequence & the most likely state sequence"
     logL::RefValue{R}
     logb::Vector{R}
     ϕ::Vector{R}
     ϕ_prev::Vector{R}
     ψ::Matrix{Int}
-    "most likely state sequence `q[t] = argmaxᵢ ℙ(X[t]=i | Y[1:T])`"
-    q::Vector{Int}
     scratch::Vector{R}
 end
 
+Base.eltype(::ViterbiStorage{R}) where {R} = R
+
 """
-    initialize_viterbi(hmm, obs_seq)
-    initialize_viterbi(hmm, MultiSeq(obs_seqs))
+$(SIGNATURES)
 """
 function initialize_viterbi(
-    hmm::AbstractHMM, obs_seq::AbstractVector, control_seq::AbstractVector
+    hmm::AbstractHMM,
+    obs_seq::AbstractVector;
+    control_seq::AbstractVector,
+    seq_ends::AbstractVector{Int},
 )
-    N, T = length(hmm), length(eachindex(obs_seq, control_seq))
+    N, T = length(hmm), length(obs_seq)
     R = eltype(hmm, obs_seq[1], control_seq[1])
+    q = Vector{Int}(undef, T)
     logL = RefValue{R}()
     logb = Vector{R}(undef, N)
     ϕ = Vector{R}(undef, N)
     ϕ_prev = Vector{R}(undef, N)
     ψ = Matrix{Int}(undef, N, T)
-    q = Vector{Int}(undef, T)
     scratch = Vector{R}(undef, N)
-    return ViterbiStorage(logL, logb, ϕ, ϕ_prev, ψ, q, scratch)
-end
-
-function initialize_viterbi(hmm::AbstractHMM, obs_seqs::MultiSeq, control_seqs::MultiSeq)
-    R = eltype(hmm, obs_seqs[1][1], control_seqs[1][1])
-    storages = Vector{ViterbiStorage{R}}(undef, length(obs_seqs))
-    for k in eachindex(storages, sequences(obs_seqs), sequences(control_seqs))
-        storages[k] = initialize_viterbi(hmm, obs_seqs[k], control_seqs[k])
-    end
-    return storages
+    return ViterbiStorage(q, logL, logb, ϕ, ϕ_prev, ψ, scratch)
 end
 
 """
-    viterbi!(storage, hmm, obs_seq)
-    viterbi!(storage, hmm, MultiSeq(obs_seqs))
+$(SIGNATURES)
 """
 function viterbi!(
-    storage::ViterbiStorage,
+    storage::ViterbiStorage{R},
     hmm::AbstractHMM,
-    obs_seq::AbstractVector,
+    obs_seq::AbstractVector;
     control_seq::AbstractVector,
-)
-    N, T = length(hmm), length(eachindex(obs_seq, control_seq))
+    seq_ends::AbstractVector{Int},
+) where {R}
     @unpack logL, logb, ϕ, ϕ_prev, ψ, q, scratch = storage
-    init = initialization(hmm)
-    obs_logdensities!(logb, hmm, obs_seq[1], control_seq[1])
-    ϕ .= log.(init) .+ logb
-    ϕ_prev .= ϕ
[…]
