+++ examples/periodic.jl
@@ -14,109 +14,114 @@ rng = Random.default_rng()
 Random.seed!(rng, 63)
 
 #-
-struct PeriodicHMM{L,T<:Number,D} <: AbstractHMM
+struct PeriodicHMM{T<:Number,D,L} <: AbstractHMM
     init::Vector{T}
-    trans_periodic::NTuple{L,Matrix{T}}
-    dists_periodic::NTuple{L,Vector{D}}
+    trans_per::NTuple{L,Matrix{T}}
+    dists_per::NTuple{L,Vector{D}}
 end
 
 #-
 
-period(::PeriodicHMM{L}) where {L} = L
+period(::PeriodicHMM{T,D,L}) where {T,D,L} = L
 
 function HMMs.initialization(hmm::PeriodicHMM)
     return hmm.init
 end
 
 function HMMs.transition_matrix(hmm::PeriodicHMM, t::Integer)
-    return hmm.trans_periodic[(t - 1) % period(hmm) + 1]
+    return hmm.trans_per[(t - 1) % period(hmm) + 1]
 end
 
 function HMMs.obs_distributions(hmm::PeriodicHMM, t::Integer)
-    return hmm.dists_periodic[(t - 1) % period(hmm) + 1]
+    return hmm.dists_per[(t - 1) % period(hmm) + 1]
 end
 
 #-
 
 init = [0.4, 0.6]
-trans_periodic = ([0.8 0.2; 0.2 0.8], [0.6 0.4; 0.4 0.6])
-dists_periodic = ([Normal(-1), Normal(+1)], [Normal(-2), Normal(+2)])
-hmm = PeriodicHMM(init, trans_periodic, dists_periodic)
+trans_per = ([0.8 0.2; 0.2 0.8], [0.6 0.4; 0.4 0.6])
+dists_per = ([Normal(-1), Normal(+1)], [Normal(-2), Normal(+2)])
+hmm = PeriodicHMM(init, trans_per, dists_per)
 
 #-
 
 T = 100
 control_seq = 1:T
 state_seq, obs_seq = rand(rng, hmm, control_seq)
+@test sum(abs, obs_seq[1:2:end]) < 0.8 * sum(abs, obs_seq[2:2:end])  #src
 
 #-
 
-viterbi(hmm, obs_seq, control_seq)
-forward(hmm, obs_seq, control_seq)
-logdensityof(hmm, obs_seq, control_seq)
-forward_backward(hmm, obs_seq, control_seq)
+viterbi(hmm, obs_seq; control_seq)
+forward(hmm, obs_seq; control_seq)
+logdensityof(hmm, obs_seq; control_seq)
+forward_backward(hmm, obs_seq; control_seq);
 
 #-
 
-function fit_states!(hmm::PeriodicHMM{L,T}, bw_storage::HMMs.BaumWelchStorage) where {L,T}
-    @unpack fb_storages, obs_seqs_concat, state_marginals_concat, seq_limits = bw_storage
-    K, N = length(fb_storages), length(hmm)
+function StatsAPI.fit!(
+    hmm::PeriodicHMM{T},
+    obs_seq::AbstractVector;
+    control_seq::AbstractVector,
+    seq_ends::AbstractVector{Int},
+    fb_storage::HMMs.ForwardBackwardStorage,
+) where {T}
+    @unpack γ, ξ = fb_storage
+    L, N = period(hmm), length(hmm)
+
     hmm.init .= zero(T)
     for l in 1:L
-        hmm.trans_periodic[l] .= zero(T)
+        hmm.trans_per[l] .= zero(T)
     end
-    for k in 1:K
-        @unpack γ, ξ = fb_storages[k]
[…]
