+++ test/types_allocations.jl
@@ -1,68 +1,102 @@
 using Distributions
 using HiddenMarkovModels
-using HiddenMarkovModels: LightDiagNormal, LightCategorical, rand_prob_vec, rand_trans_mat
 import HiddenMarkovModels as HMMs
-using JET
+using HiddenMarkovModels: LightDiagNormal, LightCategorical, rand_prob_vec, rand_trans_mat
+using JET: @test_opt, @test_call
 using LinearAlgebra
-using Random
+using Random: AbstractRNG, default_rng, seed!
 using SimpleUnPack
 using SparseArrays
 using Test
 
-Random.seed!(63)
+rng = default_rng()
+seed!(rng, 63)
 
-function test_type_stability(hmm::AbstractHMM; T::Integer)
-    state_seq, obs_seq = rand(hmm, T)
+function test_type_stability(
+    rng::AbstractRNG,
+    hmm::AbstractHMM;
+    control_seq::AbstractVector,
+    seq_ends::AbstractVector{Int}=[length(control_seq)],
+)
+    state_seq, obs_seq = rand(rng, hmm, control_seq)
 
-    @test_opt target_modules = (HMMs,) rand(hmm, T)
-    @test_call target_modules = (HMMs,) rand(hmm, T)
+    @test_opt target_modules = (HMMs,) rand(hmm, control_seq)
+    @test_call target_modules = (HMMs,) rand(hmm, control_seq)
 
-    @test_opt target_modules = (HMMs,) logdensityof(hmm, obs_seq)
-    @test_call target_modules = (HMMs,) logdensityof(hmm, obs_seq)
+    @test_opt target_modules = (HMMs,) logdensityof(hmm, obs_seq; control_seq, seq_ends)
+    @test_call target_modules = (HMMs,) logdensityof(hmm, obs_seq; control_seq, seq_ends)
+    @test_opt target_modules = (HMMs,) logdensityof(
+        hmm, obs_seq, state_seq; control_seq, seq_ends
+    )
+    @test_call target_modules = (HMMs,) logdensityof(
+        hmm, obs_seq, state_seq; control_seq, seq_ends
+    )
 
-    @test_opt target_modules = (HMMs,) forward(hmm, obs_seq)
-    @test_call target_modules = (HMMs,) forward(hmm, obs_seq)
+    @test_opt target_modules = (HMMs,) forward(hmm, obs_seq; control_seq, seq_ends)
+    @test_call target_modules = (HMMs,) forward(hmm, obs_seq; control_seq, seq_ends)
 
-    @test_opt target_modules = (HMMs,) viterbi(hmm, obs_seq)
-    @test_call target_modules = (HMMs,) viterbi(hmm, obs_seq)
+    @test_opt target_modules = (HMMs,) viterbi(hmm, obs_seq; control_seq, seq_ends)
+    @test_call target_modules = (HMMs,) viterbi(hmm, obs_seq; control_seq, seq_ends)
 
-    @test_opt target_modules = (HMMs,) forward_backward(hmm, obs_seq)
-    @test_call target_modules = (HMMs,) forward_backward(hmm, obs_seq)
+    @test_opt target_modules = (HMMs,) forward_backward(hmm, obs_seq; control_seq, seq_ends)
+    @test_call target_modules = (HMMs,) forward_backward(
+        hmm, obs_seq; control_seq, seq_ends
+    )
 
-    @test_opt target_modules = (HMMs,) baum_welch(hmm, obs_seq; max_iterations=1)
-    @test_call target_modules = (HMMs,) baum_welch(hmm, obs_seq; max_iterations=1)
[…]
