+++ docs/make.jl
@@ -45,6 +45,7 @@ pages = [
     "Home" => "index.md",
     "Tutorials" => [
         "Basics" => joinpath("examples", "basics.md"),
+        "Distributions" => joinpath("examples", "distributions.md"),
         "Periodic" => joinpath("examples", "periodic.md"),
         "Controlled" => joinpath("examples", "controlled.md"),
     ],
