+++ ext/HiddenMarkovModelsTestExt.jl
@@ -12,14 +12,17 @@ function check_equal_hmms(
     hmm2::AbstractHMM;
     control_seq=[nothing],
     atol::Real=0.1,
+    init::Bool=true,
     test::Bool=true,
 )
     equal_check = true
 
-    init1 = initialization(hmm1)
-    init2 = initialization(hmm2)
-    test && @test isapprox(init1, init2; atol, norm=infnorm)
-    equal_check = equal_check && isapprox(init1, init2; atol, norm=infnorm)
+    if init
+        init1 = initialization(hmm1)
+        init2 = initialization(hmm2)
+        test && @test isapprox(init1, init2; atol, norm=infnorm)
+        equal_check = equal_check && isapprox(init1, init2; atol, norm=infnorm)
+    end
 
     for control in control_seq
         trans1 = transition_matrix(hmm1, control)
@@ -48,9 +51,13 @@ function check_equal_hmms(
 end
 
 function HiddenMarkovModels.test_equal_hmms(
-    hmm1::AbstractHMM, hmm2::AbstractHMM; control_seq=[nothing], atol::Real=0.1
+    hmm1::AbstractHMM,
+    hmm2::AbstractHMM;
+    control_seq=[nothing],
+    atol::Real=0.1,
+    init::Bool=true,
 )
-    check_equal_hmms(hmm1, hmm2; control_seq, atol, test=true)
+    check_equal_hmms(hmm1, hmm2; control_seq, atol, init, test=true)
     return nothing
 end
 
@@ -61,6 +68,7 @@ function HiddenMarkovModels.test_coherent_algorithms(
     control_seq::AbstractVector,
     seq_ends::AbstractVector{Int},
     atol::Real=0.1,
+    init::Bool=false,
 )
     simulations = map(eachindex(seq_ends)) do k
         t1, t2 = seq_limits(seq_ends, k)
@@ -93,7 +101,7 @@ function HiddenMarkovModels.test_coherent_algorithms(
         @test !check_equal_hmms(
             hmm, hmm_guess; control_seq=control_seq[1:2], atol, test=false
         )
-        test_equal_hmms(hmm, hmm_est; control_seq=control_seq[1:2], atol)
+        test_equal_hmms(hmm, hmm_est; control_seq=control_seq[1:2], atol, init)
     end
 
     return nothing
