+++ examples/interfaces.jl
@@ -1,10 +1,11 @@
 # # Interfaces
 
 #=
-Here we discuss how to extend the observation distributions or HMM behavior to fit specific needs.
+Here we discuss how to extend the observation distributions or HMM behavior to satisfy specific needs.
 =#
 
 using DensityInterface
+using Distributions
 using HiddenMarkovModels
 using HiddenMarkovModels: test_coherent_algorithms  #src
 using LinearAlgebra
@@ -28,8 +29,6 @@ They only need to implement three methods:
 
 In addition, the observation can be arbitrary Julia types.
 So let's construct a distribution that generates stuff.
-
-If you want more sophisticated examples, check out the definitions of `HiddenMarkovModels.LightDiagNormal` and `HiddenMarkovModels.LightCategorical`, which are designed to be fast and allocation-free.
 =#
 
 struct Stuff{T}
@@ -55,13 +54,10 @@ end
 
 #=
 It is important to declare to DensityInterface.jl that the custom distribution has a density, thanks to the following trait.
+The logdensity itself can be computed up to an additive constant without issue.
 =#
 
 DensityInterface.DensityKind(::StuffDist) = HasDensity()
-
-#=
-The logdensity itself can be computed up to an additive constant without issue.
-=#
 
 function DensityInterface.logdensityof(dist::StuffDist, obs::Stuff)
     return -abs2(obs.quantity - dist.quantity_mean)
@@ -114,10 +110,9 @@ hmm_guess = HMM(init_guess, trans_guess, dists_guess);
 hmm_est, loglikelihood_evolution = baum_welch(hmm_guess, obs_seq)
 obs_distributions(hmm_est)
 
-# ## Tests  #src
-
-control_seq, seq_ends = fill(nothing, 1000), 100:10:1000  #src
-test_coherent_algorithms(rng, hmm, hmm_guess; control_seq, seq_ends, atol=0.2, init=false)  #src
+#=
+If you want more sophisticated examples, check out [`HiddenMarkovModels.LightDiagNormal`](@ref) and [`HiddenMarkovModels.LightCategorical`](@ref), which are designed to be fast and allocation-free.
+=#
 
 # ## Creating a new HMM type
 
@@ -126,13 +121,14 @@ In some scenarios, the vanilla Baum-Welch algorithm is not exactly what we want.
 For instance, we might have a prior on the parameters of our model, which we want to apply during the fitting step of the iterative procedure.
 
 Then we need to create a new type that satisfies the `AbstractHMM` interface.
-Let's make a simpler version of the built-in `HMM`m with a prior saying that each transition has been observed exactly once.
+Let's make a simpler version of the built-in `HMM`m with a prior saying that each transition has already been observed a certain number of times.
 =#
 
 struct PriorHMM{T,D} <: AbstractHMM
     init::Vector{T}
     trans::Matrix{T}
     dists::Vector{D}
+    trans_prior_count::Int
 end
 
 #=
@@ -144,16 +140,17 @@ HiddenMarkovModels.transition_matrix(hmm::PriorHMM) = hmm.trans
 HiddenMarkovModels.obs_distributions(hmm::PriorHMM) = hmm.dists
 
 #=
-In addition, we want to overload `logdensityof` to specify our prior loglikelihood.
+In addition, we want to overload [`logdensityof(hmm)`](@ref) to specify our prior loglikelihood.
+It corresponds to a Dirichlet distribution over each row of the transition matrix, where each Dirichlet parameter is one plus the number of times the corresponding transition has been observed.
 =#
 
 function DensityInterface.logdensityof(hmm::PriorHMM)
[…]
