+++ examples/basics.jl
@@ -19,12 +19,12 @@ Random.seed!(rng, 63);
 # ## Model construction
 
 #=
-The package provides a versatile `HMM` type with three attributes:
+The package provides a versatile [`HMM`](@ref) type with three attributes:
 - a vector of state initialization probabilities
 - a matrix of state transition probabilities
 - a vector of observation distributions, one for each state
 
-Here we keep it simple by using Distributions.jl, but there are other ways.
+We keep it simple for now by leveraging Distributions.jl.
 =#
 
 init = [0.8, 0.2]
@@ -35,14 +35,13 @@ hmm = HMM(init, trans, dists);
 # ## Simulation
 
 #=
-You can simulate a pair of state and observation sequences by specifying how long you want them to be.
+You can simulate a pair of state and observation sequences with [`rand`](@ref) by specifying how long you want them to be.
 =#
 
 state_seq, obs_seq = rand(rng, hmm, 20);
 
 #=
 Note that the observation sequence is a vector, whose elements have whatever type an observation distribution returns when sampled.
-As we will see, this can be more generic than just numbers or arrays.
 =#
 
 state_seq[1], obs_seq[1]
@@ -54,7 +53,7 @@ In practical applications, the state sequence is not known, which is why we need
 # ## Inference
 
 #=
-The Viterbi algorithm returns the most likely state sequence $\hat{X}_{1:T} = \underset{X_{1:T}}{\mathrm{argmax}}~\mathbb{P}(X_{1:T} \vert Y_{1:T})$, along with the joint loglikelihood $\mathbb{P}(\hat{X}_{1:T}, Y_{1:T})$.
+The Viterbi algorithm ([`viterbi`](@ref)) returns the most likely state sequence $\hat{X}_{1:T} = \underset{X_{1:T}}{\mathrm{argmax}}~\mathbb{P}(X_{1:T} \vert Y_{1:T})$, along with the joint loglikelihood $\mathbb{P}(\hat{X}_{1:T}, Y_{1:T})$.
 =#
 
 best_state_seq, best_joint_loglikelihood = viterbi(hmm, obs_seq);
@@ -66,7 +65,7 @@ As we can see, it is very close to the true state sequence, but not necessarily 
 vcat(state_seq', best_state_seq')
 
 #=
-The forward algorithm returns a matrix of filtered state marginals $\alpha[i, t] = \mathbb{P}(X_t = i | Y_{1:t})$, along with the loglikelihood $\mathbb{P}(Y_{1:T})$ of the observation sequence.
+The forward algorithm ([`forward`](@ref)) returns a matrix of filtered state marginals $\alpha[i, t] = \mathbb{P}(X_t = i | Y_{1:t})$, along with the loglikelihood $\mathbb{P}(Y_{1:T})$ of the observation sequence.
 =#
 
 filtered_state_marginals, obs_seq_loglikelihood1 = forward(hmm, obs_seq);
@@ -79,7 +78,7 @@ This is particularly useful to infer the marginal distribution of the last state
 filtered_state_marginals[:, end]
 
 #=
-Conversely, the forward-backward algorithm returns a matrix of smoothed state marginals $\gamma[i, t] = \mathbb{P}(X_t = i | Y_{1:T})$, along with the loglikelihood $\mathbb{P}(Y_{1:T})$ of the observation sequence.
+Conversely, the forward-backward algorithm ([`forward_backward`](@ref)) returns a matrix of smoothed state marginals $\gamma[i, t] = \mathbb{P}(X_t = i | Y_{1:T})$, along with the loglikelihood $\mathbb{P}(Y_{1:T})$ of the observation sequence.
 =#
 
 smoothed_state_marginals, obs_seq_loglikelihood2 = forward_backward(hmm, obs_seq);
[…]

[…]
