+++ examples/autodiff.jl
@@ -4,27 +4,25 @@
 Here we show how to compute gradients of the observation sequence loglikelihood with respect to various parameters.
 =#
 
+using ComponentArrays
 using DensityInterface
 using Distributions
-using Enzyme
-using ForwardDiff
+using Enzyme: Enzyme
+using ForwardDiff: ForwardDiff
 using HiddenMarkovModels
-using HiddenMarkovModels: test_coherent_algorithms  #src
+import HiddenMarkovModels as HMMs
 using LinearAlgebra
 using Random: Random, AbstractRNG
 using StatsAPI
 using Test  #src
+using Zygote: Zygote
 
 #-
 
 rng = Random.default_rng()
 Random.seed!(rng, 63);
 
-# ## Forward mode
-
-#=
-Since all of our code is type-generic, it is amenable to forward-mode automatic differentiation with ForwardDiff.jl.
-=#
+# ## Data generation
 
 init = [0.8, 0.2]
 trans = [0.7 0.3; 0.3 0.7]
@@ -32,28 +30,92 @@ means = [-1.0, 1.0]
 dists = Normal.(means)
 hmm = HMM(init, trans, dists);
 
-_, obs_seq = rand(rng, hmm, 10);
-
 #-
 
-f1(new_init) = logdensityof(HMM(new_init, trans, dists), obs_seq)
-ForwardDiff.gradient(f1, init)
+obs_seqs = [rand(rng, hmm, 10).obs_seq, rand(rng, hmm, 20).obs_seq];
+obs_seq = reduce(vcat, obs_seqs)
+seq_ends = cumsum(length.(obs_seqs));
 
-#-
+# ## Forward mode
 
-f2(new_trans) = logdensityof(HMM(init, new_trans, dists), obs_seq)
-ForwardDiff.gradient(f2, trans)
+#=
+Since all of our code is type-generic, it is amenable to forward-mode automatic differentiation with ForwardDiff.jl.
 
-#-
+Because this backend only accepts a single vector input, we wrap all parameters with ComponentArrays.jl, and define a new function to differentiate.
+=#
 
-f3(new_means) = logdensityof(HMM(init, trans, Normal.(new_means)), obs_seq)
-ForwardDiff.gradient(f3, means)
+params = ComponentVector(; init, trans, means)
+
+function f(params::ComponentVector)
+    new_hmm = HMM(params.init, params.trans, Normal.(params.means))
+    return logdensityof(new_hmm, obs_seq; seq_ends)
+end
+
+#=
+The gradient computation is now straightforward.
+We will use this value as a source of truth to compare with reverse mode.
+=#
+
+grad_f = ForwardDiff.gradient(f, params)
 
 # ## Reverse mode
 
 #=
 In the presence of many parameters, reverse mode automatic differentiation of the loglikelihood will be much more efficient.
-This requires using Enzyme.jl and the mutating `forward!` function.
-=# 
+The package includes a chain rule for `logdensityof`, which means backends like Zygote.jl can be used out of the box.
+=#
 
-# ## Gradient descent for estimation
+grad_z = Zygote.gradient(f, params)[1]
+
+#-
+
+grad_f ≈ grad_z
+@test grad_f ≈ grad_z  #src
+
+#=
+For increased efficiency, one can use Enzyme.jl and provide temporary storage.
+This requires going one level deeper, to the mutating [`HiddenMarkovModels.forward!`](@ref) function.
+=#
+
+control_seq = fill(nothing, length(obs_seq))
+
+function f!(storage::HMMs.ForwardStorage, params::ComponentVector)
+    new_hmm = HMM(params.init, params.trans, Normal.(params.means))
[…]
