+++ src/utils/lightdiagnormal.jl
@@ -20,7 +20,7 @@ struct LightDiagNormal{
     logσ::V3
 end
 
-function LightDiagNormal(μ, σ)
+function LightDiagNormal(μ::AbstractVector, σ::AbstractVector)
     check_positive(σ)
     return LightDiagNormal(μ, σ, log.(σ))
 end
@@ -51,11 +51,12 @@ function StatsAPI.fit!(dist::LightDiagNormal{T1,T2}, x, w) where {T1,T2}
     dist.σ .= zero(T2)
     for (xᵢ, wᵢ) in zip(x, w)
         dist.μ .+= xᵢ .* wᵢ
-        dist.σ .+= abs2.(xᵢ) .* wᵢ
     end
     dist.μ ./= w_tot
+    for (xᵢ, wᵢ) in zip(x, w)
+        dist.σ .+= abs2.(xᵢ .- dist.μ) .* wᵢ
+    end
     dist.σ ./= w_tot
-    dist.σ .-= abs2.(dist.μ)
     dist.σ .= sqrt.(dist.σ)
     dist.logσ .= log.(dist.σ)
     check_positive(dist.σ)
