+++ test/correctness.jl
@@ -1,14 +1,8 @@
 using Distributions
-using HMMBase: HMMBase
 using HiddenMarkovModels
 import HiddenMarkovModels as HMMs
-using HiddenMarkovModels:
-    LightDiagNormal,
-    LightCategorical,
-    rand_prob_vec,
-    rand_trans_mat,
-    test_equal_hmms,
-    test_coherent_algorithms
+using HiddenMarkovModels: LightDiagNormal, LightCategorical
+using HMMTest
 using LinearAlgebra
 using Random: Random, AbstractRNG, default_rng, seed!
 using SimpleUnPack
@@ -18,200 +12,85 @@ using Test
 rng = default_rng()
 seed!(rng, 63)
 
-function test_identical_hmmbase(
-    rng::AbstractRNG,
-    hmm::AbstractHMM,
-    hmm_guess::Union{Nothing,AbstractHMM}=nothing;
-    T::Integer,
-    atol::Real=1e-5,
-)
-    sim = rand(rng, hmm, T)
-    obs_mat = collect(reduce(hcat, sim.obs_seq)')
-
-    obs_seq = vcat(sim.obs_seq, sim.obs_seq)
-    seq_ends = [length(sim.obs_seq), 2 * length(sim.obs_seq)]
-
-    hmm_base = HMMBase.HMM(hmm)
-    hmm_guess_base = HMMBase.HMM(hmm_guess)
-
-    logL_base = HMMBase.forward(hmm_base, obs_mat)[2]
-    logL = logdensityof(hmm, obs_seq; seq_ends)
-    @test logL ≈ 2logL_base
-
-    α_base, logL_forward_base = HMMBase.forward(hmm_base, obs_mat)
-    α, logL_forward = forward(hmm, obs_seq; seq_ends)
-    @test isapprox(α[:, 1:T], α_base') && isapprox(α[:, (T + 1):(2T)], α_base')
-    @test logL_forward ≈ 2logL_forward_base
-
-    q_base = HMMBase.viterbi(hmm_base, obs_mat)
-    q, logL_viterbi = viterbi(hmm, obs_seq; seq_ends)
-    # Viterbi decoding can vary in case of (infrequent) ties
-    @test mean(q[1:T] .== q_base) > 0.9 && mean(q[(T + 1):(2T)] .== q_base) > 0.9
-
-    γ_base = HMMBase.posteriors(hmm_base, obs_mat)
-    γ, logL_forward_backward = forward_backward(hmm, obs_seq; seq_ends)
-    @test isapprox(γ[:, 1:T], γ_base') && isapprox(γ[:, (T + 1):(2T)], γ_base')
-
-    if !isnothing(hmm_guess)
-        hmm_est_base, hist_base = HMMBase.fit_mle(
-            hmm_guess_base, obs_mat; maxiter=10, tol=-Inf
-        )
-        logL_evolution_base = hist_base.logtots
-        hmm_est, logL_evolution = baum_welch(
-            hmm_guess, obs_seq; seq_ends, max_iterations=10, atol=-Inf
-        )
-        @test isapprox(
-            logL_evolution[(begin + 1):end], 2 * logL_evolution_base[begin:(end - 1)]
-        )
-        test_equal_hmms(hmm_est, HMM(hmm_est_base); atol, init=true)
-    end
-
-    return nothing
-end
-
 ## Settings
 
-T, K = 100, 20
+T, K = 200, 100
 
-## Distributions
+init = [0.4, 0.6]
+init_guess = [0.5, 0.5]
 
-# TODO: add uniform
+trans = [0.8 0.2; 0.2 0.8]
[…]

[…]
