### 2023-12-11 4f10b10
Fix tests 1.6
