+++ Project.toml
@@ -6,24 +6,15 @@ version = "0.4.0"
 [deps]
 ChainRulesCore = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"
 DensityInterface = "b429d917-457f-4dbc-8f4c-0cc954292b1d"
+Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
 FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 PrecompileTools = "aea7be01-6a6a-4083-8856-8a6e6704d82a"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 SimpleUnPack = "ce78b400-467f-4804-87d8-8f486da07d0a"
-StatsAPI = "82ae8749-77ed-4fe6-ae5f-f523153014b0"
-
-[weakdeps]
-Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
-HMMBase = "b2b3ca75-8444-5ffa-85e6-af70e2b64fe7"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
-
-[extensions]
-HiddenMarkovModelsDistributionsExt = "Distributions"
-HiddenMarkovModelsHMMBaseExt = "HMMBase"
-HiddenMarkovModelsSparseArraysExt = "SparseArrays"
+StatsAPI = "82ae8749-77ed-4fe6-ae5f-f523153014b0"
 
 [compat]
 ChainRulesCore = "1.16"
@@ -31,13 +22,18 @@ DensityInterface = "0.4"
 Distributions = "0.25"
 DocStringExtensions = "0.9"
 FillArrays = "1"
-HMMBase = "1"
 LinearAlgebra = "1"
 PrecompileTools = "1.1"
 Random = "1"
-Requires = "1.3"
 SimpleUnPack = "1.1"
 SparseArrays = "1"
 StatsAPI = "1.6"
 julia = "1.6"
 
+[extensions]
+HiddenMarkovModelsDistributionsExt = "Distributions"
+HiddenMarkovModelsSparseArraysExt = "SparseArrays"
+
+[weakdeps]
+Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
+SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
