+++ Project.toml
@@ -28,7 +28,7 @@ Random = "1"
 SimpleUnPack = "1.1"
 SparseArrays = "1"
 StatsAPI = "1.6"
-julia = "1.6"
+julia = "1.9"
 
 [extensions]
 HiddenMarkovModelsDistributionsExt = "Distributions"
