+++ paper/paper.md
@@ -1,114 +1,0 @@
----
-title: 'HiddenMarkovModels.jl: generic, fast and reliable latent variable modeling'
-tags:
-  - Julia
-  - statistics
-  - hmm
-  - inference
-  - estimation
-authors:
-  - name: Guillaume Dalle
-    orcid: 0000-0003-4866-1687
-    affiliation: 1
-affiliations:
- - name: EPFL (IdePHICS and INDY labs), Switzerland
-   index: 1
-date: September 2023
-bibliography: HMM.bib
-
----
-
-# Summary
-
-Hidden Markov Models (or HMMs) are a very popular statistical framework, with numerous applications ranging from speech recognition to bioinformatics.
-They model a sequence of _observations_ $Y_1, \dots, Y_T$ by assuming the existence of a hidden sequence of _states_ $X_1, \dots, X_T$.
-The distribution of a state $X_t$ can only depend on the previous state $X_{t-1}$, and the distribution of an observation $Y_t$ can only depend on the current state $X_t$.
-This is a very versatile and practical set of assumptions: see @rabinerTutorialHiddenMarkov1989 for an introduction and @cappeInferenceHiddenMarkov2005 for a book-length treatment.
-
-Given a sequence of observations and a parametric family of HMMs $\mathbb{P}_\theta$, there are several problems one can face.
-In generic graphical models, these problems are often intractable, but HMMs have a tree-like structure that yields exact solution procedures with polynomial complexity.
-
-| Problem                                                                                  | Algorithm        |
-| ---------------------------------------------------------------------------------------- | ---------------- |
-| Observation sequence likelihood $\mathbb{P}_\theta(Y_{1:T})$                             | Forward          |
-| State marginals $\mathbb{P}_\theta(X_t \vert Y_{1:T})$                                   | Forward-backward |
-| Best state sequence $\mathrm{argmax}_{X_{1:T}}~\mathbb{P}_\theta(X_{1:T} \vert Y_{1:T})$ | Viterbi          |
-| Maximum likelihood parameter $\mathrm{argmax}_\theta~\mathbb{P}_\theta(Y_{1:T})$         | Baum-Welch       |
-
-The package `HiddenMarkovModels.jl` leverages the Julia language [@bezansonJuliaFreshApproach2017] to implement those algorithms in a _generic_, _fast_ and _reliable_ way.
-
-# Statement of need
-
-The initial motivation for HiddenMarkovModels.jl was an application of HMMs to reliability analysis for the French railway company SNCF [@dalleMachineLearningCombinatorial2022].
-In this industrial use case, the observations were marked temporal point processes (sequences of timed events with structured metadata) generated by condition monitoring systems.
-
-Unfortunately, the major implementations of HMMs we surveyed (in Julia and Python) all expect the observations to be generated by a _predefined set of distributions_.
-In Julia, the reference package `HMMBase.jl` [@mouchetHMMBaseJlHidden2023] requires compliance with the `Distributions.jl` [@besanconDistributionsJlDefinition2021] interface, which precludes anything not scalar- or array-valued.
-In Python, `hmmlearn` [@hmmlearnHmmlearnHiddenMarkov2023] only offers Gaussians, mixtures of Gaussians and a few discrete distributions (categorical, multinomial, Poisson).
-Meanwhile, `pomegranate` [@schreiberPomegranateFastFlexible2018a] has a wider set of available distributions, but it doesn't allow for easy extension by the user.
-
-Focusing on Julia specifically, other downsides of `HMMBase.jl` include the lack of support for multiple observation sequences, and the mandatory use of 64-bit floating point numbers.
[…]
