+++ benchmark/benchmarks.jl
@@ -1,14 +1,17 @@
 using HMMBenchmark
 using BenchmarkTools
 
-SUITE = define_suite(;
-    implems=("HMMs.jl",),
-    algos=("logdensity", "viterbi", "forward_backward", "baum_welch"),
-    N_vals=4:4:20,
-    D_vals=fill(10, 5),
-    T_vals=fill(100, 5),
-    K_vals=fill(10, 5),
-    I_vals=fill(10, 5),
-)
+implems = ("HiddenMarkovModels.jl",)
+algos = ("rand", "logdensity", "viterbi", "forward_backward", "baum_welch")
+configurations = []
+for nb_states in (4, 16, 64)
+    push!(
+        configurations,
+        Configuration(;
+            sparse=false, nb_states, obs_dim=1, seq_length=100, nb_seqs=100, bw_iter=1
+        ),
+    )
+end
 
+SUITE = define_suite(; implems, configurations, algos)
 BenchmarkTools.save(joinpath(@__DIR__, "tune.json"), BenchmarkTools.params(SUITE));

[…]
