### 2023-12-12 cec3da5
Simplify docstrings
