### 2023-12-12 5cc6f44
Simplify docstrings (#67)
* Simplify docstrings

* Add benchmark dep
