+++ src/inference/baum_welch.jl
@@ -47,7 +47,6 @@ Return a tuple `(hmm_est, loglikelihood_evolution)` where `hmm_est` is the estim
 
 # Keyword arguments
 
-$(DESCRIBE_CONTROL_STARTS)
 - `atol`: minimum loglikelihood increase at an iteration of the algorithm (otherwise the algorithm is deemed to have converged)
 - `max_iterations`: maximum number of iterations of the algorithm
 - `loglikelihood_increasing`: whether to throw an error if the loglikelihood decreases
