+++ .github/workflows/benchmark.yml
@@ -20,6 +20,6 @@ jobs:
       - name: Run benchmarks
         run: julia -e 'using BenchmarkCI; BenchmarkCI.judge(baseline="origin/main")'
       - name: Print judgement
-        run: julia -e 'using BenchmarkCI; BenchmarkCI.displayjudgement()'
+        run: julia -e 'using BenchmarkCI; BenchmarkCI.postjudge()'
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
