### 2023-12-12 9e3179b
Typo (#68)
* Typo

* Change benchmark comment

* Print and comment
