+++ libs/HMMBenchmark/ext/HMMBenchmarkHMMBaseExt.jl
@@ -5,34 +5,24 @@ using Distributions
 using HMMBase
 using SparseArrays
 
-function rand_params_hmmbase(; N, D)
-    p = ones(N) / N
-    A = ones(N, N) / N
-    μ = randn(N, D)
-    σ = 2 * ones(N, D)
-    return (; p, A, μ, σ)
-end
+function benchmarkables_hmmbase(; configuration, algos)
+    @unpack sparse, nb_states, obs_dim, seq_length, nb_seqs, bw_iter = configuration
 
-function rand_model_hmmbase(; N, D)
-    @unpack p, A, μ, σ = rand_params_hmmbase(; N, D)
-    if D == 1
-        dists = [Normal(μ[n, 1], σ[n, 1]) for n in 1:N]
+    # Model
+    a = ones(nb_states) / nb_states
+    A = ones(nb_states, nb_states) / nb_states
+    if obs_dim == 1
+        B = [Normal(i, 1.0) for _ in 1:nb_states]
     else
-        dists = [DiagNormal(μ[n, :], PDiagMat(σ[n, :] .^ 2)) for n in 1:N]
+        B = [MvNormal(i .* ones(obs_dim), Diagonal(ones(obs_dim))) for _ in 1:nb_states]
     end
-    model = HMMBase.HMM(p, A, dists)
-    return model
-end
+    hmm = HMMBase.HMM(a, A, B)
 
-function benchmarkables_hmmbase(; algos, N, D, T, K, I)
-    rand_model_hmmbase(; N, D)
-    if D == 1
-        obs_mats = [randn(T) for k in 1:K]
-    else
-        obs_mats = [randn(T, D) for k in 1:K]
-    end
-    obs_mats_concat = randn(K * T, D)
+    obs_mat = rand(hmm, seq_length * nb_seqs)
+
+    # Benchmarks
     benchs = Dict()
+
     if "logdensity" in algos
         benchs["logdensity"] = @benchmarkable begin
             for k in 1:($K)
@@ -40,6 +30,7 @@ function benchmarkables_hmmbase(; algos, N, D, T, K, I)
             end
         end setup = (model = rand_model_hmmbase(; N=$N, D=$D))
     end
+
     if "viterbi" in algos
         benchs["viterbi"] = @benchmarkable begin
             for k in 1:($K)
@@ -47,6 +38,7 @@ function benchmarkables_hmmbase(; algos, N, D, T, K, I)
             end
         end setup = (model = rand_model_hmmbase(; N=$N, D=$D))
     end
+
     if "forward_backward" in algos
         benchs["forward_backward"] = @benchmarkable begin
             for k in 1:($K)
@@ -54,11 +46,13 @@ function benchmarkables_hmmbase(; algos, N, D, T, K, I)
             end
         end setup = (model = rand_model_hmmbase(; N=$N, D=$D))
     end
+
     if "baum_welch" in algos
         benchs["baum_welch"] = @benchmarkable begin
-            HMMBase.fit_mle(model, $obs_mats_concat; maxiter=$I, tol=-Inf)
+            HMMBase.fit_mle(model, $obs_mat; maxiter=$I, tol=-Inf)
         end setup = (model = rand_model_hmmbase(; N=$N, D=$D))
     end
+
     return benchs
 end
 
