+++ joss/plots/Project.toml
@@ -1,0 +1,5 @@
+[deps]
+CondaPkg = "992eb4ea-22a4-4c89-a5bb-47a3300528ab"
+HMMBenchmark = "557005d5-2e4a-43f9-8aa7-ba8df2d03179"
+HiddenMarkovModels = "84ca31d5-effc-45e0-bfda-5a68cd981f47"
+PythonCall = "6099a3de-0909-46bc-b1f4-468b9a2dfc0d"

[…]
