+++ libs/HMMBenchmark/src/HMMBenchmark.jl
@@ -3,7 +3,7 @@ module HMMBenchmark
 using BenchmarkTools: @benchmarkable, BenchmarkGroup
 using CSV: CSV
 using DataFrames: DataFrame
-using Distributions: Normal, DiagNormal, PDiagMat
+using Distributions: Normal, DiagNormal
 using HiddenMarkovModels
 using HiddenMarkovModels:
     LightDiagNormal,
@@ -16,9 +16,9 @@ using HiddenMarkovModels:
     initialize_forward_backward,
     forward_backward!,
     baum_welch!
-using LinearAlgebra: SymTridiagonal
+using LinearAlgebra: Diagonal, SymTridiagonal
 using Pkg: Pkg
-using SimpleUnPack: @unpack
+using Random: AbstractRNG
 
 export Configuration, define_suite, parse_results
 
@@ -27,7 +27,9 @@ include("algos.jl")
 include("suite.jl")
 
 benchmarkables_hmmbase(args...; kwargs...) = error("HMMBase not loaded")
+benchmarkables_dynamax(args...; kwargs...) = error("PythonCall not loaded")
 benchmarkables_hmmlearn(args...; kwargs...) = error("PythonCall not loaded")
 benchmarkables_pomegranate(args...; kwargs...) = error("PythonCall not loaded")
+print_python_setup(args...; kwargs...) = error("PythonCall not loaded")
 
 end
