+++ benchmark/benchmarks.jl
@@ -1,8 +1,12 @@
-using HMMBenchmark
 using BenchmarkTools
+using HMMBenchmark
+using Random
+
+rng = Random.default_rng()
+Random.seed!(rng, 63)
 
 implems = ("HiddenMarkovModels.jl",)
-algos = ("rand", "logdensity", "viterbi", "forward_backward", "baum_welch")
+algos = ("rand", "logdensity", "forward", "viterbi", "forward_backward", "baum_welch")
 configurations = []
 for nb_states in (4, 16, 64)
     push!(
@@ -13,5 +17,5 @@ for nb_states in (4, 16, 64)
     )
 end
 
-SUITE = define_suite(; implems, configurations, algos)
+SUITE = define_suite(rng; implems, configurations, algos)
 BenchmarkTools.save(joinpath(@__DIR__, "tune.json"), BenchmarkTools.params(SUITE));
