+++ libs/HMMBenchmark/ext/HMMBenchmarkPythonCallExt/pomegranate.jl
@@ -1,50 +1,54 @@
-pyimport("pomegranate.hmm")
-pyimport("pomegranate.distributions")
+function HMMBenchmark.benchmarkables_pomegranate(rng::AbstractRNG; configuration, algos)
+    (; sparse, nb_states, obs_dim, seq_length, nb_seqs, bw_iter) = configuration
 
-## pomegranate
-
-function rand_params_pomegranate(; N, D)
-    p = torch.ones(N) / N
-    A = torch.ones(N, N) / N
-    μ = torch.randn(N, D)
-    σ = 2 * torch.ones(N, D)
-    return (; p, A, μ, σ)
-end
-
-function rand_model_pomegranate(; N, D, I)
-    @unpack p, A, μ, σ = rand_params_pomegranate(; N, D)
+    # Model
+    starts = torch.ones(nb_states) / nb_states
+    edges = torch.ones(nb_states, nb_states) / nb_states
     distributions = pylist([
         pomegranate.distributions.Normal(;
-            means=μ[n], covs=torch.square(σ[n]), covariance_type="diag"
-        ) for n in 0:(N - 1)
+            means=i * torch.ones(obs_dim),
+            covs=torch.square(torch.ones(obs_dim)),
+            covariance_type="diag",
+        ) for i in 1:nb_states
     ])
-    model = pomegranate.hmm.DenseHMM(;
-        distributions=distributions, edges=A, starts=p, max_iter=I, tol=1e-10, verbose=false
+    hmm = pomegranate.hmm.DenseHMM(;
+        distributions=distributions,
+        edges=edges,
+        starts=starts,
+        max_iter=bw_iter,
+        tol=1e-10,
+        verbose=false,
     )
-    return model
-end
 
-function benchmarkables_pomegranate(; algos, N, D, T, K, I)
-    rand_model_pomegranate(; N, D, I)
-    obs_tens_py = torch.randn(K, T, D)
-    bench = Dict()
+    # Data
+    obs_tens_py = torch.randn(nb_seqs, seq_length, obs_dim)
+
+    # Benchmarks
+    benchs = Dict()
+
     if "logdensity" in algos
-        bench["logdensity"] = @benchmarkable pycall(model_forward, $obs_tens_py) setup = (
-            model_forward = rand_model_pomegranate(; N=$N, D=$D, I=$I).forward
-        )
+        benchs["logdensity"] = @benchmarkable begin
+            pycall($(hmm.forward), $obs_tens_py)
+        end
     end
+
+    if "forward" in algos
+        benchs["forward"] = @benchmarkable begin
+            pycall($(hmm.forward), $obs_tens_py)
+        end
+    end
+
     if "forward_backward" in algos
-        bench["forward_backward"] = @benchmarkable pycall(
-            model_forward_backward, $obs_tens_py
-        ) setup = (
-            model_forward_backward =
-                rand_model_pomegranate(; N=$N, D=$D, I=$I).forward_backward
-        )
+        benchs["forward_backward"] = @benchmarkable begin
+            pycall($(hmm.forward_backward), $obs_tens_py)
+        end
     end
+
     if "baum_welch" in algos
-        bench["baum_welch"] = @benchmarkable pycall(model_fit, $obs_tens_py) setup = (
-            model_fit = rand_model_pomegranate(; N=$N, D=$D, I=$I).fit
-        )
+        benchs["baum_welch"] = @benchmarkable begin
+            pycall($(hmm.fit), $obs_tens_py)
+        end
     end
-    return bench
+
+    return benchs
 end
