+++ libs/HMMBenchmark/ext/HMMBenchmarkPythonCallExt/dynamax.jl
@@ -1,0 +1,4 @@
+function HMMBenchmark.benchmarkables_dynamax(rng::AbstractRNG; configuration, algos)
+    benchs = BenchmarkGroup()
+    return benchs
+end

[…]
