+++ libs/HMMComparison/src/dynamax.jl
@@ -1,0 +1,4 @@
+function benchmarkables_dynamax(rng::AbstractRNG; configuration, algos)
+    benchs = BenchmarkGroup()
+    return benchs
+end
