+++ libs/HMMBenchmark/ext/HMMBenchmarkPythonCallExt/dynamax.jl
@@ -1,4 +1,0 @@
-function HMMBenchmark.benchmarkables_dynamax(rng::AbstractRNG; configuration, algos)
-    benchs = BenchmarkGroup()
-    return benchs
-end
