### 2023-12-21 b452fe5
Benchmarks with Python updated (#69)
* Typo

* Change benchmark comment

* Print and comment

* Work on Python benchmarks

* Python benchmarks updated

* Put in HMMComparison
