### 2023-12-21 03f6ce8
Shorter config
