+++ benchmark/run.jl
@@ -1,4 +1,4 @@
 include("benchmarks.jl")
 
-results = run(SUITE; verbose=true, samples=5)
+results = run(SUITE; verbose=true)
 data = parse_results(minimum(results); path=joinpath(@__DIR__, "results.csv"))
