+++ libs/HMMBenchmark/src/algos.jl
@@ -4,7 +4,15 @@ function benchmarkables_hiddenmarkovmodels(rng::AbstractRNG; configuration, algo
 
     # Model
     init = ones(nb_states) / nb_states
-    trans = ones(nb_states, nb_states) / nb_states
+    if sparse
+        trans = spdiagm(
+            0 => ones(nb_states) / 2,
+            +1 => ones(nb_states - 1) / 2,
+            -(nb_states - 1) => ones(1) / 2,
+        )
+    else
+        trans = ones(nb_states, nb_states) / nb_states
+    end
 
     if custom_dist
         dists = [LightDiagNormal(i .* ones(obs_dim), ones(obs_dim)) for i in 1:nb_states]
@@ -31,13 +39,13 @@ function benchmarkables_hiddenmarkovmodels(rng::AbstractRNG; configuration, algo
     if "rand" in algos
         benchs["rand"] = @benchmarkable begin
             [rand($hmm, $seq_length).obs_seq for _ in 1:($nb_seqs)]
-        end
+        end evals = 1 samples = 100
     end
 
     if "logdensity" in algos
         benchs["logdensity"] = @benchmarkable begin
             logdensityof($hmm, $obs_seq; control_seq=$control_seq, seq_ends=$seq_ends)
-        end
+        end evals = 1 samples = 100
     end
 
     if "forward" in algos
@@ -48,7 +56,7 @@ function benchmarkables_hiddenmarkovmodels(rng::AbstractRNG; configuration, algo
             forward!(
                 f_storage, $hmm, $obs_seq; control_seq=$control_seq, seq_ends=$seq_ends
             )
-        end setup = (
+        end evals = 1 samples = 100 setup = (
             f_storage = initialize_forward(
                 $hmm, $obs_seq; control_seq=$control_seq, seq_ends=$seq_ends
             )
@@ -58,12 +66,12 @@ function benchmarkables_hiddenmarkovmodels(rng::AbstractRNG; configuration, algo
     if "viterbi" in algos
         benchs["viterbi"] = @benchmarkable begin
             viterbi($hmm, $obs_seq; control_seq=$control_seq, seq_ends=$seq_ends)
-        end
+        end evals = 1 samples = 100
         benchs["viterbi!"] = @benchmarkable begin
             viterbi!(
                 v_storage, $hmm, $obs_seq; control_seq=$control_seq, seq_ends=$seq_ends
             )
-        end setup = (
+        end evals = 1 samples = 100 setup = (
             v_storage = initialize_viterbi(
                 $hmm, $obs_seq; control_seq=$control_seq, seq_ends=$seq_ends
             )
@@ -73,12 +81,12 @@ function benchmarkables_hiddenmarkovmodels(rng::AbstractRNG; configuration, algo
     if "forward_backward" in algos
         benchs["forward_backward"] = @benchmarkable begin
             forward_backward($hmm, $obs_seq; control_seq=$control_seq, seq_ends=$seq_ends)
-        end
+        end evals = 1 samples = 100
         benchs["forward_backward!"] = @benchmarkable begin
             forward_backward!(
                 fb_storage, $hmm, $obs_seq; control_seq=$control_seq, seq_ends=$seq_ends
             )
-        end setup = (
+        end evals = 1 samples = 100 setup = (
             fb_storage = initialize_forward_backward(
                 $hmm, $obs_seq; control_seq=$control_seq, seq_ends=$seq_ends
             )
@@ -96,7 +104,7 @@ function benchmarkables_hiddenmarkovmodels(rng::AbstractRNG; configuration, algo
                 atol=-Inf,
                 loglikelihood_increasing=false,
             )
-        end
+        end evals = 1 samples = 100
[…]
