+++ benchmark/benchmarks.jl
@@ -6,21 +6,19 @@ rng = Random.default_rng()
 Random.seed!(rng, 63)
 
 algos = ("rand", "logdensity", "forward", "viterbi", "forward_backward", "baum_welch")
-configurations = []
-for nb_states in (4, 16, 64), obs_dim in (1, 100), custom_dist in (true, false)
-    push!(
-        configurations,
-        Configuration(;
-            sparse=false,
-            custom_dist,
-            nb_states,
-            obs_dim,
-            seq_length=100,
-            nb_seqs=100,
-            bw_iter=1,
-        ),
-    )
-end
+configurations = [
+    # compare state numbers
+    Configuration(; nb_states=4, obs_dim=1),
+    Configuration(; nb_states=8, obs_dim=1),
+    Configuration(; nb_states=16, obs_dim=1),
+    Configuration(; nb_states=32, obs_dim=1),
+    # compare sparse
+    Configuration(; nb_states=64, obs_dim=1, sparse=false),
+    Configuration(; nb_states=64, obs_dim=1, sparse=true),
+    # compare dists
+    Configuration(; nb_states=4, obs_dim=10, custom_dist=true),
+    Configuration(; nb_states=4, obs_dim=10, custom_dist=false),
+]
 
 SUITE = define_suite(rng; configurations, algos)
 BenchmarkTools.save(joinpath(@__DIR__, "tune.json"), BenchmarkTools.params(SUITE));
