### 2023-12-21 76a71ae
More benchmarks (#70)
* More benchmarks

* Shorter config

* No tune.json
