+++ libs/HMMBenchmark/src/algos.jl
@@ -1,13 +1,29 @@
 function benchmarkables_hiddenmarkovmodels(rng::AbstractRNG; configuration, algos)
-    (; sparse, nb_states, obs_dim, seq_length, nb_seqs, bw_iter) = configuration
+    (; sparse, custom_dist, nb_states, obs_dim, seq_length, nb_seqs, bw_iter) =
+        configuration
 
     # Model
     init = ones(nb_states) / nb_states
-    trans = ones(nb_states, nb_states) / nb_states
-    if obs_dim == 1
-        dists = [Normal(i, 1.0) for i in 1:nb_states]
+    if sparse
+        trans = spdiagm(
+            0 => ones(nb_states) / 2,
+            +1 => ones(nb_states - 1) / 2,
+            -(nb_states - 1) => ones(1) / 2,
+        )
     else
+        trans = ones(nb_states, nb_states) / nb_states
+    end
+
+    if custom_dist
         dists = [LightDiagNormal(i .* ones(obs_dim), ones(obs_dim)) for i in 1:nb_states]
+    else
+        if obs_dim == 1
+            dists = [Normal(i, 1.0) for i in 1:nb_states]
+        else
+            dists = [
+                MvNormal(i .* ones(obs_dim), Diagonal(ones(obs_dim))) for i in 1:nb_states
+            ]
+        end
     end
     hmm = HiddenMarkovModels.HMM(init, trans, dists)
 
@@ -23,13 +39,13 @@ function benchmarkables_hiddenmarkovmodels(rng::AbstractRNG; configuration, algo
     if "rand" in algos
         benchs["rand"] = @benchmarkable begin
             [rand($hmm, $seq_length).obs_seq for _ in 1:($nb_seqs)]
-        end
+        end evals = 1 samples = 100
     end
 
     if "logdensity" in algos
         benchs["logdensity"] = @benchmarkable begin
             logdensityof($hmm, $obs_seq; control_seq=$control_seq, seq_ends=$seq_ends)
-        end
+        end evals = 1 samples = 100
     end
 
     if "forward" in algos
@@ -40,7 +56,7 @@ function benchmarkables_hiddenmarkovmodels(rng::AbstractRNG; configuration, algo
             forward!(
                 f_storage, $hmm, $obs_seq; control_seq=$control_seq, seq_ends=$seq_ends
             )
-        end setup = (
+        end evals = 1 samples = 100 setup = (
             f_storage = initialize_forward(
                 $hmm, $obs_seq; control_seq=$control_seq, seq_ends=$seq_ends
             )
@@ -50,12 +66,12 @@ function benchmarkables_hiddenmarkovmodels(rng::AbstractRNG; configuration, algo
     if "viterbi" in algos
         benchs["viterbi"] = @benchmarkable begin
             viterbi($hmm, $obs_seq; control_seq=$control_seq, seq_ends=$seq_ends)
-        end
+        end evals = 1 samples = 100
         benchs["viterbi!"] = @benchmarkable begin
             viterbi!(
                 v_storage, $hmm, $obs_seq; control_seq=$control_seq, seq_ends=$seq_ends
             )
-        end setup = (
+        end evals = 1 samples = 100 setup = (
             v_storage = initialize_viterbi(
                 $hmm, $obs_seq; control_seq=$control_seq, seq_ends=$seq_ends
             )
@@ -65,12 +81,12 @@ function benchmarkables_hiddenmarkovmodels(rng::AbstractRNG; configuration, algo
     if "forward_backward" in algos
         benchs["forward_backward"] = @benchmarkable begin
             forward_backward($hmm, $obs_seq; control_seq=$control_seq, seq_ends=$seq_ends)
-        end
+        end evals = 1 samples = 100
         benchs["forward_backward!"] = @benchmarkable begin
[…]
