### 2023-12-21 846636c
Replace views with batch
