+++ Project.toml
@@ -10,11 +10,20 @@ Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
 FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+Polyester = "f517fe37-dbe3-4b94-8317-1923a5111588"
 PrecompileTools = "aea7be01-6a6a-4083-8856-8a6e6704d82a"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 SimpleUnPack = "ce78b400-467f-4804-87d8-8f486da07d0a"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 StatsAPI = "82ae8749-77ed-4fe6-ae5f-f523153014b0"
+
+[weakdeps]
+Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
+SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
+
+[extensions]
+HiddenMarkovModelsDistributionsExt = "Distributions"
+HiddenMarkovModelsSparseArraysExt = "SparseArrays"
 
 [compat]
 ChainRulesCore = "1.16"
@@ -23,17 +32,10 @@ Distributions = "0.25"
 DocStringExtensions = "0.9"
 FillArrays = "1"
 LinearAlgebra = "1"
+Polyester = "0.7"
 PrecompileTools = "1.1"
 Random = "1"
 SimpleUnPack = "1.1"
 SparseArrays = "1"
 StatsAPI = "1.6"
 julia = "1.9"
-
-[extensions]
-HiddenMarkovModelsDistributionsExt = "Distributions"
-HiddenMarkovModelsSparseArraysExt = "SparseArrays"
-
-[weakdeps]
-Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
-SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
