### 2023-12-21 c310db7
Fix manifest
