+++ src/inference/viterbi.jl
@@ -50,7 +50,7 @@ function viterbi!(
 ) where {R}
     @unpack q, logL, logB, ϕ, ψ = storage
 
-    @batch for k in eachindex(seq_ends)
+    @threads for k in eachindex(seq_ends)
         t1, t2 = seq_limits(seq_ends, k)
 
         obs_logdensities!(view(logB, :, t1), hmm, obs_seq[t1], control_seq[t1])
