### 2023-12-21 c38c049
Make it easier to parallelize by dispatching on seq_ends
