### 2023-12-21 ee33d20
Threads (#72)
* Replace views with batch

* Fix manifest

* Add views

* batch becomes threads

* Back to batch, higher workload in benchmark

* Make it easier to parallelize by dispatching on seq_ends

* Resolve benchmarks
