### 2023-12-26 #5874 @gdalle
I will, so sorry for the delay. The package will be infinitely more generic and interesting as a result 😊
