+++ examples/controlled.jl
@@ -10,7 +10,6 @@ import HiddenMarkovModels as HMMs
 using HMMTest  #src
 using LinearAlgebra
 using Random
-using SimpleUnPack
 using StatsAPI
 using Test  #src
 
@@ -96,7 +95,7 @@ function StatsAPI.fit!(
     control_seq::AbstractVector,
     seq_ends::AbstractVector{Int},
 ) where {T}
-    @unpack γ, ξ = fb_storage
+    (; γ, ξ) = fb_storage
     N = length(hmm)
 
     hmm.init .= 0
