### 2024-01-12 2b592f7
Skip allocation test on sparse matrices
