+++ src/inference/chainrules.jl
@@ -2,8 +2,8 @@ _dcat(M1, M2) = cat(M1, M2; dims=3)
 
 function _params_and_loglikelihoods(
     hmm::AbstractHMM,
-    obs_seq::Vector;
-    control_seq::AbstractVector=Fill(nothing, length(obs_seq)),
+    obs_seq::Vector,
+    control_seq::AbstractVector=Fill(nothing, length(obs_seq));
     seq_ends::AbstractVector{Int}=Fill(length(obs_seq), 1),
 )
     init = initialization(hmm)
@@ -20,16 +20,16 @@ function ChainRulesCore.rrule(
     rc::RuleConfig,
     ::typeof(logdensityof),
     hmm::AbstractHMM,
-    obs_seq::AbstractVector;
-    control_seq::AbstractVector=Fill(nothing, length(obs_seq)),
+    obs_seq::AbstractVector,
+    control_seq::AbstractVector=Fill(nothing, length(obs_seq));
     seq_ends::AbstractVector{Int}=Fill(length(obs_seq), 1),
 )
     _, pullback = rrule_via_ad(
-        rc, _params_and_loglikelihoods, hmm, obs_seq; control_seq, seq_ends
+        rc, _params_and_loglikelihoods, hmm, obs_seq, control_seq; seq_ends
     )
-    fb_storage = initialize_forward_backward(hmm, obs_seq; control_seq, seq_ends)
-    forward_backward!(fb_storage, hmm, obs_seq; control_seq, seq_ends)
-    (; logL, α, β, γ, Bβ) = fb_storage
+    fb_storage = initialize_forward_backward(hmm, obs_seq, control_seq; seq_ends)
+    forward_backward!(fb_storage, hmm, obs_seq, control_seq; seq_ends)
+    (; logL, α, γ, Bβ) = fb_storage
     N, T = length(hmm), length(obs_seq)
     R = eltype(α)
 
@@ -45,11 +45,11 @@ function ChainRulesCore.rrule(
     ΔlogB = γ
 
     function logdensityof_hmm_pullback(ΔlogL)
-        _, Δhmm, Δobs_seq = pullback((
+        _, Δhmm, Δobs_seq, Δcontrol_seq = pullback((
             ΔlogL .* Δinit, ΔlogL .* Δtrans_by_time, ΔlogL .* ΔlogB
         ))
         Δlogdensityof = NoTangent()
-        return Δlogdensityof, Δhmm, Δobs_seq
+        return Δlogdensityof, Δhmm, Δobs_seq, Δcontrol_seq
     end
 
     return sum(logL), logdensityof_hmm_pullback
