### 2024-01-12 36b3313
Remove deps and fix Enzyme (#73)
* Remove deps and fix Enzyme

* Skip allocation test on sparse matrices

* Move control_seq to positional argument
