+++ libs/HMMComparison/test/comparison.jl
@@ -6,15 +6,18 @@ using Random
 rng = Random.default_rng()
 Random.seed!(rng, 63)
 
-implems = ("HiddenMarkovModels.jl", "HMMBase.jl", "dynamax", "hmmlearn", "pomegranate")
-algos = ("logdensity", "baum_welch")
-configurations = [
-    Configuration(;
-        sparse=false, nb_states=4, obs_dim=1, seq_length=100, nb_seqs=100, bw_iter=10
-    ),
+implems = [
+    HiddenMarkovModelsImplem(),  #
+    HMMBaseImplem(),  #
+    hmmlearnImplem(),  #
+    pomegranateImplem(),  #
+]
+algos = ["logdensity", "forward", "viterbi", "forward_backward", "baum_welch"]
+instances = [
+    Instance(; sparse=false, nb_states=4, obs_dim=2, seq_length=100, nb_seqs=5, bw_iter=10)
 ]
 
-SUITE = define_full_suite(rng; implems, configurations, algos)
-# BenchmarkTools.save(joinpath(@__DIR__, "tune.json"), BenchmarkTools.params(SUITE));
+SUITE = define_suite(rng, implems; instances, algos)
+
 results = BenchmarkTools.run(SUITE; verbose=true)
-data = parse_results(minimum(results); path=joinpath(@__DIR__, "results.csv"))
+data = parse_results(minimum(results))
