+++ libs/HMMComparison/src/hmmlearn.jl
@@ -1,9 +1,15 @@
-function benchmarkables_hmmlearn(rng::AbstractRNG; configuration, algos)
-    np = pyimport("numpy")
-    (; sparse, nb_states, obs_dim, seq_length, nb_seqs, bw_iter) = configuration
+struct hmmlearnImplem <: Implementation end
 
-    # Model
-    hmm = pyimport("hmmlearn.hmm").GaussianHMM(;
+function HMMBenchmark.build_model(
+    rng::AbstractRNG, implem::hmmlearnImplem; instance::Instance
+)
+    np = pyimport("numpy")
+    hmmlearn_hmm = pyimport("hmmlearn.hmm")
+
+    (; bw_iter, nb_states) = instance
+    (; init, trans, means, stds) = build_params(rng; instance)
+
+    hmm = hmmlearn_hmm.GaussianHMM(;
         n_components=nb_states,
         covariance_type="diag",
         n_iter=bw_iter,
@@ -11,49 +17,60 @@ function benchmarkables_hmmlearn(rng::AbstractRNG; configuration, algos)
         implementation="scaling",
         init_params="",
     )
-    hmm.startprob_ = np.ones(nb_states) / nb_states
-    hmm.transmat_ = np.ones((nb_states, nb_states)) / nb_states
-    hmm.means_ =
-        np.ones((nb_states, obs_dim)) *
-        np.arange(1, nb_states + 1)[0:(nb_states - 1), np.newaxis]
-    hmm.covars_ = np.ones((nb_states, obs_dim))
 
-    # Data
-    obs_mats_list_py = pylist([hmm.sample(seq_length)[0] for _ in 1:nb_seqs])
-    obs_mat_concat_py = np.concatenate(obs_mats_list_py)
+    hmm.startprob_ = np.array(init)
+    hmm.transmat_ = np.array(trans)
+    hmm.means_ = np.array(means')
+    hmm.means_ = np.array(means')
+    hmm.covars_ = np.array((stds .^ 2)')
+    return hmm
+end
+
+function HMMBenchmark.build_benchmarkables(
+    rng::AbstractRNG, implem::hmmlearnImplem; instance::Instance, algos::Vector{String}
+)
+    np = pyimport("numpy")
+    (; obs_dim, seq_length, nb_seqs) = instance
+
+    hmm = build_model(rng, implem; instance)
+    data = randn(rng, nb_seqs, seq_length, obs_dim)
+
+    obs_mat_concat = reduce(vcat, data[k, :, :] for k in 1:nb_seqs)
+    obs_mat_concat_py = np.array(obs_mat_concat)
     obs_mat_len_py = np.full(nb_seqs, seq_length)
 
-    # Benchmarks
     benchs = Dict()
 
     if "logdensity" in algos
         benchs["logdensity"] = @benchmarkable begin
             pycall($(hmm.score), $obs_mat_concat_py, $obs_mat_len_py)
-        end
+        end evals = 1 samples = 100
     end
 
     if "forward" in algos
         benchs["forward"] = @benchmarkable begin
             pycall($(hmm.score), $obs_mat_concat_py, $obs_mat_len_py)
-        end
+        end evals = 1 samples = 100
     end
 
     if "viterbi" in algos
         benchs["viterbi"] = @benchmarkable begin
             pycall($(hmm.decode), $obs_mat_concat_py, $obs_mat_len_py)
-        end
+        end evals = 1 samples = 100
     end
 
     if "forward_backward" in algos
         benchs["forward_backward"] = @benchmarkable begin
             pycall($(hmm.predict_proba), $obs_mat_concat_py, $obs_mat_len_py)
-        end
+        end evals = 1 samples = 100
     end
 
[…]
