### 2024-02-02 c7134e1
Clean up benchmarks (#77)
* Rephrase benchmarks with dispatch

* Dynamax benchmarks

* Update Manifest
