+++ .github/workflows/benchmark.yml
@@ -18,7 +18,7 @@ jobs:
       - name: Install dependencies
         run: julia -e 'using Pkg; pkg"add PkgBenchmark BenchmarkCI@0.1"'
       - name: Run benchmarks
-        run: julia -t -e 'using BenchmarkCI, PkgBenchmark; BenchmarkCI.judge(baseline="origin/main", target=BenchmarkConfig(env=Dict("JULIA_NUM_THREADS" => 4)))'
+        run: julia -e 'using BenchmarkCI, PkgBenchmark; BenchmarkCI.judge(baseline="origin/main", target=BenchmarkConfig(env=Dict("JULIA_NUM_THREADS" => 4)))'
       - name: Print judgement
         run: julia -e 'using BenchmarkCI; BenchmarkCI.displayjudgement()'
       - name: Comment judgement
