### 2024-02-02 20511c4
Add multithreading (#78)
* Add multithreading

* Multithreaded judgement

* Fix Enzyme and multithread benchmarks

* Fix -t

* auto threads

* 1 BLAS thread

* Fix parsing
