### 2024-02-02 5ce0c5f
Parallelize fitting
