+++ Project.toml
@@ -26,9 +26,6 @@ DensityInterface = "0.4"
 Distributions = "0.25"
 DocStringExtensions = "0.9"
 FillArrays = "1"
-LinearAlgebra = "1"
 PrecompileTools = "1.1"
-Random = "1"
-SparseArrays = "1"
 StatsAPI = "1.6"
 julia = "1.9"
