### 2024-02-02 8bf278a
Parallelize fitting (#80)
* Parallelize fitting

* Stdlib compatds
