+++ libs/HMMTest/src/coherence.jl
@@ -2,8 +2,8 @@ infnorm(x) = maximum(abs, x)
 
 function check_equal_hmms(
     hmm1::AbstractHMM,
-    hmm2::AbstractHMM;
-    control_seq=[nothing],
+    hmm2::AbstractHMM,
+    control_seq=[nothing];
     atol::Real=0.1,
     init::Bool=true,
     test::Bool=true,
@@ -45,21 +45,53 @@ end
 
 function test_equal_hmms(
     hmm1::AbstractHMM,
-    hmm2::AbstractHMM;
-    control_seq=[nothing],
+    hmm2::AbstractHMM,
+    control_seq=[nothing];
     atol::Real=0.1,
     init::Bool=true,
 )
-    check_equal_hmms(hmm1, hmm2; control_seq, atol, init, test=true)
+    check_equal_hmms(hmm1, hmm2, control_seq; atol, init, test=true)
     return nothing
+end
+
+function test_coherent_algorithms_aux(
+    hmm::AbstractHMM,
+    obs_seq::AbstractVecOrMat,
+    state_seq::AbstractVector{<:Integer},
+    control_seq::AbstractVecOrMat;
+    seq_ends::AbstractVector{Int},
+    hmm_guess::Union{Nothing,AbstractHMM},
+    atol::Real,
+    init::Bool,
+)
+    logL = logdensityof(hmm, obs_seq, control_seq; seq_ends)
+    logL_joint = joint_logdensityof(hmm, obs_seq, state_seq, control_seq; seq_ends)
+
+    q, logL_viterbi = viterbi(hmm, obs_seq, control_seq; seq_ends)
+    @test logL_viterbi > logL_joint
+    @test logL_viterbi ≈ joint_logdensityof(hmm, obs_seq, q, control_seq; seq_ends)
+
+    α, logL_forward = forward(hmm, obs_seq, control_seq; seq_ends)
+    @test logL_forward ≈ logL
+
+    γ, logL_forward_backward = forward_backward(hmm, obs_seq, control_seq; seq_ends)
+    @test logL_forward_backward ≈ logL
+    @test all(α[:, seq_ends[k]] ≈ γ[:, seq_ends[k]] for k in eachindex(seq_ends))
+
+    if !isnothing(hmm_guess)
+        hmm_est, logL_evolution = baum_welch(hmm_guess, obs_seq, control_seq; seq_ends)
+        @test all(>=(0), diff(logL_evolution))
+        @test !check_equal_hmms(hmm, hmm_guess, control_seq[1:2]; atol, test=false)
+        test_equal_hmms(hmm, hmm_est, control_seq[1:2]; atol, init)
+    end
 end
 
 function test_coherent_algorithms(
     rng::AbstractRNG,
     hmm::AbstractHMM,
-    hmm_guess::Union{Nothing,AbstractHMM}=nothing;
-    control_seq::AbstractVector,
+    control_seq::AbstractVecOrMat;
     seq_ends::AbstractVector{Int},
+    hmm_guess::Union{Nothing,AbstractHMM}=nothing,
     atol::Real=0.1,
     init::Bool=true,
 )
@@ -75,27 +107,18 @@ function test_coherent_algorithms(
         state_seq = reduce(vcat, state_seqs)
         obs_seq = reduce(vcat, obs_seqs)
 
-        logL = logdensityof(hmm, obs_seq, control_seq; seq_ends)
-        logL_joint = joint_logdensityof(hmm, obs_seq, state_seq, control_seq; seq_ends)
-
-        q, logL_viterbi = viterbi(hmm, obs_seq, control_seq; seq_ends)
-        @test logL_viterbi > logL_joint
[…]
