### 2024-02-03 e48a742
Fairer Python benchmarks
