+++ libs/HMMComparison/test/comparison.jl
@@ -1,6 +1,10 @@
 using BenchmarkTools
 using HMMComparison
 using HMMBenchmark
+using LinearAlgebra
+using StableRNGs
+
+BLAS.set_num_threads(1)
 
 rng = StableRNG(63)
 
@@ -13,9 +17,7 @@ implems = [
 ]
 algos = ["logdensity", "forward", "viterbi", "forward_backward", "baum_welch"]
 instances = [
-    Instance(;
-        sparse=false, nb_states=5, obs_dim=10, seq_length=100, nb_seqs=50, bw_iter=10
-    ),
+    Instance(; sparse=false, nb_states=5, obs_dim=10, seq_length=1000, nb_seqs=10, bw_iter=10)
 ]
 
 SUITE = define_suite(rng, implems; instances, algos)
