### 2024-02-03 1ed97cf
Fairer Python benchmarks
* Fairer Python benchmarks
