+++ libs/HMMBenchmark/src/HMMBenchmark.jl
@@ -23,11 +23,11 @@ using Random: AbstractRNG
 using SparseArrays: spdiagm
 using Statistics: mean, median, std
 
-export AbstractImplementation, Instance
-export define_suite, parse_results, print_julia_setup
+export Implementation, Instance, Params, HiddenMarkovModelsImplem
+export build_params, build_data, build_model, build_benchmarkables
+export define_suite, parse_results, read_results, print_julia_setup
 
 abstract type Implementation end
-Base.string(implem::Implementation) = string(typeof(implem))[begin:(end - length("Implem"))]
 
 include("instance.jl")
 include("params.jl")
