+++ libs/HMMComparison/test/plots.jl
@@ -1,11 +1,9 @@
 using DataFrames
 using Plots
 
-include("comparison.jl")
+data = read_results(joinpath(@__DIR__, "results.csv"))
 
-data
-
-algo = "forward"
+algo = "baum_welch"
 metric = :time_minimum
 data_algo = data[data[!, :algo] .== algo, :]
 bar(data_algo[!, :implem], data_algo[!, metric]; title=algo, label=string(metric))
