+++ libs/HMMComparison/test/performance.jl
@@ -1,6 +1,5 @@
 using BenchmarkTools
 using HMMComparison
-using HMMBenchmark
 using LinearAlgebra
 using StableRNGs
 
@@ -21,15 +20,18 @@ algos = ["forward", "viterbi", "forward_backward", "baum_welch"]
 instances = Instance[]
 
 for nb_states in 2:3:24
-    push!(instances, Instance(;
-        custom_dist=true,
-        sparse=false,
-        nb_states=nb_states,
-        obs_dim=5,
-        seq_length=100,
-        nb_seqs=10,
-        bw_iter=10,
-    ))
+    push!(
+        instances,
+        Instance(;
+            custom_dist=true,
+            sparse=false,
+            nb_states=nb_states,
+            obs_dim=5,
+            seq_length=100,
+            nb_seqs=10,
+            bw_iter=10,
+        ),
+    )
 end
 
 SUITE = define_suite(rng, implems; instances, algos)
