### 2024-02-03 0b7f2ac
Correct benchmarks (#84)
* Correct benchmarks

* Plots
