+++ libs/HMMComparison/src/pomegranate.jl
@@ -1,15 +1,16 @@
 struct pomegranateImplem <: Implementation end
+Base.string(::pomegranateImplem) = "pomegranate"
 
 function HMMBenchmark.build_model(
-    rng::AbstractRNG, implem::pomegranateImplem; instance::Instance
+    implem::pomegranateImplem, instance::Instance, params::Params
 )
     torch = pyimport("torch")
-    torch.set_default_dtype(torch.float64)
     pomegranate_distributions = pyimport("pomegranate.distributions")
     pomegranate_hmm = pyimport("pomegranate.hmm")
+    torch.set_default_dtype(torch.float64)
 
     (; nb_states, bw_iter) = instance
-    (; init, trans, means, stds) = build_params(rng; instance)
+    (; init, trans, means, stds) = params
 
     starts = torch.tensor(Py(init).to_numpy())
     ends = torch.ones(nb_states) * 1e-10
@@ -37,43 +38,39 @@ function HMMBenchmark.build_model(
 end
 
 function HMMBenchmark.build_benchmarkables(
-    rng::AbstractRNG, implem::pomegranateImplem; instance::Instance, algos::Vector{String}
+    implem::pomegranateImplem,
+    instance::Instance,
+    params::Params,
+    data::AbstractArray{<:Real,3},
+    algos::Vector{String},
 )
-    np = pyimport("numpy")
     torch = pyimport("torch")
     torch.set_default_dtype(torch.float64)
+
     (; obs_dim, seq_length, nb_seqs) = instance
+    hmm = build_model(implem, instance, params)
 
-    hmm = build_model(rng, implem; instance)
-    data = randn(rng, nb_seqs, seq_length, obs_dim)
-
-    obs_tens_py = torch.tensor(Py(data).to_numpy())
+    obs_tens_torch_py = torch.tensor(Py(data).to_numpy())
 
     benchs = Dict()
 
-    if "logdensity" in algos
-        benchs["logdensity"] = @benchmarkable begin
-            $(hmm.forward)($obs_tens_py)
-        end evals = 1 samples = 100
-    end
-
     if "forward" in algos
         benchs["forward"] = @benchmarkable begin
-            $(hmm.forward)($obs_tens_py)
+            $(hmm.forward)($obs_tens_torch_py)
         end evals = 1 samples = 100
     end
 
     if "forward_backward" in algos
         benchs["forward_backward"] = @benchmarkable begin
-            $(hmm.forward_backward)($obs_tens_py)
+            $(hmm.forward_backward)($obs_tens_torch_py)
         end evals = 1 samples = 100
     end
 
     if "baum_welch" in algos
         benchs["baum_welch"] = @benchmarkable begin
-            hmm_guess.fit($obs_tens_py)
+            hmm_guess.fit($obs_tens_torch_py)
         end evals = 1 samples = 100 setup = (
-            hmm_guess = build_model($rng, $implem; instance=$instance)
+            hmm_guess = build_model($implem, $instance, $params)
         )
     end
 
