+++ src/utils/lightcategorical.jl
@@ -16,8 +16,8 @@ struct LightCategorical{T1,T2,V1<:AbstractVector{T1},V2<:AbstractVector{T2}}
     logp::V2
 end
 
-function LightCategorical(p::AbstractVector)
-    check_prob_vec(p)
+function LightCategorical(p::AbstractVector{T}) where {T}
+    @argcheck valid_prob_vec(p)
     return LightCategorical(p, log.(p))
 end
 
@@ -55,6 +55,6 @@ function StatsAPI.fit!(dist::LightCategorical{T1}, x, w) where {T1}
     end
     dist.p ./= w_tot
     dist.logp .= log.(dist.p)
-    check_prob_vec(dist.p)
+    @argcheck valid_prob_vec(dist.p)
     return nothing
 end
