+++ src/utils/linalg.jl
@@ -1,7 +1,5 @@
 sum_to_one!(x) = ldiv!(sum(x), x)
 
-mysimilar_mutable(x::AbstractArray, ::Type{R}) where {R} = similar(x, R)
-
 mynonzeros(x::AbstractArray) = x
 mynonzeros(x::AbstractSparseArray) = nonzeros(x)
 
@@ -17,8 +15,8 @@ end
 function mul_rows_cols!(
     B::SparseMatrixCSC, l::AbstractVector, A::SparseMatrixCSC, r::AbstractVector
 )
-    @assert size(B) == size(A) == (length(l), length(r))
-    @assert nnz(B) == nnz(A)
+    @argcheck size(B) == size(A) == (length(l), length(r))
+    @argcheck nnz(B) == nnz(A)
     for j in axes(B, 2)
         for k in nzrange(B, j)
             i = B.rowval[k]
