+++ examples/autodiff.jl
@@ -1,7 +1,7 @@
 # # Autodiff
 
 #=
-Here we show how to compute gradients of the observation sequence loglikelihood with respect to various parameters.
+Here we show how to compute gradients of the observation sequence loglikelihood with respect to various inputs.
 =#
 
 using ComponentArrays
@@ -18,90 +18,189 @@ using StatsAPI
 using Test  #src
 using Zygote: Zygote
 
-Enzyme.API.runtimeActivity!(true)
-
 #-
 
 rng = StableRNG(63);
 
-# ## Data generation
+# ## Diffusion HMM
+
+#=
+To play around with automatic differentiation, we define a simple controlled HMM.
+=#
+
+struct DiffusionHMM{V1<:AbstractVector,M2<:AbstractMatrix,V3<:AbstractVector} <: AbstractHMM
+    init::V1
+    trans::M2
+    means::V3
+end
+
+#=
+Both its transition matrix and its vector of observation means result from a convex combination between the corresponding field and a base value (aka diffusion).
+The coefficient $\lambda$ of this convex combination is given as a control. 
+=#
+
+HMMs.initialization(hmm::DiffusionHMM) = hmm.init
+
+function HMMs.transition_matrix(hmm::DiffusionHMM, λ::Number)
+    N = length(hmm)
+    return (1 - λ) * hmm.trans + λ * ones(N, N) / N
+end
+
+function HMMs.obs_distributions(hmm::DiffusionHMM, λ::Number)
+    return [Normal((1 - λ) * hmm.means[i] + λ * 0) for i in 1:length(hmm)]
+end
+
+#=
+We now construct an instance of this object and draw samples from it.
+=#
 
 init = [0.6, 0.4]
 trans = [0.7 0.3; 0.3 0.7]
 means = [-1.0, 1.0]
-dists = Normal.(means)
-hmm = HMM(init, trans, dists);
+hmm = DiffusionHMM(init, trans, means);
 
-#-
+#=
+It is essential that the controls are taken between $0$ and $1$.
+=#
 
-obs_seqs = [rand(rng, hmm, 10).obs_seq, rand(rng, hmm, 20).obs_seq];
+control_seqs = [rand(rng, 3), rand(rng, 5)];
+obs_seqs = [rand(rng, hmm, control_seqs[k]).obs_seq for k in 1:2];
+
+control_seq = reduce(vcat, control_seqs)
 obs_seq = reduce(vcat, obs_seqs)
 seq_ends = cumsum(length.(obs_seqs));
+
+# ## What to differentiate?
+
+#=
+The key function we are interested in is the loglikelihood of the observation sequence.
+We can differentiate it with respect to
+- the model itself (`hmm`), or more precisely its parameters
+- the observation sequence (`obs_seq`)
+- the control sequence (`control_seq`).
+- but not with respect to the sequence limits (`seq_ends`), which are discrete.
+=#
+
+logdensityof(hmm, obs_seq, control_seq; seq_ends)
+
+#=
+To ensure compatibility with backends that only accept a single input, we wrap all parameters inside a `ComponentVector` from [ComponentArrays.jl](https://github.com/jonniedie/ComponentArrays.jl), and define a new function to differentiate.
+=#
+
+parameters = ComponentVector(; init, trans, means)
+
+function f(parameters::ComponentVector, obs_seq, control_seq; seq_ends)
+    new_hmm = DiffusionHMM(parameters.init, parameters.trans, parameters.means)
+    return logdensityof(new_hmm, obs_seq, control_seq; seq_ends)
+end;
+
+f(parameters, obs_seq, control_seq; seq_ends)
 
[…]
