+++ src/utils/lightdiagnormal.jl
@@ -54,14 +54,14 @@ function StatsAPI.fit!(dist::LightDiagNormal{T1,T2}, x, w) where {T1,T2}
     w_tot = sum(w)
     dist.μ .= zero(T1)
     dist.σ .= zero(T2)
-    for (xᵢ, wᵢ) in zip(x, w)
-        dist.μ .+= xᵢ .* wᵢ
-        dist.σ .+= abs2.(xᵢ) .* wᵢ
+    @inbounds @simd for i in eachindex(x, w)
+        dist.μ .+= x[i] .* w[i]
     end
     dist.μ ./= w_tot
-    dist.σ ./= w_tot
-    dist.σ .-= min.(abs2.(dist.μ), dist.σ)
-    dist.σ .= sqrt.(dist.σ)
+    @inbounds @simd for i in eachindex(x, w)
+        dist.σ .+= abs2.(x[i] .- dist.μ) .* w[i]
+    end
+    dist.σ .= sqrt.(dist.σ ./ w_tot)
     dist.logσ .= log.(dist.σ)
     @argcheck all(isfinite, dist.μ)
     @argcheck all(isfinite, dist.σ)
