+++ docs/src/alternatives.md
@@ -4,26 +4,25 @@
 
 We compare features among the following Julia packages:
 
-* HiddenMarkovModels.jl (abbreviated to HMMs.jl)
+* HiddenMarkovModels.jl
 * [HMMBase.jl](https://github.com/maxmouchet/HMMBase.jl)
 * [HMMGradients.jl](https://github.com/idiap/HMMGradients.jl)
 
 We discard [MarkovModels.jl](https://github.com/FAST-ASR/MarkovModels.jl) because its focus is GPU computation.
 There are also more generic packages for probabilistic programming, which are able to perform MCMC or variational inference (eg. [Turing.jl](https://github.com/TuringLang/Turing.jl)) but we leave those aside.
 
-|                           | HMMs.jl             | HMMBase.jl       | HMMGradients.jl |
-| ------------------------- | ------------------- | ---------------- | --------------- |
-| Algorithms[^1]            | V, FB, BW           | V, FB, BW        | FB              |
-| Number types              | anything            | `Float64`        | `AbstractFloat` |
-| Observation types         | anything            | number or vector | anything        |
-| Observation distributions | DensityInterface.jl | Distributions.jl | manual          |
-| Multiple sequences        | yes                 | no               | yes             |
-| Priors / structures       | possible            | no               | possible        |
-| Temporal dependency       | yes                 | no               | no              |
-| Control dependency        | yes                 | no               | no              |
-| Automatic differentiation | yes                 | no               | yes             |
-| Linear algebra speedup    | yes                 | yes              | no              |
-| Numerical stability       | scaling+            | scaling+         | log             |
+|                           | HiddenMarkovModels.jl | HMMBase.jl       | HMMGradients.jl |
+| ------------------------- | --------------------- | ---------------- | --------------- |
+| Algorithms[^1]            | V, FB, BW             | V, FB, BW        | FB              |
+| Number types              | anything              | `Float64`        | `AbstractFloat` |
+| Observation types         | anything              | number or vector | anything        |
+| Observation distributions | DensityInterface.jl   | Distributions.jl | manual          |
+| Multiple sequences        | yes                   | no               | yes             |
+| Priors / structures       | possible              | no               | possible        |
+| Control dependency        | yes                   | no               | no              |
+| Automatic differentiation | yes                   | no               | yes             |
+| Linear algebra speedup    | yes                   | yes              | no              |
+| Numerical stability       | scaling+              | scaling+         | log             |
 
 
 !!! info "Very small probabilities"
@@ -43,17 +42,16 @@ We compare features among the following Python packages:
 
 |                           | hmmlearn             | pomegranate           | dynamax              |
 | ------------------------- | -------------------- | --------------------- | -------------------- |
-| Algorithms[^1]            | V, FB, BW, VI        | V, FB, BW             | FB, V, BW, GD        |
-| Number types              | NumPy format         | PyTorch format        | JAX format           |
+| Algorithms[^1]            | V, FB, BW, VI        | FB, BW                | FB, V, BW, GD        |
+| Number types              | NumPy formats        | PyTorch formats       | JAX formats          |
 | Observation types         | number or vector     | number or vector      | number or vector     |
 | Observation distributions | discrete or Gaussian | pomegranate catalogue | discrete or Gaussian |
 | Multiple sequences        | yes                  | yes                   | yes                  |
-| Priors / structures       | yes                  | no                    | ?                    |
-| Temporal dependency       | no                   | no                    | no                   |
+| Priors / structures       | yes                  | no                    | yes                  |
 | Control dependency        | no                   | no                    | no                   |
 | Automatic differentiation | no                   | yes                   | yes                  |
[…]
