### 2024-02-22 ba666fb
Better docs and numerical stability (#87)
* Better docs and numerical stability

* Fix leaky tests

* Better autodiff tuto

* Typos

* Fix table
