+++ docs/make.jl
@@ -43,15 +43,15 @@ end
 
 pages = [
     "Home" => "index.md",
-    "API reference" => "api.md",
     "Tutorials" => [
         "Basics" => joinpath("examples", "basics.md"),
         "Types" => joinpath("examples", "types.md"),
         "Interfaces" => joinpath("examples", "interfaces.md"),
-        "Autodiff" => joinpath("examples", "autodiff.md"),
         "Time dependency" => joinpath("examples", "temporal.md"),
         "Control dependency" => joinpath("examples", "controlled.md"),
+        "Autodiff" => joinpath("examples", "autodiff.md"),
     ],
+    "API reference" => "api.md",
     "Advanced" => [
         "Alternatives" => "alternatives.md",
         "Debugging" => "debugging.md",
