### 2024-02-22 d08346b
Collapse docstrings
