+++ docs/src/api.md
@@ -1,3 +1,7 @@
+```@meta
+CollapsedDocStrings = true
+```
+
 # API reference
 
 ```@docs
