### 2024-02-22 dbdcee6
fit_in_sequence
