### 2024-02-22 5ad5d81
Fix benchmarks (#90)
* Fix benchmarks

* Don't ignore png

* Remove JOSS

* Commit benchmark manifest

* Remove Pkg business

* Fix paths
