+++ libs/HMMBenchmark/src/setup.jl
@@ -1,7 +1,7 @@
-function print_julia_setup(; path)
+function print_julia_setup(path)
     open(path, "w") do file
         redirect_stdout(file) do
-            versioninfo()
+            InteractiveUtils.versioninfo()
             println("\n# Multithreading\n")
             println("Julia threads = $(Threads.nthreads())")
             println("OpenBLAS threads = $(BLAS.get_num_threads())")
