+++ libs/HMMComparison/src/setup.jl
@@ -1,8 +1,8 @@
 
-function print_python_setup(; path)
+function print_python_setup(path)
     open(path, "w") do file
         redirect_stdout(file) do
-            println("Pytorch threads = $(torch.get_num_threads())")
+            println("Pytorch threads = $(pyimport("torch").get_num_threads())")
             println("\n# Python packages\n")
         end
         redirect_stderr(file) do
