### 2024-02-22 #5874 @gdalle
I have released a new version with everything I want in it, cleaned up the benchmark set and revamped the paper.
I only have doubts on the performance of the latest Python competitor, so I have asked its devs for feedback on my benchmark code in order to be as fair as possible. If they don't answer within a week I'll just add my current plots to the paper and submit for good.

See the issue:
- https://github.com/probml/dynamax/issues/359
