### 2024-02-24 64262cd
Speed up viterbi
