### 2024-02-24 5c0367c
Remove sparse special case
