+++ docs/src/api.md
@@ -36,13 +36,6 @@ transition_matrix
 obs_distributions
 ```
 
-Optional log versions:
-
-```@docs
-log_initialization
-log_transition_matrix
-```
-
 ## Utils
 
 ```@docs
@@ -101,20 +94,22 @@ HiddenMarkovModels.forward_backward!
 HiddenMarkovModels.baum_welch!
 ```
 
-## Misc
+## Miscellaneous
 
 ```@docs
 HiddenMarkovModels.valid_hmm
 HiddenMarkovModels.rand_prob_vec
 HiddenMarkovModels.rand_trans_mat
-HiddenMarkovModels.LightDiagNormal
-HiddenMarkovModels.LightCategorical
 HiddenMarkovModels.fit_in_sequence!
 ```
 
 ## Internals
 
 ```@docs
+HiddenMarkovModels.LightDiagNormal
+HiddenMarkovModels.LightCategorical
+HiddenMarkovModels.log_initialization
+HiddenMarkovModels.log_transition_matrix
 HiddenMarkovModels.mul_rows_cols!
 HiddenMarkovModels.argmaxplus_mul!
 ```

[…]
