### 2024-02-24 2746f09
Give PR write permissions to benchmark workflow
