### 2024-02-24 083bbf9
Special case for sparse matrices
