### 2024-02-24 2f975b6
Speed up viterbi (#93)
