+++ src/inference/viterbi.jl
@@ -13,7 +13,7 @@ struct ViterbiStorage{R}
     "one joint loglikelihood per pair of observation sequence and most likely state sequence"
     logL::Vector{R}
     logB::Matrix{R}
-    logϕ::Matrix{R}
+    ϕ::Matrix{R}
     ψ::Matrix{Int}
 end
 
@@ -33,9 +33,9 @@ function initialize_viterbi(
     q = Vector{Int}(undef, T)
     logL = Vector{R}(undef, K)
     logB = Matrix{R}(undef, N, T)
-    logϕ = Matrix{R}(undef, N, T)
+    ϕ = Matrix{R}(undef, N, T)
     ψ = Matrix{Int}(undef, N, T)
-    return ViterbiStorage(q, logL, logB, logϕ, ψ)
+    return ViterbiStorage(q, logL, logB, ϕ, ψ)
 end
 
 """
@@ -49,31 +49,26 @@ function viterbi!(
     t1::Integer,
     t2::Integer;
 ) where {R}
-    (; q, logB, logϕ, ψ) = storage
+    (; q, logB, ϕ, ψ) = storage
 
-    obs_logdensities!(view(logB, :, t1), hmm, obs_seq[t1], control_seq[t1])
-    init = initialization(hmm)
-    logϕ[:, t1] .= log.(init) .+ view(logB, :, t1)
+    logBₜ₁ = view(logB, :, t1)
+    obs_logdensities!(logBₜ₁, hmm, obs_seq[t1], control_seq[t1])
+    loginit = log_initialization(hmm)
+    ϕ[:, t1] .= loginit .+ logBₜ₁
 
     for t in (t1 + 1):t2
-        obs_logdensities!(view(logB, :, t), hmm, obs_seq[t], control_seq[t])
-        trans = transition_matrix(hmm, control_seq[t - 1])
-        for j in 1:length(hmm)
-            i_max = 1
-            score_max = logϕ[i_max, t - 1] + log(trans[i_max, j])
-            for i in 2:length(hmm)
-                score = logϕ[i, t - 1] + log(trans[i, j])
-                if score > score_max
-                    score_max, i_max = score, i
-                end
-            end
-            ψ[j, t] = i_max
-            logϕ[j, t] = score_max + logB[j, t]
-        end
+        logBₜ = view(logB, :, t)
+        obs_logdensities!(logBₜ, hmm, obs_seq[t], control_seq[t])
+        logtrans = log_transition_matrix(hmm, control_seq[t - 1])
+        ϕₜ, ϕₜ₋₁ = view(ϕ, :, t), view(ϕ, :, t - 1)
+        ψₜ = view(ψ, :, t)
+        argmaxplus_transmul!(ϕₜ, ψₜ, logtrans, ϕₜ₋₁)
+        ϕₜ .+= logBₜ
     end
 
-    q[t2] = argmax(view(logϕ, :, t2))
-    logL = logϕ[q[t2], t2]
+    ϕₜ₂ = view(ϕ, :, t2)
+    q[t2] = argmax(ϕₜ₂)
+    logL = ϕ[q[t2], t2]
     for t in (t2 - 1):-1:t1
         q[t] = ψ[q[t + 1], t + 1]
     end
