+++ libs/HMMComparison/README.md
@@ -1,0 +1,10 @@
+# HMMComparison
+
+To re-run the experiments from the JOSS paper, clone the repository, start a Julia REPL and run
+
+```julia
+include("libs/HMMComparison/experiments/measurements.jl")
+include("libs/HMMComparison/experiments/plots.jl")
+```
+
+The results and plots will appear in `libs/HMMComparison/experiments/results/`.
