### 2024-02-29 9e0b7ab
Final benchmarks (#94)
* Final benchmarks

* Fix formatting
