### 2024-02-29 #5874 @gdalle
I also would like to change the title to

> HiddenMarkovModels.jl: generic, fast and reliable state space modeling

if that's not a problem
