### 2024-02-29 #5874 @jbytecode
👋👋👋 Dear @dmbates, @DanielRivasMD 👋👋👋

**[HiddenMarkovModels.jl: generic, fast and reliable state space modeling](https://github.com/openjournals/joss-reviews/issues/5874)**

You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/5874).

Thank you in advance!
