### 2024-02-29 #5874 @DanielRivasMD
Dear editor,

I will be glad to assist in the review process.

Best regards,

Daniel Rivas

> On 29 Feb 2024, at 10:23, Mehmet Hakan Satman ***@***.***> wrote:
>
>
> 👋👋👋 Dear @dmbates <https://github.com/dmbates>, @DanielRivasMD <https://github.com/DanielRivasMD> 👋👋👋
>
> Would you be willing to assist in reviewing this submission for JOSS (Journal of Open Source Software)?
>
> JOSS publishes articles about open source research software. The submission I'd like you to review is titled:
>
> HiddenMarkovModels.jl: generic, fast and reliable state space modeling <https://github.com/openjournals/joss-reviews/issues/5874>
> You can find more information at the top of this Github issue (#5874 <https://github.com/openjournals/joss-reviews/issues/5874>).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged. If you have any questions please let me know.
>
> This is the pre-review issue. After setting at least 2 reviewers we will start the review process in a separate thread. In that thread, there will be 23 check items for each single reviewer.
>
> Thank you in advance!
>
