### 2024-03-01 #6436 @editorialbot
[REVIEW]: HiddenMarkovModels.jl: generic, fast and reliable latent variable modeling
**Submitting author:** @gdalle (<a href="http://orcid.org/0000-0003-4866-1687">Guillaume Dalle</a>)
**Repository:** https://github.com/gdalle/HiddenMarkovModels.jl
**Branch with paper.md** (empty if default branch): joss2
**Version:** v0.5.1
