### 2024-03-01 #6436 @gdalle
I'll take a look at the citation. The bot seems to expect that the paper is located on a branch where the rest of the code lives too. I'm gonna merge `main` into `joss` to fix it
