### 2024-03-01 #6436 @jbytecode
@gdalle - As expressed at the top of the page, the default branch of your paper is `joss`.
