### 2024-03-01 #6436 @gdalle
Yeah but there's no code on it, just the markdown, so that's why the bot can't find the license or get an accurate LOC count
