### 2024-03-01 #6436 @gdalle
Also the version of the package is now 0.5.0
