+++ paper/paper.md
@@ -1,0 +1,113 @@
+---
+title: 'HiddenMarkovModels.jl: generic, fast and reliable state space modeling'
+tags:
+  - Julia
+  - statistics
+  - hmm
+  - inference
+  - estimation
+authors:
+  - name: Guillaume Dalle
+    orcid: 0000-0003-4866-1687
+    affiliation: "1, 2, 3"
+affiliations:
+  - name: Information, Learning and Physics laboratory, Ecole Polytechnique Fédérale de Lausanne (EPFL), Station 11, CH-1015 Lausanne
+    index: 1
+  - name: Information and Network Dynamics laboratory, Ecole Polytechnique Fédérale de Lausanne (EPFL), Station 14, CH-1015 Lausanne
+    index: 2
+  - name: Statistical Physics of Computation laboratory, Ecole Polytechnique Fédérale de Lausanne (EPFL), CH-1015 Lausanne
+    index: 3
+date: 22 February 2024
+bibliography: HMM.bib
+
+---
+
+# Summary
+
+Hidden Markov Models (or HMMs) are a very popular statistical framework, with numerous applications ranging from speech recognition to bioinformatics.
+They characterize a sequence of _observations_ $Y_1, \dots, Y_T$ by assuming the existence of a hidden sequence of _states_ $X_1, \dots, X_T$.
+The distribution of a state $X_t$ can only depend on the previous state $X_{t-1}$, and the distribution of an observation $Y_t$ can only depend on the current state $X_t$.
+In addition, both of these dynamics may be influenced by exogenous control variables $U_1, \dots, U_T$.
+This is a very versatile and practical set of assumptions: see @rabinerTutorialHiddenMarkov1989 for an introduction, @cappeInferenceHiddenMarkov2005 for a book-length treatment and @bengioInputOutputHMM1994 for a seminal discussion of HMMs with controls.
+
+Given a sequence of observations and a parametric family of HMMs $\mathbb{P}_\theta$, there are several problems one can face.
+In generic graphical models, these problems are often intractable, but HMMs have a tree-like structure that yields exact solution procedures with polynomial complexity.
+The package `HiddenMarkovModels.jl` leverages the Julia language [@bezansonJuliaFreshApproach2017] to implement those algorithms in a _generic_, _fast_ and _reliable_ way.
+
+| Inference problem                                                                                              | Algorithm        |
+| -------------------------------------------------------------------------------------------------------------- | ---------------- |
+| Best state sequence              $\mathrm{argmax}_{X_{1:T}}~\mathbb{P}_\theta(X_{1:T} \vert Y_{1:T}, U_{1:T})$ | Viterbi          |
+| Observation sequence likelihood  $\mathbb{P}_\theta(Y_{1:T} \vert U_{1:T})$                                    | Forward          |
+| State marginals                  $\mathbb{P}_\theta(X_t \vert Y_{1:T}, U_{1:T})$                               | Forward-backward |
+| Maximum likelihood parameter     $\mathrm{argmax}_\theta~\mathbb{P}_\theta(Y_{1:T} \vert U_{1:T})$             | Baum-Welch       |
+
+# Statement of need
+
+The initial motivation for HiddenMarkovModels.jl was an application of HMMs to reliability analysis for the French railway company SNCF [@dalleMachineLearningCombinatorial2022].
+In this industrial use case, the observations were marked temporal point processes (sequences of timed events with structured metadata) generated by condition monitoring systems, possibly influenced by the daily activity of the train unit.
+
+Unfortunately, nearly all implementations of HMMs we surveyed (in Julia and Python) expect the observations to be generated by a _predefined set of distributions_, with _no temporal heterogeneity_.
[…]
