### 2024-03-01 #6436 @gdalle
@jbytecode all good
