### 2024-03-03 #6436 @dmbates
The only questions I have about the paper and the package is the form of citation in the paper for other Julia packages.  Those like Makie.jl and ChainRules.jl that have recommended citation forms for the package are fine.  Those that don't (HMMBase.jl, PythonCall.jl, and LogarithmicNumbers.jl) are cited by the author, year, package name, and package title.  Those experienced with Julia packages may know to check at, e.g. https://juliahub.com/ui/packages, to determine the repository but a reader unfamiliar with Julia packages may not know this.  How are registered Julia packages without citation info cited in other JOSS papers?
