### 2024-03-03 #6436 @gdalle
I don't know, I pulled the information directly from the GitHub repository with Zotero whenever there was no CITATION.bib or Zenodo DOI available. If there's a better practice I'm curious too
