### 2024-03-11 #6436 @gdalle
@jbytecode can we maybe get some guidance on this issue, so that @dmbates's review may be completed?
