### 2024-03-12 #6436 @dmbates
@gdalle There's this wonderful resource https://modernjuliaworkflows.github.io/ by G. Dalle et al. that mentions `PkgCite.jl` as a way to generate citations for the packages that you currently have loaded.  It would suggest as a bibtex entry
```bibtex

@misc{PythonCall.jl,
 author        = {Rowley, Christopher},
 year          = {2022},
 url           = {https://github.com/JuliaPy/PythonCall.jl},
 title         = {PythonCall.jl: Python and Julia in harmony}
}

```
