+++ paper/HMM.bib
@@ -5,10 +5,11 @@
   year = {2021},
   month = jun,
   doi = {10.5281/zenodo.4454565},
+  url = {https://doi.org/10.5281/zenodo.4454565},
   urldate = {2023-09-12},
   howpublished = {Zenodo},
   keywords = {hmm},
-  file = {/Users/guillaume/Zotero/storage/PEFYSLF7/4906644.html}
+  file = {/home/guillaume/Zotero/storage/PEFYSLF7/4906644.html}
 }
 
 @inproceedings{bengioInputOutputHMM1994,
@@ -17,12 +18,17 @@
   author = {Bengio, Yoshua and Frasconi, Paolo},
   year = {1994},
   volume = {7},
-  publisher = {{MIT Press}},
+  publisher = {MIT Press},
+  url = {https://proceedings.neurips.cc/paper/1994/hash/8065d07da4a77621450aa84fee5656d9-Abstract.html},
   urldate = {2023-03-12},
   abstract = {We  introduce  a  recurrent  architecture  having  a  modular structure  and we formulate a training procedure based on the EM  algorithm.  The resulting model has similarities to hidden  Markov models, but  supports  recurrent  networks  processing style and allows  to exploit  the supervised  learning paradigm while using maximum likelihood  estimation.},
   keywords = {hmm,thesis},
-  file = {/Users/guillaume/Zotero/storage/68UNNYP2/Bengio_Frasconi_1994_An Input Output HMM Architecture.pdf}
+  file = {/home/guillaume/Zotero/storage/68UNNYP2/Bengio_Frasconi_1994_An Input Output HMM Architecture.pdf}
 }
+% == BibTeX quality report for bengioInputOutputHMM1994:
+% ? Unsure about the formatting of the booktitle
+% ? Title looks like it was stored in title-case in Zotero
+% ? unused Library catalog ("Neural Information Processing Systems")
 
 @article{besanconDistributionsJlDefinition2021,
   title = {Distributions{{.jl}}: {{Definition}} and {{Modeling}} of {{Probability Distributions}} in the {{JuliaStats Ecosystem}}},
@@ -35,13 +41,18 @@
   pages = {1--30},
   issn = {1548-7660},
   doi = {10.18637/jss.v098.i16},
+  url = {https://doi.org/10.18637/jss.v098.i16},
   urldate = {2022-09-19},
   abstract = {Random variables and their distributions are a central part in many areas of statistical methods. The Distributions.jl package provides Julia users and developers tools for working with probability distributions, leveraging Julia features for their intuitive and flexible manipulation, while remaining highly efficient through zero-cost abstractions.},
   copyright = {Copyright (c) 2021 Mathieu Besan{\c c}on, Theodore Papamarkou, David Anthoff, Alex Arslan, Simon Byrne, Dahua Lin, John Pearson},
   langid = {english},
+  language = {en},
   keywords = {hmm,thesis},
-  file = {/Users/guillaume/Zotero/storage/FZ5V2QNZ/Besancon et al_2021_Distributions.pdf}
+  file = {/home/guillaume/Zotero/storage/FZ5V2QNZ/Besancon et al_2021_Distributions.pdf}
 }
+% == BibTeX quality report for besanconDistributionsJlDefinition2021:
+% ? Title looks like it was stored in title-case in Zotero
+% ? unused Library catalog ("www.jstatsoft.org")
 
 @article{bezansonJuliaFreshApproach2017,
   title = {Julia: {{A Fresh Approach}} to {{Numerical Computing}}},
@@ -55,39 +66,55 @@
   pages = {65--98},
   issn = {0036-1445, 1095-7200},
   doi = {10.1137/141000671},
[…]
