### 2024-03-12 #6436 @gdalle
I found out where the issue was: my Zotero settings had disabled URL export.
