### 2024-03-13 f7cf63b
Fix benchmark reproducibility (#95)
* Commit comparison manifest

* Ensure reproducibility
