### 2024-03-13 083c26b
Fix benchmarks and citations
