+++ libs/HMMComparison/experiments/results/python_setup.txt
@@ -1,0 +1,16 @@
+Pytorch threads = 1
+
+# Python packages
+
+CondaPkg Status /home/guillaume/Work/GitHub/Julia/HiddenMarkovModels.jl/libs/HMMComparison/CondaPkg.toml
+Environment
+  /home/guillaume/Work/GitHub/Julia/HiddenMarkovModels.jl/libs/HMMComparison/.CondaPkg/env
+Packages
+  cpuonly v2.0 (channel=pytorch)
+  jax v0.4.25 (channel=conda-forge)
+  numpy v1.26.4 (channel=conda-forge)
+  pytorch v2.2.1 (channel=pytorch)
+Pip packages
+  dynamax v0.1.3
+  hmmlearn v0.3.2
+  pomegranate v1.0.4
