### 2024-03-13 #6436 @gdalle
@dmbates now the citations all have links, and I fixed a minor detail for benchmark reproducibility, which slightly changed the plots cause I re-ran them.
