### 2024-04-01 #6436 @gdalle
We can keep trying the suggestions here

https://github.com/openjournals/joss-reviews/issues/5874#issuecomment-1735023132
