### 2024-04-01 #6436 @DanielRivasMD
Hi,

      I deeply apologize for the enormoys delay. Something urgent and personal came up that grab my focus for a couple of weeks.

      I can get the review done tomorrow since I have already check the paper and make my own notes. I hope this is fine with you all. Please let me know.

      Once again, I am sorry about the delay.

      Cheers,

      Daniel Rivas

       On Monday, Apr 01, 2024 at 9:50 AM, Mehmet Hakan Satman ***@***.***> wrote:

      @gdalle - I regret to say that it seems one of our reviewers ***@***.***) is unable to find a proper time to respond. They haven't even created their task list. I believe we should look for another suitable reviewer at this point. Do you have any suggestions for reviewers?
      ***@***.***": ***@***.***": "EmailMessage","potentialAction": ***@***.***": "ViewAction","target": "https://github.com/openjournals/joss-reviews/issues/6436#issuecomment-2030035230","url": "https://github.com/openjournals/joss-reviews/issues/6436#issuecomment-2030035230","name": "View Issue"},"description": "View this Issue on GitHub","publisher": ***@***.***": "Organization","name": "GitHub","url": "https://github.com"}}]
