### 2024-04-03 #6436 @gdalle
I think it's `editorialbot generate my checklist` (with an at)
