### 2024-04-04 #6436 @gdalle
Hi @DanielRivasMD, thank you for the kind words! Glad my package can be put to good use.

> There are some character subindexes on the files forward.jl & viterbi.jl of the inference folder that are troublesome to render. Not suggesting you changing them, just an observation.

Yeah, I sometimes use Unicode a bit more than I should. But I just checked and the symbols render correctly on GitHub, so even if users want to check out the code it should be fine. And if they want to look at it in their IDE, fonts such as [JuliaMono](https://juliamono.netlify.app/) enable perfect unicode rendering locally.

> I am puzzle about the precompilation, particularly about the functions with those values. Why is this the case, and how does it actually impact your performance benchmarks?

I assume you're referring to these lines?

https://github.com/gdalle/HiddenMarkovModels.jl/blob/f7cf63b48fb4853376071772ce35c55a73f57e5c/src/precompile.jl#L1-L15

With precompilation, I had hoped to improve user experience by reducing latency. At package install (or first package load, I'm not 100% sure), precompilation allows selected functions to run on the provided values, so that they are compiled once and for all, instead of having to be compiled again in every Julia session.
But now that you mention it, I don't know if it helps, because every compiled method will be specialized for the case of a Gaussian HMM. So it's probably the same to remove it.

Regardless, _precompilation does not affect the performance benchmarks_, because as is the norm in Julia world, _measurements exclude the first (compiling) run_. The idea is that those functions will be run numerous times anyway, so the startup latency becomes negligible. But the first run is precisely what precompilation would speed up, the other ones are not affected.

Note that just-in-time compilation of the benchmark code is also an issue for the JAX-based competitor `dynamax`. I have done my best to mitigate bias, and have even asked the developers for advice (https://github.com/probml/dynamax/issues/359), but without answer.

> Another observation. I personally would have not benchmark inside Julia using PythonCall.jl because of the overhead. However, I appreciate you explictly state in the paper.

Yeah, it's an opinionated choice, and it mostly comes down to simplicity, because running all the benchmarks from the same place saves a lot of boilerplate and simplifies comparison.

Of course there is the question of the overhead. In an earlier version of the paper, I had _subtracted the time for the smallest instance from every Python-generated curve, as a way to (over)estimate the Julia-to-Python overhead_. I can do that again if you think it necessary, it's a matter of 30 minutes or so.

> Is it possible to add more documentation in libs as to what they are meant for and how to use them, please?

Those libraries are _not meant to be user-facing_, which is why they are not documented:
- `HMMTest` is used for internal testing.
- `HMMBenchmark` is used for internal benchmarking
- `HMMComparison` was written specifically to perform the comparative benchmarks in this paper.

Such toolkits are relevant for me as a developer, but for the vast majority of users they should be pretty useless (after all, users should trust that the values returned by HiddenMarkovModels.jl are correct).

> Lastly, probably the most critical point. Although you clearly state the reliability and speed of this package as key points, I believe the former overseeds the latter. In the same sense, GitHub batches state 93% coverage, but looking at the test directory I am not convienced this would be the case. Is it the case that this package is extensibly tested, and could you do something to improve its reliability?

Good catch, but the `test` directory is _not the only source of tests_, not even the major one.
The [`example` directory](https://github.com/gdalle/HiddenMarkovModels.jl/tree/main/examples) contains six Julia files that are used both for documentation and for testing.

For the documentation website, they are parsed with [Literate.jl](https://github.com/fredrikekre/Literate.jl) and converted to Markdown files, which in turn generate HTML. For instance, the [`examples/controlled.jl` file](https://github.com/gdalle/HiddenMarkovModels.jl/blob/main/examples/controlled.jl) becomes the ["Control dependency" tutorial](https://gdalle.github.io/HiddenMarkovModels.jl/stable/examples/controlled/).

But these files are also included in the test suite, as shown in the following lines from the `test/runtests.jl` file:

https://github.com/gdalle/HiddenMarkovModels.jl/blob/f7cf63b48fb4853376071772ce35c55a73f57e5c/test/runtests.jl#L36-L41

Each of them contains lines that end with `#src`: those bits of code are voluntarily ignored by Literate.jl for building the website. You don't see these parts in the docs, but they do run every time in the tests:

https://github.com/gdalle/HiddenMarkovModels.jl/blob/f7cf63b48fb4853376071772ce35c55a73f57e5c/examples/controlled.jl#L150-L155

These example files do the heavy lifting of ensuring correctness and reliability, including all the weird cases like sparse arrays, time and control dependency, autodiff, etc. Since those weird cases also need to be documented, this is a "2 birds, 1 stone" kind of trick.

As a side note, the code coverage percentage is usually rather reliable, even if it has a few quirks. You can [click the badge](https://app.codecov.io/gh/gdalle/HiddenMarkovModels.jl/tree/main/) to check the individual lines on Codecov for more details.

---

Does that answer your questions? Happy to provide more precisions if needed.
