### 2024-04-04 #6436 @jbytecode
@gdalle - Do we have anything to do with missing DOIs? Could you please add them as suggested, if possible?
