### 2024-04-04 #6436 @jbytecode
@gdalle - The whole paper seems good to me. Thank you!

Now, we need a tagged release in the target repository and a Zenodo archive linked to this version tag. Please make sure that the author list, the manuscript title, author DOIs match in the Zenodo archive. Please report here the version, Zenodo DOI, and the Zenodo link.

Please consider adding missing DOIs as much as you can.

After that, please perform a full grammar check. Tell me when you have done with it and report the changes if exist.

Thank you in advance.
