+++ paper/paper.md
@@ -48,8 +48,8 @@ In this industrial use case, the observations were marked temporal point process
 
 Unfortunately, nearly all implementations of HMMs we surveyed (in Julia and Python) expect the observations to be generated by a _predefined set of distributions_, with _no temporal heterogeneity_.
 In Julia, the previous reference package `HMMBase.jl` [@mouchetHMMBaseJlHidden2023] requires compliance with the `Distributions.jl` [@besanconDistributionsJlDefinition2021] interface, which precludes anything not scalar- or array-valued, let alone point processes.
-In Python, the `numpy`-based `hmmlearn` [@hmmlearndevelopersHmmlearnHiddenMarkov2023] and the `PyTorch`-based `pomegranate` [@schreiberPomegranateFastFlexible2018,@schreiberJmschreiPomegranate2024] each offer a catalogue of discrete and continuous distributions, but do not allow for easy extension by the user.
-The more recent `JAX`-based `dynamax` [@changDynamaxStateSpace2024,@murphyProbabilisticMachineLearning2023,@sarkkaBayesianFilteringSmoothing2023] is the only package adopting an extensible interface with optional controls, similar to ours.
+In Python, the `numpy`-based `hmmlearn` [@hmmlearndevelopersHmmlearnHiddenMarkov2023] and the `PyTorch`-based `pomegranate` [@schreiberPomegranateFastFlexible2018; @schreiberJmschreiPomegranate2024] each offer a catalogue of discrete and continuous distributions, but do not allow for easy extension by the user.
+The more recent `JAX`-based `dynamax` [@changDynamaxStateSpace2024; @murphyProbabilisticMachineLearning2023; @sarkkaBayesianFilteringSmoothing2023] is the only package adopting an extensible interface with optional controls, similar to ours.
 
 Focusing on Julia specifically, other downsides of `HMMBase.jl` include the lack of support for _multiple observation sequences_, _automatic differentiation_, _sparse transition matrices_ or _number types beyond 64-bit floating point_.
 Two other Julia packages each provide a subset of functionalities that `HMMBase.jl` lacks, namely `HMMGradients.jl` [@antonelloHMMGradientsJlEnables2021] and `MarkovModels.jl` [@ondelGPUAcceleratedForwardBackwardAlgorithm2022], but they are less developed and ill-suited to uninformed users.
