+++ paper/HMM.bib
@@ -1,12 +1,12 @@
-@misc{antonelloHMMGradientsJlEnables2021,
+@software{antonelloHMMGradientsJlEnables2021,
   title = {{{HMMGradients}}.Jl: {{Enables}} Computing the Gradient of the Parameters of {{Hidden Markov Models}} ({{HMMs}})},
   shorttitle = {Idiap/{{HMMGradients}}.Jl},
-  author = {Antonello, Niccol{\`o}},
-  year = {2021},
-  month = jun,
+  author = {Antonello, Niccolò},
+  date = {2021-06-07},
   doi = {10.5281/zenodo.4454565},
+  url = {https://doi.org/10.5281/zenodo.4454565},
   urldate = {2023-09-12},
-  howpublished = {Zenodo},
+  organization = {Zenodo},
   keywords = {hmm},
   file = {/home/gdalle/snap/zotero-snap/common/Zotero/storage/PEFYSLF7/4906644.html}
 }
@@ -15,29 +15,29 @@
   title = {An {{Input Output HMM Architecture}}},
   booktitle = {Advances in {{Neural Information Processing Systems}}},
   author = {Bengio, Yoshua and Frasconi, Paolo},
-  year = {1994},
+  date = {1994},
   volume = {7},
   publisher = {MIT Press},
+  url = {https://proceedings.neurips.cc/paper/1994/hash/8065d07da4a77621450aa84fee5656d9-Abstract.html},
   urldate = {2023-03-12},
   abstract = {We  introduce  a  recurrent  architecture  having  a  modular structure  and we formulate a training procedure based on the EM  algorithm.  The resulting model has similarities to hidden  Markov models, but  supports  recurrent  networks  processing style and allows  to exploit  the supervised  learning paradigm while using maximum likelihood  estimation.},
-  keywords = {hmm,No DOI found,thesis},
+  keywords = {⛔ No DOI found,hmm,thesis},
   file = {/home/gdalle/snap/zotero-snap/common/Zotero/storage/68UNNYP2/Bengio_Frasconi_1994_An Input Output HMM Architecture.pdf}
 }
 
 @article{besanconDistributionsJlDefinition2021,
   title = {Distributions{{.jl}}: {{Definition}} and {{Modeling}} of {{Probability Distributions}} in the {{JuliaStats Ecosystem}}},
   shorttitle = {Distributions.Jl},
-  author = {Besan{\c c}on, Mathieu and Papamarkou, Theodore and Anthoff, David and Arslan, Alex and Byrne, Simon and Lin, Dahua and Pearson, John},
-  year = {2021},
-  month = jul,
-  journal = {Journal of Statistical Software},
+  author = {Besançon, Mathieu and Papamarkou, Theodore and Anthoff, David and Arslan, Alex and Byrne, Simon and Lin, Dahua and Pearson, John},
+  date = {2021-07-25},
+  journaltitle = {Journal of Statistical Software},
   volume = {98},
   pages = {1--30},
   issn = {1548-7660},
   doi = {10.18637/jss.v098.i16},
+  url = {https://doi.org/10.18637/jss.v098.i16},
   urldate = {2022-09-19},
   abstract = {Random variables and their distributions are a central part in many areas of statistical methods. The Distributions.jl package provides Julia users and developers tools for working with probability distributions, leveraging Julia features for their intuitive and flexible manipulation, while remaining highly efficient through zero-cost abstractions.},
-  copyright = {Copyright (c) 2021 Mathieu Besan{\c c}on, Theodore Papamarkou, David Anthoff, Alex Arslan, Simon Byrne, Dahua Lin, John Pearson},
[…]
