+++ src/HiddenMarkovModels.jl
@@ -17,7 +17,6 @@ using DensityInterface: DensityInterface, DensityKind, HasDensity, NoDensity, lo
 using DocStringExtensions
 using FillArrays: Fill
 using LinearAlgebra: Transpose, dot, ldiv!, lmul!, mul!, parent
-using PrecompileTools: @compile_workload
 using Random: Random, AbstractRNG, default_rng
 using SparseArrays: AbstractSparseArray, SparseMatrixCSC, nonzeros, nnz, nzrange, rowvals
 using StatsAPI: StatsAPI, fit, fit!
@@ -47,7 +46,5 @@ include("inference/logdensity.jl")
 include("inference/chainrules.jl")
 
 include("types/hmm.jl")
-
-include("precompile.jl")
 
 end
