### 2024-04-05 #6436 @gdalle
I have released v0.5.1 of the package

https://github.com/gdalle/HiddenMarkovModels.jl/releases/tag/v0.5.1

The associated Zenodo DOI is

https://doi.org/10.5281/zenodo.10931812

> Please make sure that the author list, the manuscript title, author DOIs match in the Zenodo archive.

What do you mean by that? The Zenodo archive doesn't know anything about the manuscript since it is on a separate branch
