### 2024-04-05 #6436 @jbytecode
It's just how JOSS handles this issue as mentioned in [here](https://joss.readthedocs.io/en/latest/editing.html#message-to-authors-at-the-end-of-a-review). Thank you for applying the change!
