### 2024-04-05 #6436 @jbytecode
@gdalle - The problem is about the DOI entry of the bibtex item

```
@article{rabinerTutorialHiddenMarkov1989,
  title = {A Tutorial on Hidden {{Markov}} Models and Selected Applications in Speech Recognition},
  author = {Rabiner, L.R.},
  date = {1989-02},
  journaltitle = {Proceedings of the IEEE},
  volume = {77},
  number = {2},
  pages = {257--286},
  issn = {1558-2256},
  doi = {10/cswph2},
  ...
}
```

Could you please fix it and then ping me again?
