### 2024-04-05 #6436 @jbytecode
@openjournals/dev - Interestingly the doi is resolved by DOI.org however we get this error:

```
!! Invalid Crossref XML v5.3.1 in tmp/3a94f67f00ebb84ab56aec5ba38026f1/paper/paper.crossref. Errors: 1.
  - 268:0: ERROR: Element {http://www.crossref.org/schema/5.3.1}doi: [facet pattern] The value 10/cswph2 is not accepted by the pattern 10.[0-9]{4,9}/.{1,200}.
```
