### 2024-04-05 #6436 @jbytecode
@gdalle - using `10.1109/5.18626` instead may be fix that because the bibtex is reported as

```
@ARTICLE{18626,
  author={Rabiner, L.R.},
  journal={Proceedings of the IEEE},
  title={A tutorial on hidden Markov models and selected applications in speech recognition},
  year={1989},
  volume={77},
  number={2},
  pages={257-286},
  keywords={Tutorial;Hidden Markov models;Speech recognition},
  doi={10.1109/5.18626}}
```
