+++ paper/HMM.bib
@@ -221,7 +221,7 @@
   number = {2},
   pages = {257--286},
   issn = {1558-2256},
-  doi = {10/cswph2},
+  doi = {10.1109/5.18626},
   abstract = {This tutorial provides an overview of the basic theory of hidden Markov models (HMMs) as originated by L.E. Baum and T. Petrie (1966) and gives practical details on methods of implementation of the theory along with a description of selected applications of the theory to distinct problems in speech recognition. Results from a number of original sources are combined to provide a single source of acquiring the background required to pursue further this area of research. The author first reviews the theory of discrete Markov chains and shows how the concept of hidden states, where the observation is a probabilistic function of the state, can be used effectively. The theory is illustrated with two simple examples, namely coin-tossing, and the classic balls-in-urns system. Three fundamental problems of HMMs are noted and several practical techniques for solving these problems are given. The various types of HMMs that have been studied, including ergodic as well as left-right models, are described.{$<>$}},
   eventtitle = {Proceedings of the {{IEEE}}},
   keywords = {done,hmm,thesis,viva},
