### 2024-04-05 #6436 @gdalle
I just made the modification, anything to make our little bot happy ;)
